refactor: split shenanigans effects into individual files
Extract all 27 visual effects from the monolithic shenanigans.go (7200+ lines)
into individual effect_<name>.go files for better readability and maintainability.

Each effect file has a descriptive comment explaining what the effect does:
- effect_matrix.go — Matrix digital rain (falling Katakana characters)
- effect_fire.go — Doom-style procedural fire simulation
- effect_snow.go — Snowfall with wind drift and wobble
- effect_plasma.go — Classic demoscene plasma (layered sine waves)
- effect_starfield.go — 3D starfield with parallax depth
- effect_aurora.go — Northern lights with shimmering curtains
- effect_glitch.go — VHS-style digital glitch artifacts
- effect_pride.go — Animated rainbow pride flag
- effect_football.go — Football pitch with bouncing ball
- effect_fireworks.go — Celebratory fireworks with particle trails
- effect_pong.go — Classic Pong with two AI paddles
- effect_life.go — Conway's Game of Life
- effect_invaders.go — Space Invaders arcade game
- effect_snake.go — Classic Snake game
- effect_rain.go — Rainy window with droplets and splashes
- effect_breakout.go — Breakout/Arkanoid arcade game
- effect_dino.go — Chrome dinosaur runner
- effect_asteroids.go — Atari Asteroids arcade game
- effect_pacman.go — Pac-Man with ghosts
- effect_tetris.go — Tetris with falling tetrominoes
- effect_frogger.go — Frogger arcade game
- effect_minesweeper.go — Minesweeper auto-solve
- effect_flappy.go — Flappy Bird
- effect_lava.go — Lava lamp with soft blobs
- effect_sinescroll.go — Retro sine-wave text scroller
- effect_fireflies.go — Glowing fireflies with bioluminescence
- effect_boids.go — Boids flocking simulation
- effect_raycast.go — Wolfenstein 3D-style raycaster

Shared math helpers (sinApprox, cosApprox, absF, sinNorm, atan2Approx)
moved to helpers.go. Main shenanigans.go now contains only the plugin
struct, constants, dispatch switch, and shared utilities (238 lines).

Also updated test to include all 29 effect options in metadata assertion.
