+++ plugins/shenanigans/effect_matrix.go
@@ -1,0 +1,224 @@
+//nolint:mnd,gosec // Visual effects plugin: magic numbers and weak random are by design.
+package main
+
+import (
+	"math/rand/v2"
+	"time"
+
+	"github.com/strobotti/linkquisition"
+)
+
+// Matrix digital rain — falling green Katakana characters like in The Matrix.
+// --- Matrix Rain Effect ---
+
+type matrixState struct {
+	columns []matrixColumn
+	width   int
+	height  int
+}
+
+type matrixColumn struct {
+	y     float64
+	speed float64
+	chars []rune
+}
+
+func (p *shenanigans) startMatrixRain(pc linkquisition.PickerCanvas) {
+	state := &matrixState{}
+	state.width = pc.Width()
+	state.height = pc.Height()
+	state.initColumns()
+
+	pc.AddRasterOverlay(0.6, func(w, h int) []uint8 {
+		if w != state.width || h != state.height {
+			state.width = w
+			state.height = h
+			state.initColumns()
+		}
+		return p.invertForLight(state.render())
+	})
+
+	go func() {
+		ticker := time.NewTicker(frameInterval)
+		defer ticker.Stop()
+
+		for range ticker.C {
+			if p.stopped.Load() {
+				return
+			}
+			state.update()
+			pc.ScheduleRefresh()
+		}
+	}()
+}
+
+func (s *matrixState) initColumns() {
+	colCount := matrixColumns
+	if s.width > 0 {
+		colCount = s.width / 14
+		if colCount < 1 {
+			colCount = 1
+		}
+	}
+
+	s.columns = make([]matrixColumn, colCount)
+	for i := range s.columns {
+		s.columns[i] = matrixColumn{
+			y:     -rand.Float64() * float64(s.height),
+			speed: 2 + rand.Float64()*6,
+			chars: generateMatrixChars(20 + rand.IntN(15)),
+		}
+	}
+}
+
+func (s *matrixState) update() {
+	h := float64(s.height)
+	if h == 0 {
+		h = 400
+	}
+
+	for i := range s.columns {
+		s.columns[i].y += s.columns[i].speed
+		if s.columns[i].y > h+float64(len(s.columns[i].chars)*16) {
+			s.columns[i].y = -float64(len(s.columns[i].chars) * 16)
+			s.columns[i].speed = 2 + rand.Float64()*6
+			s.columns[i].chars = generateMatrixChars(20 + rand.IntN(15))
+		}
+	}
+}
+
+func (s *matrixState) render() []uint8 {
+	w, h := s.width, s.height
+	if w == 0 || h == 0 {
+		return make([]uint8, rgbaChannels)
+	}
+
+	pixels := make([]uint8, w*h*rgbaChannels)
+
+	colWidth := w / len(s.columns)
+	if colWidth < 1 {
+		colWidth = 1
+	}
+
+	for i, col := range s.columns {
+		x := i * colWidth
+		charHeight := 16
+
+		for j := range col.chars {
+			cy := int(col.y) - j*charHeight
+			if cy >= h {
+				continue
+			}
+			if cy+charHeight <= 0 {
+				break // this and all subsequent trail chars are above viewport
+			}
+
+			// Fade out older characters
+			brightness := uint8(255 - min(j*12, 230))
+
+			// Render the character glyph
+			glyph := matrixGlyph(col.chars[j])
+			glyphW := 8
+			glyphH := 12
+			// Center glyph within the column cell
+			offsetX := (colWidth - glyphW) / 2
+			offsetY := (charHeight - glyphH) / 2
+
+			for gy := range glyphH {
+				for gx := range glyphW {
+					if glyph[gy]&(1<<(7-gx)) == 0 {
+						continue
+					}
+					px := x + offsetX + gx
+					py := cy + offsetY + gy
+					if px >= 0 && px < w && py >= 0 && py < h {
+						offset := (py*w + px) * rgbaChannels
+						pixels[offset] = 0                // R
+						pixels[offset+1] = brightness     // G
+						pixels[offset+2] = brightness / 4 // B
+						pixels[offset+3] = brightness     // A
+					}
+				}
+			}
+		}
+	}
+
+	return pixels
+}
+
+func generateMatrixChars(length int) []rune {
+	chars := []rune("アイウエオカキクケコサシスセソタチツテトナニヌネノハヒフヘホマミムメモ0123456789")
+	result := make([]rune, length)
+	for i := range result {
+		result[i] = chars[rand.IntN(len(chars))]
+	}
+	return result
+}
+
+// matrixGlyphs contains 8x12 bitmap glyphs for the Matrix rain characters.
+// Each glyph is 12 rows of uint8 where bits represent pixels (MSB = leftmost).
+var matrixGlyphs = map[rune][12]uint8{
+	// Katakana
+	'ア': {0x00, 0x7E, 0x02, 0x02, 0x3E, 0x20, 0x20, 0x10, 0x08, 0x04, 0x02, 0x00},
+	'イ': {0x00, 0x02, 0x04, 0x08, 0x18, 0x28, 0x08, 0x08, 0x08, 0x08, 0x08, 0x00},
+	'ウ': {0x00, 0x10, 0x7E, 0x42, 0x42, 0x42, 0x22, 0x22, 0x14, 0x08, 0x04, 0x00},
+	'エ': {0x00, 0x7E, 0x00, 0x00, 0x18, 0x18, 0x18, 0x18, 0x00, 0x00, 0x7E, 0x00},
+	'オ': {0x00, 0x08, 0x7E, 0x08, 0x08, 0x1C, 0x2A, 0x4A, 0x08, 0x08, 0x08, 0x00},
+	'カ': {0x00, 0x10, 0x7E, 0x12, 0x12, 0x22, 0x22, 0x22, 0x42, 0x42, 0x02, 0x00},
+	'キ': {0x00, 0x08, 0x08, 0x7E, 0x08, 0x08, 0x7E, 0x08, 0x08, 0x08, 0x08, 0x00},
+	'ク': {0x00, 0x20, 0x3E, 0x22, 0x22, 0x22, 0x14, 0x14, 0x08, 0x04, 0x02, 0x00},
+	'ケ': {0x00, 0x20, 0x3E, 0x22, 0x04, 0x04, 0x08, 0x08, 0x10, 0x20, 0x40, 0x00},
+	'コ': {0x00, 0x7E, 0x02, 0x02, 0x02, 0x02, 0x02, 0x02, 0x02, 0x7E, 0x00, 0x00},
+	'サ': {0x00, 0x24, 0x24, 0x24, 0x7E, 0x24, 0x24, 0x04, 0x04, 0x08, 0x10, 0x00},
+	'シ': {0x00, 0x42, 0x22, 0x02, 0x02, 0x02, 0x04, 0x08, 0x10, 0x20, 0x40, 0x00},
+	'ス': {0x00, 0x7E, 0x02, 0x04, 0x08, 0x0C, 0x14, 0x22, 0x42, 0x02, 0x02, 0x00},
+	'セ': {0x00, 0x10, 0x10, 0x7E, 0x12, 0x12, 0x12, 0x22, 0x22, 0x42, 0x02, 0x00},
+	'ソ': {0x00, 0x42, 0x22, 0x22, 0x12, 0x04, 0x04, 0x08, 0x08, 0x10, 0x20, 0x00},
+	'タ': {0x00, 0x10, 0x3E, 0x22, 0x22, 0x3E, 0x10, 0x08, 0x04, 0x02, 0x02, 0x00},
+	'チ': {0x00, 0x08, 0x7E, 0x08, 0x08, 0x3C, 0x08, 0x08, 0x10, 0x20, 0x40, 0x00},
+	'ツ': {0x00, 0x44, 0x24, 0x24, 0x04, 0x04, 0x04, 0x08, 0x08, 0x10, 0x20, 0x00},
+	'テ': {0x00, 0x7E, 0x00, 0x08, 0x08, 0x08, 0x08, 0x10, 0x10, 0x20, 0x40, 0x00},
+	'ト': {0x00, 0x20, 0x20, 0x20, 0x38, 0x24, 0x20, 0x20, 0x20, 0x20, 0x20, 0x00},
+	'ナ': {0x00, 0x08, 0x08, 0x7E, 0x08, 0x08, 0x10, 0x10, 0x20, 0x20, 0x40, 0x00},
+	'ニ': {0x00, 0x00, 0x3C, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x7E, 0x00, 0x00},
+	'ヌ': {0x00, 0x7E, 0x02, 0x04, 0x0C, 0x14, 0x24, 0x04, 0x02, 0x02, 0x02, 0x00},
+	'ネ': {0x00, 0x08, 0x7E, 0x08, 0x08, 0x3E, 0x49, 0x08, 0x08, 0x08, 0x08, 0x00},
+	'ノ': {0x00, 0x02, 0x02, 0x04, 0x04, 0x08, 0x08, 0x10, 0x10, 0x20, 0x40, 0x00},
+	'ハ': {0x00, 0x00, 0x10, 0x10, 0x28, 0x28, 0x44, 0x44, 0x82, 0x02, 0x02, 0x00},
+	'ヒ': {0x00, 0x20, 0x20, 0x20, 0x3E, 0x20, 0x20, 0x20, 0x20, 0x1E, 0x00, 0x00},
+	'フ': {0x00, 0x7E, 0x42, 0x02, 0x02, 0x04, 0x04, 0x08, 0x10, 0x20, 0x40, 0x00},
+	'ヘ': {0x00, 0x00, 0x00, 0x10, 0x28, 0x44, 0x82, 0x02, 0x00, 0x00, 0x00, 0x00},
+	'ホ': {0x00, 0x08, 0x7E, 0x08, 0x08, 0x2A, 0x2A, 0x4A, 0x08, 0x08, 0x08, 0x00},
+	'マ': {0x00, 0x7E, 0x02, 0x02, 0x04, 0x08, 0x10, 0x10, 0x08, 0x04, 0x02, 0x00},
+	'ミ': {0x00, 0x00, 0x1E, 0x00, 0x00, 0x3C, 0x00, 0x00, 0x7E, 0x00, 0x00, 0x00},
+	'ム': {0x00, 0x10, 0x10, 0x10, 0x18, 0x14, 0x14, 0x22, 0x42, 0x42, 0x7E, 0x00},
+	'メ': {0x00, 0x02, 0x04, 0x08, 0x14, 0x24, 0x14, 0x08, 0x10, 0x20, 0x40, 0x00},
+	'モ': {0x00, 0x7E, 0x08, 0x08, 0x7E, 0x08, 0x08, 0x08, 0x10, 0x20, 0x40, 0x00},
+	// Digits
+	'0': {0x00, 0x3C, 0x42, 0x46, 0x4A, 0x52, 0x62, 0x42, 0x42, 0x3C, 0x00, 0x00},
+	'1': {0x00, 0x08, 0x18, 0x28, 0x08, 0x08, 0x08, 0x08, 0x08, 0x3E, 0x00, 0x00},
+	'2': {0x00, 0x3C, 0x42, 0x02, 0x04, 0x08, 0x10, 0x20, 0x40, 0x7E, 0x00, 0x00},
+	'3': {0x00, 0x3C, 0x42, 0x02, 0x1C, 0x02, 0x02, 0x02, 0x42, 0x3C, 0x00, 0x00},
+	'4': {0x00, 0x04, 0x0C, 0x14, 0x24, 0x44, 0x7E, 0x04, 0x04, 0x04, 0x00, 0x00},
+	'5': {0x00, 0x7E, 0x40, 0x40, 0x7C, 0x02, 0x02, 0x02, 0x42, 0x3C, 0x00, 0x00},
+	'6': {0x00, 0x1C, 0x20, 0x40, 0x7C, 0x42, 0x42, 0x42, 0x42, 0x3C, 0x00, 0x00},
+	'7': {0x00, 0x7E, 0x02, 0x04, 0x04, 0x08, 0x08, 0x10, 0x10, 0x10, 0x00, 0x00},
+	'8': {0x00, 0x3C, 0x42, 0x42, 0x3C, 0x42, 0x42, 0x42, 0x42, 0x3C, 0x00, 0x00},
+	'9': {0x00, 0x3C, 0x42, 0x42, 0x42, 0x3E, 0x02, 0x02, 0x04, 0x38, 0x00, 0x00},
+}
+
+// matrixGlyph returns the 8x12 bitmap for a rune, falling back to a random glyph.
+func matrixGlyph(r rune) [12]uint8 {
+	if g, ok := matrixGlyphs[r]; ok {
+		return g
+	}
+	// Fallback: generate a pseudo-random glyph pattern from the rune value
+	var g [12]uint8
+	seed := uint64(r)
+	for i := range g {
+		seed = seed*6364136223846793005 + 1
+		g[i] = uint8((seed >> 33) & 0x7E) // keep within 8px width, avoid edges
+	}
+	return g
+}
+
