+++ plugins/shenanigans/shenanigans.go
@@ -229,7030 +229,6 @@ func (p *shenanigans) invertForLight(pixels []uint8) []uint8 {
 	return pixels
 }
 
-// --- Matrix Rain Effect ---
-
-type matrixState struct {
-	columns []matrixColumn
-	width   int
-	height  int
-}
-
-type matrixColumn struct {
-	y     float64
-	speed float64
-	chars []rune
-}
-
-func (p *shenanigans) startMatrixRain(pc linkquisition.PickerCanvas) {
-	state := &matrixState{}
-	state.width = pc.Width()
-	state.height = pc.Height()
-	state.initColumns()
-
-	pc.AddRasterOverlay(0.6, func(w, h int) []uint8 {
-		if w != state.width || h != state.height {
-			state.width = w
-			state.height = h
-			state.initColumns()
-		}
-		return p.invertForLight(state.render())
-	})
-
-	go func() {
-		ticker := time.NewTicker(frameInterval)
-		defer ticker.Stop()
-
-		for range ticker.C {
-			if p.stopped.Load() {
-				return
-			}
-			state.update()
-			pc.ScheduleRefresh()
-		}
-	}()
-}
-
-func (s *matrixState) initColumns() {
-	colCount := matrixColumns
-	if s.width > 0 {
-		colCount = s.width / 14
-		if colCount < 1 {
-			colCount = 1
-		}
-	}
-
-	s.columns = make([]matrixColumn, colCount)
-	for i := range s.columns {
-		s.columns[i] = matrixColumn{
-			y:     -rand.Float64() * float64(s.height),
-			speed: 2 + rand.Float64()*6,
-			chars: generateMatrixChars(20 + rand.IntN(15)),
-		}
-	}
-}
-
-func (s *matrixState) update() {
-	h := float64(s.height)
-	if h == 0 {
-		h = 400
-	}
-
-	for i := range s.columns {
-		s.columns[i].y += s.columns[i].speed
-		if s.columns[i].y > h+float64(len(s.columns[i].chars)*16) {
-			s.columns[i].y = -float64(len(s.columns[i].chars) * 16)
-			s.columns[i].speed = 2 + rand.Float64()*6
-			s.columns[i].chars = generateMatrixChars(20 + rand.IntN(15))
-		}
-	}
-}
-
-func (s *matrixState) render() []uint8 {
-	w, h := s.width, s.height
-	if w == 0 || h == 0 {
-		return make([]uint8, rgbaChannels)
-	}
-
-	pixels := make([]uint8, w*h*rgbaChannels)
-
-	colWidth := w / len(s.columns)
-	if colWidth < 1 {
-		colWidth = 1
-	}
-
-	for i, col := range s.columns {
-		x := i * colWidth
-		charHeight := 16
-
-		for j := range col.chars {
-			cy := int(col.y) - j*charHeight
-			if cy >= h {
-				continue
-			}
-			if cy+charHeight <= 0 {
-				break // this and all subsequent trail chars are above viewport
-			}
-
-			// Fade out older characters
-			brightness := uint8(255 - min(j*12, 230))
-
-			// Render the character glyph
-			glyph := matrixGlyph(col.chars[j])
-			glyphW := 8
-			glyphH := 12
-			// Center glyph within the column cell
-			offsetX := (colWidth - glyphW) / 2
-			offsetY := (charHeight - glyphH) / 2
-
-			for gy := range glyphH {
-				for gx := range glyphW {
-					if glyph[gy]&(1<<(7-gx)) == 0 {
-						continue
-					}
-					px := x + offsetX + gx
-					py := cy + offsetY + gy
-					if px >= 0 && px < w && py >= 0 && py < h {
-						offset := (py*w + px) * rgbaChannels
-						pixels[offset] = 0                // R
-						pixels[offset+1] = brightness     // G
-						pixels[offset+2] = brightness / 4 // B
-						pixels[offset+3] = brightness     // A
-					}
-				}
-			}
-		}
-	}
-
-	return pixels
-}
-
-func generateMatrixChars(length int) []rune {
-	chars := []rune("アイウエオカキクケコサシスセソタチツテトナニヌネノハヒフヘホマミムメモ0123456789")
-	result := make([]rune, length)
-	for i := range result {
-		result[i] = chars[rand.IntN(len(chars))]
-	}
-	return result
-}
-
-// matrixGlyphs contains 8x12 bitmap glyphs for the Matrix rain characters.
-// Each glyph is 12 rows of uint8 where bits represent pixels (MSB = leftmost).
-var matrixGlyphs = map[rune][12]uint8{
-	// Katakana
-	'ア': {0x00, 0x7E, 0x02, 0x02, 0x3E, 0x20, 0x20, 0x10, 0x08, 0x04, 0x02, 0x00},
-	'イ': {0x00, 0x02, 0x04, 0x08, 0x18, 0x28, 0x08, 0x08, 0x08, 0x08, 0x08, 0x00},
-	'ウ': {0x00, 0x10, 0x7E, 0x42, 0x42, 0x42, 0x22, 0x22, 0x14, 0x08, 0x04, 0x00},
-	'エ': {0x00, 0x7E, 0x00, 0x00, 0x18, 0x18, 0x18, 0x18, 0x00, 0x00, 0x7E, 0x00},
-	'オ': {0x00, 0x08, 0x7E, 0x08, 0x08, 0x1C, 0x2A, 0x4A, 0x08, 0x08, 0x08, 0x00},
-	'カ': {0x00, 0x10, 0x7E, 0x12, 0x12, 0x22, 0x22, 0x22, 0x42, 0x42, 0x02, 0x00},
-	'キ': {0x00, 0x08, 0x08, 0x7E, 0x08, 0x08, 0x7E, 0x08, 0x08, 0x08, 0x08, 0x00},
-	'ク': {0x00, 0x20, 0x3E, 0x22, 0x22, 0x22, 0x14, 0x14, 0x08, 0x04, 0x02, 0x00},
-	'ケ': {0x00, 0x20, 0x3E, 0x22, 0x04, 0x04, 0x08, 0x08, 0x10, 0x20, 0x40, 0x00},
-	'コ': {0x00, 0x7E, 0x02, 0x02, 0x02, 0x02, 0x02, 0x02, 0x02, 0x7E, 0x00, 0x00},
-	'サ': {0x00, 0x24, 0x24, 0x24, 0x7E, 0x24, 0x24, 0x04, 0x04, 0x08, 0x10, 0x00},
-	'シ': {0x00, 0x42, 0x22, 0x02, 0x02, 0x02, 0x04, 0x08, 0x10, 0x20, 0x40, 0x00},
-	'ス': {0x00, 0x7E, 0x02, 0x04, 0x08, 0x0C, 0x14, 0x22, 0x42, 0x02, 0x02, 0x00},
-	'セ': {0x00, 0x10, 0x10, 0x7E, 0x12, 0x12, 0x12, 0x22, 0x22, 0x42, 0x02, 0x00},
-	'ソ': {0x00, 0x42, 0x22, 0x22, 0x12, 0x04, 0x04, 0x08, 0x08, 0x10, 0x20, 0x00},
-	'タ': {0x00, 0x10, 0x3E, 0x22, 0x22, 0x3E, 0x10, 0x08, 0x04, 0x02, 0x02, 0x00},
-	'チ': {0x00, 0x08, 0x7E, 0x08, 0x08, 0x3C, 0x08, 0x08, 0x10, 0x20, 0x40, 0x00},
-	'ツ': {0x00, 0x44, 0x24, 0x24, 0x04, 0x04, 0x04, 0x08, 0x08, 0x10, 0x20, 0x00},
-	'テ': {0x00, 0x7E, 0x00, 0x08, 0x08, 0x08, 0x08, 0x10, 0x10, 0x20, 0x40, 0x00},
-	'ト': {0x00, 0x20, 0x20, 0x20, 0x38, 0x24, 0x20, 0x20, 0x20, 0x20, 0x20, 0x00},
-	'ナ': {0x00, 0x08, 0x08, 0x7E, 0x08, 0x08, 0x10, 0x10, 0x20, 0x20, 0x40, 0x00},
-	'ニ': {0x00, 0x00, 0x3C, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x7E, 0x00, 0x00},
-	'ヌ': {0x00, 0x7E, 0x02, 0x04, 0x0C, 0x14, 0x24, 0x04, 0x02, 0x02, 0x02, 0x00},
-	'ネ': {0x00, 0x08, 0x7E, 0x08, 0x08, 0x3E, 0x49, 0x08, 0x08, 0x08, 0x08, 0x00},
-	'ノ': {0x00, 0x02, 0x02, 0x04, 0x04, 0x08, 0x08, 0x10, 0x10, 0x20, 0x40, 0x00},
-	'ハ': {0x00, 0x00, 0x10, 0x10, 0x28, 0x28, 0x44, 0x44, 0x82, 0x02, 0x02, 0x00},
-	'ヒ': {0x00, 0x20, 0x20, 0x20, 0x3E, 0x20, 0x20, 0x20, 0x20, 0x1E, 0x00, 0x00},
-	'フ': {0x00, 0x7E, 0x42, 0x02, 0x02, 0x04, 0x04, 0x08, 0x10, 0x20, 0x40, 0x00},
-	'ヘ': {0x00, 0x00, 0x00, 0x10, 0x28, 0x44, 0x82, 0x02, 0x00, 0x00, 0x00, 0x00},
-	'ホ': {0x00, 0x08, 0x7E, 0x08, 0x08, 0x2A, 0x2A, 0x4A, 0x08, 0x08, 0x08, 0x00},
-	'マ': {0x00, 0x7E, 0x02, 0x02, 0x04, 0x08, 0x10, 0x10, 0x08, 0x04, 0x02, 0x00},
-	'ミ': {0x00, 0x00, 0x1E, 0x00, 0x00, 0x3C, 0x00, 0x00, 0x7E, 0x00, 0x00, 0x00},
-	'ム': {0x00, 0x10, 0x10, 0x10, 0x18, 0x14, 0x14, 0x22, 0x42, 0x42, 0x7E, 0x00},
-	'メ': {0x00, 0x02, 0x04, 0x08, 0x14, 0x24, 0x14, 0x08, 0x10, 0x20, 0x40, 0x00},
-	'モ': {0x00, 0x7E, 0x08, 0x08, 0x7E, 0x08, 0x08, 0x08, 0x10, 0x20, 0x40, 0x00},
-	// Digits
-	'0': {0x00, 0x3C, 0x42, 0x46, 0x4A, 0x52, 0x62, 0x42, 0x42, 0x3C, 0x00, 0x00},
-	'1': {0x00, 0x08, 0x18, 0x28, 0x08, 0x08, 0x08, 0x08, 0x08, 0x3E, 0x00, 0x00},
-	'2': {0x00, 0x3C, 0x42, 0x02, 0x04, 0x08, 0x10, 0x20, 0x40, 0x7E, 0x00, 0x00},
-	'3': {0x00, 0x3C, 0x42, 0x02, 0x1C, 0x02, 0x02, 0x02, 0x42, 0x3C, 0x00, 0x00},
-	'4': {0x00, 0x04, 0x0C, 0x14, 0x24, 0x44, 0x7E, 0x04, 0x04, 0x04, 0x00, 0x00},
-	'5': {0x00, 0x7E, 0x40, 0x40, 0x7C, 0x02, 0x02, 0x02, 0x42, 0x3C, 0x00, 0x00},
-	'6': {0x00, 0x1C, 0x20, 0x40, 0x7C, 0x42, 0x42, 0x42, 0x42, 0x3C, 0x00, 0x00},
-	'7': {0x00, 0x7E, 0x02, 0x04, 0x04, 0x08, 0x08, 0x10, 0x10, 0x10, 0x00, 0x00},
-	'8': {0x00, 0x3C, 0x42, 0x42, 0x3C, 0x42, 0x42, 0x42, 0x42, 0x3C, 0x00, 0x00},
-	'9': {0x00, 0x3C, 0x42, 0x42, 0x42, 0x3E, 0x02, 0x02, 0x04, 0x38, 0x00, 0x00},
-}
-
-// matrixGlyph returns the 8x12 bitmap for a rune, falling back to a random glyph.
-func matrixGlyph(r rune) [12]uint8 {
-	if g, ok := matrixGlyphs[r]; ok {
-		return g
-	}
-	// Fallback: generate a pseudo-random glyph pattern from the rune value
-	var g [12]uint8
-	seed := uint64(r)
-	for i := range g {
-		seed = seed*6364136223846793005 + 1
-		g[i] = uint8((seed >> 33) & 0x7E) // keep within 8px width, avoid edges
-	}
-	return g
-}
-
-// --- Fire Effect ---
-
-type fireState struct {
-	grid   [][]uint8
-	width  int
-	height int
-}
-
-func (p *shenanigans) startFire(pc linkquisition.PickerCanvas) {
-	state := &fireState{
-		width:  fireWidth,
-		height: fireHeight,
-	}
-	state.init()
-
-	pc.AddRasterOverlay(0.4, func(w, h int) []uint8 {
-		return state.render(w, h)
-	})
-
-	go func() {
-		ticker := time.NewTicker(fireFrameInterval)
-		defer ticker.Stop()
-
-		for range ticker.C {
-			if p.stopped.Load() {
-				return
-			}
-			state.update()
-			state.update() // double-step for faster movement
-			pc.ScheduleRefresh()
-		}
-	}()
-}
-
-func (s *fireState) init() {
-	s.grid = make([][]uint8, s.height)
-	for i := range s.grid {
-		s.grid[i] = make([]uint8, s.width)
-	}
-}
-
-func (s *fireState) update() {
-	// Set the bottom two rows to random hot values (wider fuel source)
-	for x := range s.width {
-		s.grid[s.height-1][x] = uint8(180 + rand.IntN(76))
-		s.grid[s.height-2][x] = uint8(150 + rand.IntN(106))
-	}
-
-	// Propagate fire upward with averaging and cooling
-	for y := range s.height - 2 {
-		for x := range s.width {
-			// Sample a wider neighborhood for smoother spread
-			l2 := max(x-2, 0)
-			l1 := max(x-1, 0)
-			r1 := min(x+1, s.width-1)
-			r2 := min(x+2, s.width-1)
-
-			// Weighted average: center-heavy for more vertical flames
-			sum := int(s.grid[y+1][l1]) +
-				int(s.grid[y+1][x])*3 +
-				int(s.grid[y+1][r1]) +
-				int(s.grid[y+2][l2]) +
-				int(s.grid[y+2][x])*2 +
-				int(s.grid[y+2][r2])
-
-			avg := sum / 9
-
-			// Cooling increases toward the top for natural fadeout
-			coolBase := 2 + (s.height-y)/15
-			cooling := rand.IntN(coolBase + 1)
-			val := avg - cooling
-
-			if val < 0 {
-				val = 0
-			}
-			s.grid[y][x] = uint8(val)
-		}
-	}
-}
-
-func (s *fireState) render(targetW, targetH int) []uint8 {
-	if targetW == 0 || targetH == 0 {
-		return make([]uint8, rgbaChannels)
-	}
-
-	pixels := make([]uint8, targetW*targetH*rgbaChannels)
-
-	// Fire occupies the bottom 2/3 of the window
-	fireStartY := targetH / 3
-
-	fireH := targetH - fireStartY
-	if fireH <= 0 {
-		return pixels
-	}
-
-	for py := fireStartY; py < targetH; py++ {
-		for px := 0; px < targetW; px++ {
-			// Map to fire grid with bilinear interpolation
-			fy := float64(py-fireStartY) * float64(s.height-1) / float64(fireH-1)
-			fx := float64(px) * float64(s.width-1) / float64(targetW-1)
-
-			val := s.sampleBilinear(fx, fy)
-
-			r, g, b, a := fireColor(val)
-			offset := (py*targetW + px) * rgbaChannels
-			pixels[offset] = r
-			pixels[offset+1] = g
-			pixels[offset+2] = b
-			pixels[offset+3] = a
-		}
-	}
-
-	return pixels
-}
-
-// sampleBilinear performs bilinear interpolation on the fire grid for smooth scaling.
-func (s *fireState) sampleBilinear(fx, fy float64) uint8 {
-	x0 := int(fx)
-	y0 := int(fy)
-	x1 := min(x0+1, s.width-1)
-	y1 := min(y0+1, s.height-1)
-
-	xFrac := fx - float64(x0)
-	yFrac := fy - float64(y0)
-
-	// Four corners
-	v00 := float64(s.grid[y0][x0])
-	v10 := float64(s.grid[y0][x1])
-	v01 := float64(s.grid[y1][x0])
-	v11 := float64(s.grid[y1][x1])
-
-	// Interpolate
-	top := v00*(1-xFrac) + v10*xFrac
-	bottom := v01*(1-xFrac) + v11*xFrac
-	val := top*(1-yFrac) + bottom*yFrac
-
-	return uint8(min(max(int(val), 0), 255))
-}
-
-// fireColor maps a heat value (0-255) to a realistic fire palette.
-// Gradient: transparent → dark red/brown → red → orange → gold → pale yellow
-func fireColor(val uint8) (r, g, b, a uint8) {
-	if val < 24 {
-		return 0, 0, 0, 0
-	}
-
-	// Normalize to 0.0-1.0 range (24-255 → 0.0-1.0)
-	t := float64(val-24) / 231.0
-
-	// Piecewise palette for realistic fire
-	switch {
-	case t < 0.2:
-		// Black → dark maroon/brown
-		p := t / 0.2
-		r = uint8(p * 80)
-		g = uint8(p * 10)
-		a = uint8(p * 180)
-		return r, g, 0, a
-	case t < 0.45:
-		// Dark maroon → bright red
-		p := (t - 0.2) / 0.25
-		r = uint8(80 + p*175)
-		g = uint8(10 + p*20)
-		a = uint8(180 + p*75)
-		return r, g, 0, a
-	case t < 0.7:
-		// Red → orange
-		p := (t - 0.45) / 0.25
-		r = 255
-		g = uint8(30 + p*170)
-		a = 255
-		return r, g, 0, a
-	case t < 0.9:
-		// Orange → golden yellow
-		p := (t - 0.7) / 0.2
-		r = 255
-		g = uint8(200 + p*55)
-		b = uint8(p * 30)
-		a = 255
-		return r, g, b, a
-	default:
-		// Golden yellow → pale yellow/white tips
-		p := (t - 0.9) / 0.1
-		r = 255
-		g = 255
-		b = uint8(30 + p*120)
-		a = uint8(255 - p*80)
-		return r, g, b, a
-	}
-}
-
-// --- Snow Effect ---
-
-const (
-	snowFlakeCount = 150
-	snowMaxSize    = 4
-)
-
-type snowflake struct {
-	x, y   float64
-	size   float64
-	speed  float64
-	drift  float64
-	wobble float64
-	phase  float64
-}
-
-type snowState struct {
-	flakes      []snowflake
-	width       int
-	height      int
-	initialized bool
-}
-
-func (p *shenanigans) startSnow(pc linkquisition.PickerCanvas) {
-	state := &snowState{
-		width:  pc.Width(),
-		height: pc.Height(),
-	}
-	if state.width == 0 {
-		state.width = 600
-	}
-	if state.height == 0 {
-		state.height = 400
-	}
-
-	pc.AddRasterOverlay(0.3, func(w, h int) []uint8 {
-		if !state.initialized || w != state.width || h != state.height {
-			state.width = w
-			state.height = h
-			state.init()
-			state.initialized = true
-		}
-		return p.invertForLight(state.render())
-	})
-
-	go func() {
-		ticker := time.NewTicker(frameInterval)
-		defer ticker.Stop()
-
-		for range ticker.C {
-			if p.stopped.Load() {
-				return
-			}
-			state.update()
-			pc.ScheduleRefresh()
-		}
-	}()
-}
-
-func (s *snowState) init() {
-	s.flakes = make([]snowflake, snowFlakeCount)
-	for i := range s.flakes {
-		// Start most flakes above the window so they drift in gradually.
-		// A few start on-screen so it's not completely empty at first.
-		onScreen := i < snowFlakeCount/5
-		s.flakes[i] = s.newFlake(onScreen)
-	}
-}
-
-func (s *snowState) newFlake(onScreen bool) snowflake {
-	// Default: spawn above the viewport at varying distances
-	y := -(rand.Float64() * float64(s.height))
-	if onScreen {
-		y = rand.Float64() * float64(s.height)
-	}
-
-	return snowflake{
-		x:      rand.Float64() * float64(s.width),
-		y:      y,
-		size:   1 + rand.Float64()*float64(snowMaxSize-1),
-		speed:  0.5 + rand.Float64()*2.0,
-		drift:  (rand.Float64() - 0.5) * 0.3,
-		wobble: 0.3 + rand.Float64()*0.7,
-		phase:  rand.Float64() * 6.28,
-	}
-}
-
-func (s *snowState) update() {
-	for i := range s.flakes {
-		f := &s.flakes[i]
-		f.y += f.speed
-		f.phase += 0.05
-
-		// Gentle sine-wave wobble for horizontal drift
-		f.x += f.drift + f.wobble*sinApprox(f.phase)*0.3
-
-		// Respawn at top if fallen below window
-		if f.y > float64(s.height)+10 {
-			*f = s.newFlake(false)
-		}
-
-		// Wrap horizontally
-		if f.x < 0 {
-			f.x += float64(s.width)
-		} else if f.x >= float64(s.width) {
-			f.x -= float64(s.width)
-		}
-	}
-}
-
-func (s *snowState) render() []uint8 {
-	w, h := s.width, s.height
-	if w == 0 || h == 0 {
-		return make([]uint8, rgbaChannels)
-	}
-
-	pixels := make([]uint8, w*h*rgbaChannels)
-
-	for _, f := range s.flakes {
-		s.drawFlake(pixels, f)
-	}
-
-	return pixels
-}
-
-func (s *snowState) drawFlake(pixels []uint8, f snowflake) {
-	w, h := s.width, s.height
-	cx, cy := int(f.x), int(f.y)
-	radius := int(f.size)
-
-	// Draw a soft circle with alpha falloff
-	for dy := -radius; dy <= radius; dy++ {
-		for dx := -radius; dx <= radius; dx++ {
-			px := cx + dx
-			py := cy + dy
-
-			if px < 0 || px >= w || py < 0 || py >= h {
-				continue
-			}
-
-			// Distance from center (0.0 to 1.0+)
-			dist := float64(dx*dx+dy*dy) / float64(radius*radius+1)
-			if dist > 1.0 {
-				continue
-			}
-
-			// Soft edge: alpha falls off near the border
-			alpha := uint8((1.0 - dist*dist) * 220)
-
-			offset := (py*w + px) * rgbaChannels
-			// Blend: white snowflake with alpha
-			existing := pixels[offset+3]
-			if alpha > existing {
-				pixels[offset] = 255   // R
-				pixels[offset+1] = 255 // G
-				pixels[offset+2] = 255 // B
-				pixels[offset+3] = alpha
-			}
-		}
-	}
-}
-
-// sinApprox is a fast sine approximation (Bhaskara I's formula) avoiding math import.
-func sinApprox(x float64) float64 {
-	// Normalize to [0, 2π)
-	const twoPi = 6.283185307
-	const pi = 3.141592654
-
-	for x < 0 {
-		x += twoPi
-	}
-	for x >= twoPi {
-		x -= twoPi
-	}
-
-	// Map to [0, π] with sign
-	sign := 1.0
-	if x > pi {
-		x -= pi
-		sign = -1.0
-	}
-
-	// Bhaskara I's approximation: sin(x) ≈ 16x(π-x) / (5π²-4x(π-x))
-	num := 16 * x * (pi - x)
-	den := 5*pi*pi - 4*x*(pi-x)
-	return sign * num / den
-}
-
-// --- Plasma Effect ---
-
-type plasmaState struct {
-	time float64
-}
-
-func (p *shenanigans) startPlasma(pc linkquisition.PickerCanvas) {
-	state := &plasmaState{}
-
-	pc.AddRasterOverlay(0.5, func(w, h int) []uint8 {
-		return state.render(w, h)
-	})
-
-	go func() {
-		ticker := time.NewTicker(frameInterval)
-		defer ticker.Stop()
-
-		for range ticker.C {
-			if p.stopped.Load() {
-				return
-			}
-			state.time += 0.06
-			pc.ScheduleRefresh()
-		}
-	}()
-}
-
-func (s *plasmaState) render(w, h int) []uint8 {
-	if w == 0 || h == 0 {
-		return make([]uint8, rgbaChannels)
-	}
-
-	pixels := make([]uint8, w*h*rgbaChannels)
-	t := s.time
-
-	for py := 0; py < h; py++ {
-		fy := float64(py) / float64(h)
-		for px := 0; px < w; px++ {
-			fx := float64(px) / float64(w)
-
-			// Overlapping sine waves at different frequencies and phases
-			v1 := sinApprox((fx*4 + t) * 3.14159)
-			v2 := sinApprox((fy*4 + t*0.7) * 3.14159)
-			v3 := sinApprox(((fx+fy)*3 + t*1.3) * 3.14159)
-			v4 := sinApprox(((fx-fy)*2 + t*0.5) * 3.14159)
-			v5 := sinApprox(((fx*fx+fy*fy)*4 - t*0.9) * 3.14159)
-
-			// Combine waves (result in -1 to 1 range, normalize to 0-1)
-			val := (v1 + v2 + v3 + v4 + v5) / 5.0
-			val = (val + 1.0) / 2.0
-
-			// Map to color using three phase-shifted sine waves for RGB
-			r := uint8(sinNorm(val*3.14159*2+t*0.3) * 255)
-			g := uint8(sinNorm(val*3.14159*2+t*0.3+2.09) * 255)
-			b := uint8(sinNorm(val*3.14159*2+t*0.3+4.19) * 255)
-
-			offset := (py*w + px) * rgbaChannels
-			pixels[offset] = r
-			pixels[offset+1] = g
-			pixels[offset+2] = b
-			pixels[offset+3] = 200
-		}
-	}
-
-	return pixels
-}
-
-// --- Starfield Effect ---
-
-const starCount = 200
-
-type star struct {
-	x, y, z float64
-}
-
-type starfieldState struct {
-	stars  []star
-	width  int
-	height int
-}
-
-func (p *shenanigans) startStarfield(pc linkquisition.PickerCanvas) {
-	state := &starfieldState{}
-
-	pc.AddRasterOverlay(0.2, func(w, h int) []uint8 {
-		if !state.isInitialized() || w != state.width || h != state.height {
-			state.width = w
-			state.height = h
-			state.init()
-		}
-		return p.invertForLight(state.render())
-	})
-
-	go func() {
-		ticker := time.NewTicker(frameInterval)
-		defer ticker.Stop()
-
-		for range ticker.C {
-			if p.stopped.Load() {
-				return
-			}
-			state.update()
-			pc.ScheduleRefresh()
-		}
-	}()
-}
-
-func (s *starfieldState) isInitialized() bool {
-	return len(s.stars) > 0
-}
-
-func (s *starfieldState) init() {
-	s.stars = make([]star, starCount)
-	for i := range s.stars {
-		s.stars[i] = s.newStar(true)
-	}
-}
-
-func (s *starfieldState) newStar(randomDepth bool) star {
-	z := 0.01 + rand.Float64()*0.99
-	if randomDepth {
-		z = 0.1 + rand.Float64()*0.9
-	}
-
-	return star{
-		x: (rand.Float64() - 0.5) * 2.0,
-		y: (rand.Float64() - 0.5) * 2.0,
-		z: z,
-	}
-}
-
-func (s *starfieldState) update() {
-	for i := range s.stars {
-		s.stars[i].z -= 0.015
-
-		// Respawn stars that have passed the viewer
-		if s.stars[i].z <= 0.001 {
-			s.stars[i] = s.newStar(false)
-			s.stars[i].z = 0.9 + rand.Float64()*0.1
-		}
-	}
-}
-
-func (s *starfieldState) render() []uint8 {
-	w, h := s.width, s.height
-	if w == 0 || h == 0 {
-		return make([]uint8, rgbaChannels)
-	}
-
-	pixels := make([]uint8, w*h*rgbaChannels)
-	cx := float64(w) / 2.0
-	cy := float64(h) / 2.0
-
-	for _, st := range s.stars {
-		// Perspective projection
-		screenX := int(cx + (st.x/st.z)*cx)
-		screenY := int(cy + (st.y/st.z)*cy)
-
-		if screenX < 0 || screenX >= w || screenY < 0 || screenY >= h {
-			continue
-		}
-
-		// Size and brightness increase as stars get closer (z → 0)
-		brightness := uint8(min(int((1.0-st.z)*255), 255))
-		size := int(1 + (1.0-st.z)*3)
-
-		s.drawStar(pixels, screenX, screenY, size, brightness)
-	}
-
-	return pixels
-}
-
-func (s *starfieldState) drawStar(pixels []uint8, screenX, screenY, size int, brightness uint8) {
-	w, h := s.width, s.height
-	for dy := -size; dy <= size; dy++ {
-		for dx := -size; dx <= size; dx++ {
-			px := screenX + dx
-			py := screenY + dy
-
-			if px < 0 || px >= w || py < 0 || py >= h {
-				continue
-			}
-
-			dist := dx*dx + dy*dy
-			maxDist := size * size
-			if dist > maxDist {
-				continue
-			}
-
-			// Alpha falls off with distance from center
-			falloff := 1.0 - float64(dist)/float64(maxDist+1)
-			alpha := uint8(float64(brightness) * falloff)
-
-			offset := (py*w + px) * rgbaChannels
-			if alpha > pixels[offset+3] {
-				pixels[offset] = brightness
-				pixels[offset+1] = brightness
-				pixels[offset+2] = 255 // slight blue tint
-				pixels[offset+3] = alpha
-			}
-		}
-	}
-}
-
-// --- Aurora Effect ---
-
-const auroraLayers = 4
-
-type auroraState struct {
-	time float64
-}
-
-func (p *shenanigans) startAurora(pc linkquisition.PickerCanvas) {
-	state := &auroraState{}
-
-	pc.AddRasterOverlay(0.4, func(w, h int) []uint8 {
-		return state.render(w, h)
-	})
-
-	go func() {
-		ticker := time.NewTicker(frameInterval)
-		defer ticker.Stop()
-
-		for range ticker.C {
-			if p.stopped.Load() {
-				return
-			}
-			state.time += 0.03
-			pc.ScheduleRefresh()
-		}
-	}()
-}
-
-func (s *auroraState) render(w, h int) []uint8 {
-	if w == 0 || h == 0 {
-		return make([]uint8, rgbaChannels)
-	}
-
-	pixels := make([]uint8, w*h*rgbaChannels)
-	t := s.time
-
-	// Aurora occupies the top 2/3 of the window
-	auroraEndY := h * 2 / 3
-
-	for py := 0; py < auroraEndY; py++ {
-		// Vertical position normalized (0 at top, 1 at aurora bottom)
-		fy := float64(py) / float64(auroraEndY)
-
-		for px := 0; px < w; px++ {
-			fx := float64(px) / float64(w)
-
-			// Combine multiple curtain layers
-			var intensity float64
-			for layer := range auroraLayers {
-				fl := float64(layer)
-				// Each layer has different frequency, speed, and phase
-				wave := sinApprox((fx*(3+fl) + t*(0.4+fl*0.15) + fl*1.7) * 3.14159)
-				wave2 := sinApprox((fx*(2+fl*0.7) - t*(0.3+fl*0.1) + fl*2.3) * 3.14159)
-
-				// Curtain shape: thin band that undulates
-				curtainCenter := 0.2 + 0.15*fl + 0.1*(wave*0.5+0.5)
-				curtainWidth := 0.08 + 0.04*wave2
-
-				// Gaussian-like falloff from the curtain center
-				dist := (fy - curtainCenter) / curtainWidth
-				layerIntensity := fastExp(-dist * dist * 0.5)
-
-				intensity += layerIntensity * (0.6 + 0.4/(fl+1))
-			}
-
-			if intensity < 0.01 {
-				continue
-			}
-			if intensity > 1.0 {
-				intensity = 1.0
-			}
-
-			// Aurora color: shift from green to purple/blue based on position and time
-			colorPhase := fx*0.5 + fy*0.3 + t*0.1
-			r, g, b := auroraColor(colorPhase, intensity)
-
-			alpha := uint8(intensity * 180)
-
-			offset := (py*w + px) * rgbaChannels
-			pixels[offset] = r
-			pixels[offset+1] = g
-			pixels[offset+2] = b
-			pixels[offset+3] = alpha
-		}
-	}
-
-	return pixels
-}
-
-// auroraColor maps a phase and intensity to northern lights colors.
-// Shifts between green, teal, blue, and purple.
-func auroraColor(phase, intensity float64) (r, g, b uint8) {
-	// Cycle through aurora palette
-	p := sinApprox(phase * 3.14159 * 2)
-	p = (p + 1.0) / 2.0 // normalize to 0-1
-
-	// Blend between green-dominant and purple-dominant
-	var rf, gf, bf float64
-	switch {
-	case p < 0.33:
-		// Green to teal
-		t := p / 0.33
-		rf = 0.1 * t
-		gf = 0.8 + 0.2*t
-		bf = 0.2 + 0.5*t
-	case p < 0.66:
-		// Teal to purple
-		t := (p - 0.33) / 0.33
-		rf = 0.1 + 0.5*t
-		gf = 1.0 - 0.6*t
-		bf = 0.7 + 0.3*t
-	default:
-		// Purple back to green
-		t := (p - 0.66) / 0.34
-		rf = 0.6 - 0.5*t
-		gf = 0.4 + 0.4*t
-		bf = 1.0 - 0.8*t
-	}
-
-	r = uint8(rf * intensity * 255)
-	g = uint8(gf * intensity * 255)
-	b = uint8(bf * intensity * 255)
-	return r, g, b
-}
-
-// fastExp approximates e^x for negative x values (used for Gaussian falloff).
-func fastExp(x float64) float64 {
-	if x < -6 {
-		return 0
-	}
-	// Padé approximation: (1 + x/n)^n for small |x|
-	// Using n=8 for reasonable accuracy
-	t := 1.0 + x/8.0
-	t *= t // ^2
-	t *= t // ^4
-	t *= t // ^8
-	if t < 0 {
-		return 0
-	}
-	return t
-}
-
-// --- Glitch Effect ---
-
-type glitchState struct {
-	frame      int
-	slices     []glitchSlice
-	burstTimer int
-	isBursting bool
-}
-
-type glitchSlice struct {
-	y, height int
-	offsetX   int
-	channel   int // 0=R shift, 1=G shift, 2=B shift
-	alpha     uint8
-}
-
-func (p *shenanigans) startGlitch(pc linkquisition.PickerCanvas) {
-	state := &glitchState{}
-
-	pc.AddRasterOverlay(0.0, func(w, h int) []uint8 {
-		return state.render(w, h)
-	})
-
-	go func() {
-		ticker := time.NewTicker(frameInterval)
-		defer ticker.Stop()
-
-		for range ticker.C {
-			if p.stopped.Load() {
-				return
-			}
-			state.update()
-			pc.ScheduleRefresh()
-		}
-	}()
-}
-
-func (s *glitchState) update() {
-	s.frame++
-	s.burstTimer--
-
-	// Trigger glitch bursts periodically
-	if s.burstTimer <= 0 {
-		if s.isBursting {
-			// End burst
-			s.isBursting = false
-			s.slices = nil
-			s.burstTimer = 20 + rand.IntN(40)
-		} else {
-			// Start burst
-			s.isBursting = true
-			s.burstTimer = 3 + rand.IntN(8)
-			s.generateSlices()
-		}
-	} else if s.isBursting && s.frame%2 == 0 {
-		// Regenerate slices during burst for flicker
-		s.generateSlices()
-	}
-}
-
-func (s *glitchState) generateSlices() {
-	count := 3 + rand.IntN(8)
-	s.slices = make([]glitchSlice, count)
-
-	for i := range s.slices {
-		s.slices[i] = glitchSlice{
-			y:       rand.IntN(400),
-			height:  2 + rand.IntN(20),
-			offsetX: -30 + rand.IntN(60),
-			channel: rand.IntN(3),
-			alpha:   uint8(80 + rand.IntN(176)),
-		}
-	}
-}
-
-func (s *glitchState) render(w, h int) []uint8 {
-	if w == 0 || h == 0 {
-		return make([]uint8, rgbaChannels)
-	}
-
-	pixels := make([]uint8, w*h*rgbaChannels)
-
-	if !s.isBursting {
-		return pixels
-	}
-
-	// Draw glitch slices
-	for _, slice := range s.slices {
-		s.drawSlice(pixels, slice, w, h)
-	}
-
-	// Add random static noise during bursts
-	s.addNoise(pixels, w, h)
-
-	return pixels
-}
-
-func (s *glitchState) drawSlice(pixels []uint8, slice glitchSlice, w, h int) {
-	sy := slice.y * h / 400
-	sh := slice.height * h / 400
-
-	for py := sy; py < sy+sh && py < h; py++ {
-		if py < 0 {
-			continue
-		}
-		for px := 0; px < w; px++ {
-			offset := (py*w + px) * rgbaChannels
-
-			// RGB channel separation effect
-			switch slice.channel {
-			case 0: // Red shift
-				srcX := px - slice.offsetX
-				if srcX >= 0 && srcX < w {
-					pixels[offset] = slice.alpha
-					pixels[offset+3] = slice.alpha / 2
-				}
-			case 1: // Green shift
-				srcX := px + slice.offsetX
-				if srcX >= 0 && srcX < w {
-					pixels[offset+1] = slice.alpha
-					pixels[offset+3] = slice.alpha / 2
-				}
-			default: // Blue/cyan shift
-				pixels[offset+2] = slice.alpha
-				pixels[offset+1] = slice.alpha / 3
-				pixels[offset+3] = slice.alpha / 2
-			}
-		}
-	}
-}
-
-func (s *glitchState) addNoise(pixels []uint8, w, h int) {
-	noiseCount := w * h / 40
-	for range noiseCount {
-		px := rand.IntN(w)
-		py := rand.IntN(h)
-		offset := (py*w + px) * rgbaChannels
-		v := uint8(rand.IntN(256))
-		pixels[offset] = v
-		pixels[offset+1] = v
-		pixels[offset+2] = v
-		pixels[offset+3] = uint8(rand.IntN(100))
-	}
-}
-
-// --- Pride Effect ---
-
-type prideState struct {
-	time float64
-}
-
-// Pride flag colors (6-stripe rainbow)
-var prideColors = [][3]uint8{
-	{228, 3, 3},   // Red
-	{255, 140, 0}, // Orange
-	{255, 237, 0}, // Yellow
-	{0, 128, 38},  // Green
-	{0, 77, 255},  // Blue
-	{117, 7, 135}, // Purple
-}
-
-func (p *shenanigans) startPride(pc linkquisition.PickerCanvas) {
-	state := &prideState{}
-
-	pc.AddRasterOverlay(0.45, func(w, h int) []uint8 {
-		return state.render(w, h)
-	})
-
-	go func() {
-		ticker := time.NewTicker(frameInterval)
-		defer ticker.Stop()
-
-		for range ticker.C {
-			if p.stopped.Load() {
-				return
-			}
-			state.time += 0.04
-			pc.ScheduleRefresh()
-		}
-	}()
-}
-
-func (s *prideState) render(w, h int) []uint8 {
-	if w == 0 || h == 0 {
-		return make([]uint8, rgbaChannels)
-	}
-
-	pixels := make([]uint8, w*h*rgbaChannels)
-	t := s.time
-	stripeCount := float64(len(prideColors))
-
-	for py := 0; py < h; py++ {
-		for px := 0; px < w; px++ {
-			fx := float64(px) / float64(w)
-			fy := float64(py) / float64(h)
-
-			// Flag wave: pinned on the left edge, wave amplitude increases to the right
-			// This simulates fabric attached to a pole on the left side
-			amplitude := fx * fx * 0.12
-			wave := sinApprox((fx*2.0-t*1.5)*3.14159) * amplitude
-			wave2 := sinApprox((fx*3.0-t*2.0)*3.14159) * amplitude * 0.5
-
-			// Determine which stripe this pixel belongs to (with wave offset)
-			stripePos := (fy + wave + wave2) * stripeCount
-			stripeIdx := int(stripePos)
-
-			if stripeIdx < 0 {
-				stripeIdx = 0
-			}
-			if stripeIdx >= len(prideColors) {
-				stripeIdx = len(prideColors) - 1
-			}
-
-			// Smooth blending between stripes
-			blend := stripePos - float64(stripeIdx)
-			nextIdx := stripeIdx + 1
-			if nextIdx >= len(prideColors) {
-				nextIdx = len(prideColors) - 1
-			}
-
-			c1 := prideColors[stripeIdx]
-			c2 := prideColors[nextIdx]
-
-			// Smooth interpolation (smoothstep-like)
-			blend = blend * blend * (3 - 2*blend)
-
-			r := uint8(float64(c1[0])*(1-blend) + float64(c2[0])*blend)
-			g := uint8(float64(c1[1])*(1-blend) + float64(c2[1])*blend)
-			b := uint8(float64(c1[2])*(1-blend) + float64(c2[2])*blend)
-
-			// Subtle shading to simulate fabric folds (stronger toward free edge)
-			foldDepth := fx * 0.2
-			shade := 1.0 - foldDepth + foldDepth*sinApprox((fx*3-t*1.5)*3.14159)
-			r = uint8(float64(r) * shade)
-			g = uint8(float64(g) * shade)
-			b = uint8(float64(b) * shade)
-
-			offset := (py*w + px) * rgbaChannels
-			pixels[offset] = r
-			pixels[offset+1] = g
-			pixels[offset+2] = b
-			pixels[offset+3] = 200
-		}
-	}
-
-	return pixels
-}
-
-// --- Football Effect ---
-
-type footballState struct {
-	time   float64
-	width  int
-	height int
-}
-
-func (p *shenanigans) startFootball(pc linkquisition.PickerCanvas) {
-	state := &footballState{}
-
-	pc.AddRasterOverlay(0.4, func(w, h int) []uint8 {
-		state.width = w
-		state.height = h
-		return state.render()
-	})
-
-	go func() {
-		ticker := time.NewTicker(frameInterval)
-		defer ticker.Stop()
-
-		for range ticker.C {
-			if p.stopped.Load() {
-				return
-			}
-			state.time += 0.03
-			pc.ScheduleRefresh()
-		}
-	}()
-}
-
-func (s *footballState) render() []uint8 {
-	w, h := s.width, s.height
-	if w == 0 || h == 0 {
-		return make([]uint8, rgbaChannels)
-	}
-
-	pixels := make([]uint8, w*h*rgbaChannels)
-	t := s.time
-
-	for py := 0; py < h; py++ {
-		fy := float64(py) / float64(h)
-		for px := 0; px < w; px++ {
-			fx := float64(px) / float64(w)
-
-			// Green pitch base with subtle stripe pattern (mowed grass look)
-			stripeWidth := 0.08
-			stripe := int(fx/stripeWidth) % 2
-			var gr, gg, gb uint8
-			if stripe == 0 {
-				gr, gg, gb = 34, 139, 34
-			} else {
-				gr, gg, gb = 30, 124, 30
-			}
-
-			// Animated element: a "spotlight" sweeping across the pitch
-			spotX := 0.5 + 0.4*sinApprox(t*1.5)
-			spotY := 0.5 + 0.3*sinApprox(t*1.1+1.0)
-			spotDist := (fx-spotX)*(fx-spotX) + (fy-spotY)*(fy-spotY)
-			spotLight := fastExp(-spotDist*15) * 0.3
-
-			var r, g, b uint8
-			if isPitchLine(fx, fy) {
-				r, g, b = 255, 255, 255
-			} else {
-				r = uint8(min(int(float64(gr)*(1+spotLight)), 255))
-				g = uint8(min(int(float64(gg)*(1+spotLight)), 255))
-				b = uint8(min(int(float64(gb)*(1+spotLight)), 255))
-			}
-
-			offset := (py*w + px) * rgbaChannels
-			pixels[offset] = r
-			pixels[offset+1] = g
-			pixels[offset+2] = b
-			pixels[offset+3] = 180
-		}
-	}
-
-	return pixels
-}
-
-// isPitchLine determines if a normalized coordinate is on a white pitch marking.
-func isPitchLine(fx, fy float64) bool {
-	return isPitchCenterMarkings(fx, fy) ||
-		isPitchBoundary(fx, fy) ||
-		isPitchPenaltyArea(fx, fy)
-}
-
-func isPitchCenterMarkings(fx, fy float64) bool {
-	// Center line (vertical)
-	if absF(fx-0.5) < 0.004 {
-		return true
-	}
-
-	// Center circle
-	cx, cy := 0.5, 0.5
-	dist := (fx-cx)*(fx-cx)*1.5 + (fy-cy)*(fy-cy)
-	if absF(dist-0.04) < 0.003 {
-		return true
-	}
-
-	// Center dot
-	return dist < 0.002
-}
-
-func isPitchBoundary(fx, fy float64) bool {
-	if fx < 0.02 || fx > 0.98 || fy < 0.03 || fy > 0.97 {
-		if fx > 0.015 && fx < 0.985 && fy > 0.025 && fy < 0.975 {
-			return true
-		}
-	}
-	return false
-}
-
-func isPitchPenaltyArea(fx, fy float64) bool {
-	penaltyW := 0.15
-	penaltyH := 0.35
-	penaltyTop := 0.5 - penaltyH
-	penaltyBot := 0.5 + penaltyH
-
-	// Left penalty area
-	if fx < penaltyW && fy > penaltyTop && fy < penaltyBot {
-		if absF(fx-penaltyW) < 0.004 || absF(fy-penaltyTop) < 0.005 || absF(fy-penaltyBot) < 0.005 {
-			return true
-		}
-	}
-
-	// Right penalty area
-	if fx > (1-penaltyW) && fy > penaltyTop && fy < penaltyBot {
-		if absF(fx-(1-penaltyW)) < 0.004 || absF(fy-penaltyTop) < 0.005 || absF(fy-penaltyBot) < 0.005 {
-			return true
-		}
-	}
-
-	return false
-}
-
-// --- Fireworks Effect ---
-
-const (
-	fireworksMaxRockets   = 5
-	fireworksParticles    = 60
-	fireworksLaunchChance = 8 // percent chance per frame to launch a new rocket
-)
-
-type fireworksParticle struct {
-	x, y    float64
-	vx, vy  float64
-	life    float64
-	r, g, b uint8
-}
-
-type fireworksRocket struct {
-	particles []fireworksParticle
-	exploded  bool
-	// Pre-explosion rocket position
-	x, y    float64
-	vy      float64
-	targetY float64
-	color   [3]uint8
-}
-
-type fireworksState struct {
-	rockets []fireworksRocket
-	width   int
-	height  int
-}
-
-var fireworksColors = [][3]uint8{
-	{255, 200, 50},  // Gold
-	{255, 80, 80},   // Red
-	{80, 150, 255},  // Blue
-	{80, 255, 80},   // Green
-	{200, 100, 255}, // Purple
-	{255, 150, 200}, // Pink
-	{255, 255, 255}, // White
-}
-
-func (p *shenanigans) startFireworks(pc linkquisition.PickerCanvas) {
-	state := &fireworksState{}
-
-	pc.AddRasterOverlay(0.2, func(w, h int) []uint8 {
-		state.width = w
-		state.height = h
-		return state.render()
-	})
-
-	go func() {
-		ticker := time.NewTicker(frameInterval)
-		defer ticker.Stop()
-
-		for range ticker.C {
-			if p.stopped.Load() {
-				return
-			}
-			state.update()
-			pc.ScheduleRefresh()
-		}
-	}()
-}
-
-func (s *fireworksState) update() {
-	if s.width == 0 || s.height == 0 {
-		return
-	}
-
-	// Chance to launch a new rocket
-	if len(s.rockets) < fireworksMaxRockets && rand.IntN(100) < fireworksLaunchChance {
-		color := fireworksColors[rand.IntN(len(fireworksColors))]
-		s.rockets = append(s.rockets, fireworksRocket{
-			x:       0.2 + rand.Float64()*0.6,
-			y:       1.0,
-			vy:      -0.025 - rand.Float64()*0.015,
-			targetY: 0.15 + rand.Float64()*0.35,
-			color:   color,
-		})
-	}
-
-	// Update rockets
-	alive := s.rockets[:0]
-	for i := range s.rockets {
-		r := &s.rockets[i]
-
-		if !r.exploded {
-			r.y += r.vy
-			// Explode when reaching target height
-			if r.y <= r.targetY {
-				r.exploded = true
-				r.particles = make([]fireworksParticle, fireworksParticles)
-				for j := range r.particles {
-					angle := rand.Float64() * 6.283
-					speed := 0.005 + rand.Float64()*0.015
-					r.particles[j] = fireworksParticle{
-						x:    r.x,
-						y:    r.y,
-						vx:   sinApprox(angle) * speed,
-						vy:   sinApprox(angle+1.5708) * speed,
-						life: 1.0,
-						r:    r.color[0],
-						g:    r.color[1],
-						b:    r.color[2],
-					}
-				}
-			}
-		} else {
-			// Update particles
-			allDead := true
-			for j := range r.particles {
-				p := &r.particles[j]
-				if p.life <= 0 {
-					continue
-				}
-				p.x += p.vx
-				p.y += p.vy
-				p.vy += 0.0004 // gravity
-				p.vx *= 0.98   // drag
-				p.vy *= 0.98
-				p.life -= 0.015
-				if p.life > 0 {
-					allDead = false
-				}
-			}
-			if allDead {
-				continue // don't keep this rocket
-			}
-		}
-		alive = append(alive, *r)
-	}
-	s.rockets = alive
-}
-
-func (s *fireworksState) render() []uint8 {
-	w, h := s.width, s.height
-	if w == 0 || h == 0 {
-		return make([]uint8, rgbaChannels)
-	}
-
-	pixels := make([]uint8, w*h*rgbaChannels)
-
-	for _, rocket := range s.rockets {
-		if !rocket.exploded {
-			// Draw rising rocket as a small bright dot with trail
-			px := int(rocket.x * float64(w))
-			py := int(rocket.y * float64(h))
-			s.drawDot(pixels, px, py, 2, 255, 220, 150, 255)
-			// Trail
-			s.drawDot(pixels, px, py+3, 1, 255, 150, 50, 150)
-			s.drawDot(pixels, px, py+6, 1, 255, 100, 30, 80)
-		} else {
-			// Draw particles
-			for _, p := range rocket.particles {
-				if p.life <= 0 {
-					continue
-				}
-				px := int(p.x * float64(w))
-				py := int(p.y * float64(h))
-				alpha := uint8(p.life * 255)
-				size := 1
-				if p.life > 0.7 {
-					size = 2
-				}
-				s.drawDot(pixels, px, py, size, p.r, p.g, p.b, alpha)
-			}
-		}
-	}
-
-	return pixels
-}
-
-func (s *fireworksState) drawDot(pixels []uint8, cx, cy, radius int, r, g, b, a uint8) {
-	w, h := s.width, s.height
-	for dy := -radius; dy <= radius; dy++ {
-		for dx := -radius; dx <= radius; dx++ {
-			px := cx + dx
-			py := cy + dy
-			if px < 0 || px >= w || py < 0 || py >= h {
-				continue
-			}
-			if dx*dx+dy*dy > radius*radius {
-				continue
-			}
-			offset := (py*w + px) * rgbaChannels
-			if a > pixels[offset+3] {
-				pixels[offset] = r
-				pixels[offset+1] = g
-				pixels[offset+2] = b
-				pixels[offset+3] = a
-			}
-		}
-	}
-}
-
-// absF returns the absolute value of a float64.
-func absF(x float64) float64 {
-	if x < 0 {
-		return -x
-	}
-	return x
-}
-
-// sinNorm returns sinApprox mapped to 0.0-1.0 range.
-func sinNorm(x float64) float64 {
-	return (sinApprox(x) + 1.0) / 2.0
-}
-
-// --- Pong Effect ---
-
-const (
-	pongPaddleWidth  = 6
-	pongPaddleHeight = 40
-	pongBallSize     = 6
-	pongPaddleMargin = 12
-	pongPaddleSpeed  = 3.5
-	pongBallAlpha    = 140
-	pongPaddleAlpha  = 120
-	pongNetAlpha     = 50
-	pongNetDash      = 8
-	pongNetGap       = 6
-)
-
-type pongState struct {
-	width, height int
-
-	// Ball
-	ballX, ballY   float64
-	ballVX, ballVY float64
-
-	// Paddles (y is the center)
-	leftY, rightY float64
-
-	// Score
-	leftScore, rightScore int
-}
-
-func (p *shenanigans) startPong(pc linkquisition.PickerCanvas) {
-	state := &pongState{
-		width:  pc.Width(),
-		height: pc.Height(),
-	}
-	if state.width == 0 {
-		state.width = 600
-	}
-	if state.height == 0 {
-		state.height = 400
-	}
-	state.resetBall()
-
-	pc.AddRasterOverlay(0.4, func(w, h int) []uint8 {
-		if w != state.width || h != state.height {
-			state.width = w
-			state.height = h
-		}
-		return p.invertForLight(state.render())
-	})
-
-	go func() {
-		ticker := time.NewTicker(frameInterval)
-		defer ticker.Stop()
-
-		for range ticker.C {
-			if p.stopped.Load() {
-				return
-			}
-			state.update()
-			pc.ScheduleRefresh()
-		}
-	}()
-}
-
-func (s *pongState) resetBall() {
-	s.ballX = float64(s.width) / 2
-	s.ballY = float64(s.height) / 2
-	s.leftY = float64(s.height) / 2
-	s.rightY = float64(s.height) / 2
-
-	// Random initial direction
-	s.ballVX = 3.0 + rand.Float64()*2.0
-	if rand.IntN(2) == 0 {
-		s.ballVX = -s.ballVX
-	}
-	s.ballVY = (rand.Float64() - 0.5) * 4.0
-}
-
-func (s *pongState) update() { //nolint:gocyclo
-	h := float64(s.height)
-	w := float64(s.width)
-
-	// Move ball
-	s.ballX += s.ballVX
-	s.ballY += s.ballVY
-
-	// Bounce off top/bottom walls
-	if s.ballY <= 0 {
-		s.ballY = -s.ballY
-		s.ballVY = -s.ballVY
-	} else if s.ballY >= h-1 {
-		s.ballY = 2*(h-1) - s.ballY
-		s.ballVY = -s.ballVY
-	}
-
-	// AI for paddles — follow the ball with imperfect tracking
-	s.leftY += clampPaddleMove(s.ballY-s.leftY, pongPaddleSpeed*0.85)
-	s.rightY += clampPaddleMove(s.ballY-s.rightY, pongPaddleSpeed*0.9)
-
-	// Clamp paddles within bounds
-	halfPaddle := float64(pongPaddleHeight) / 2
-	s.leftY = clampFloat(s.leftY, halfPaddle, h-halfPaddle)
-	s.rightY = clampFloat(s.rightY, halfPaddle, h-halfPaddle)
-
-	// Left paddle collision
-	leftPaddleX := float64(pongPaddleMargin + pongPaddleWidth)
-	if s.ballX <= leftPaddleX && s.ballVX < 0 {
-		if s.ballY >= s.leftY-halfPaddle && s.ballY <= s.leftY+halfPaddle {
-			s.ballX = leftPaddleX
-			s.ballVX = -s.ballVX * (1.0 + rand.Float64()*0.1)
-			// Add spin based on where ball hits the paddle
-			offset := (s.ballY - s.leftY) / halfPaddle
-			s.ballVY += offset * 1.5
-		}
-	}
-
-	// Right paddle collision
-	rightPaddleX := w - float64(pongPaddleMargin+pongPaddleWidth)
-	if s.ballX >= rightPaddleX && s.ballVX > 0 {
-		if s.ballY >= s.rightY-halfPaddle && s.ballY <= s.rightY+halfPaddle {
-			s.ballX = rightPaddleX
-			s.ballVX = -s.ballVX * (1.0 + rand.Float64()*0.1)
-			offset := (s.ballY - s.rightY) / halfPaddle
-			s.ballVY += offset * 1.5
-		}
-	}
-
-	// Cap ball speed to prevent it from getting too fast
-	maxSpeed := 8.0
-	if s.ballVX > maxSpeed {
-		s.ballVX = maxSpeed
-	} else if s.ballVX < -maxSpeed {
-		s.ballVX = -maxSpeed
-	}
-	if s.ballVY > maxSpeed {
-		s.ballVY = maxSpeed
-	} else if s.ballVY < -maxSpeed {
-		s.ballVY = -maxSpeed
-	}
-
-	// Score — ball leaves the field
-	if s.ballX < 0 {
-		s.rightScore++
-		s.resetBall()
-	} else if s.ballX > w {
-		s.leftScore++
-		s.resetBall()
-	}
-}
-
-func (s *pongState) render() []uint8 {
-	w, h := s.width, s.height
-	if w == 0 || h == 0 {
-		return make([]uint8, rgbaChannels)
-	}
-
-	pixels := make([]uint8, w*h*rgbaChannels)
-
-	// Draw center net (dashed line)
-	centerX := w / 2
-	for y := 0; y < h; y++ {
-		segment := y % (pongNetDash + pongNetGap)
-		if segment < pongNetDash {
-			s.setPixel(pixels, centerX, y, 255, 255, 255, pongNetAlpha)
-		}
-	}
-
-	// Draw paddles
-	halfPaddle := pongPaddleHeight / 2
-	// Left paddle
-	for dy := -halfPaddle; dy <= halfPaddle; dy++ {
-		for dx := 0; dx < pongPaddleWidth; dx++ {
-			px := pongPaddleMargin + dx
-			py := int(s.leftY) + dy
-			s.setPixel(pixels, px, py, 255, 255, 255, pongPaddleAlpha)
-		}
-	}
-	// Right paddle
-	for dy := -halfPaddle; dy <= halfPaddle; dy++ {
-		for dx := 0; dx < pongPaddleWidth; dx++ {
-			px := w - pongPaddleMargin - pongPaddleWidth + dx
-			py := int(s.rightY) + dy
-			s.setPixel(pixels, px, py, 255, 255, 255, pongPaddleAlpha)
-		}
-	}
-
-	// Draw ball
-	halfBall := pongBallSize / 2
-	for dy := -halfBall; dy <= halfBall; dy++ {
-		for dx := -halfBall; dx <= halfBall; dx++ {
-			px := int(s.ballX) + dx
-			py := int(s.ballY) + dy
-			s.setPixel(pixels, px, py, 255, 255, 255, pongBallAlpha)
-		}
-	}
-
-	// Draw score (simple dot-based digits)
-	s.drawScore(pixels)
-
-	return pixels
-}
-
-func (s *pongState) setPixel(pixels []uint8, x, y int, r, g, b, a uint8) { //nolint:unparam
-	if x < 0 || x >= s.width || y < 0 || y >= s.height {
-		return
-	}
-	offset := (y*s.width + x) * rgbaChannels
-	if a > pixels[offset+3] {
-		pixels[offset] = r
-		pixels[offset+1] = g
-		pixels[offset+2] = b
-		pixels[offset+3] = a
-	}
-}
-
-func (s *pongState) drawScore(pixels []uint8) {
-	// Simple score display near the top center
-	scoreAlpha := uint8(70)
-	y := 15
-	// Left score — draw dots on the left of center
-	cx := s.width/2 - 20
-	for i := range s.leftScore {
-		s.drawDot(pixels, cx-i*8, y, scoreAlpha)
-	}
-	// Right score — draw dots on the right of center
-	cx = s.width/2 + 20
-	for i := range s.rightScore {
-		s.drawDot(pixels, cx+i*8, y, scoreAlpha)
-	}
-}
-
-func (s *pongState) drawDot(pixels []uint8, cx, cy int, alpha uint8) {
-	for dy := -2; dy <= 2; dy++ {
-		for dx := -2; dx <= 2; dx++ {
-			s.setPixel(pixels, cx+dx, cy+dy, 255, 255, 255, alpha)
-		}
-	}
-}
-
-func clampPaddleMove(delta, maxMove float64) float64 {
-	if delta > maxMove {
-		return maxMove
-	}
-	if delta < -maxMove {
-		return -maxMove
-	}
-	return delta
-}
-
-func clampFloat(v, lo, hi float64) float64 {
-	if v < lo {
-		return lo
-	}
-	if v > hi {
-		return hi
-	}
-	return v
-}
-
-// --- Game of Life Effect ---
-
-const (
-	lifeCellSize      = 4
-	lifeAlpha         = 100
-	lifeFadeAlpha     = 40
-	lifeFrameInterval = 80 * time.Millisecond
-	lifeInitDensity   = 0.35
-)
-
-type lifeState struct {
-	width, height int
-	cols, rows    int
-	cells         []bool
-	prev          []bool
-	generation    int
-	staleCount    int
-}
-
-func (p *shenanigans) startLife(pc linkquisition.PickerCanvas) {
-	state := &lifeState{
-		width:  pc.Width(),
-		height: pc.Height(),
-	}
-	if state.width == 0 {
-		state.width = 600
-	}
-	if state.height == 0 {
-		state.height = 400
-	}
-	state.cols = state.width / lifeCellSize
-	state.rows = state.height / lifeCellSize
-	state.randomize()
-
-	pc.AddRasterOverlay(0.5, func(w, h int) []uint8 {
-		if w != state.width || h != state.height {
-			state.width = w
-			state.height = h
-			state.cols = w / lifeCellSize
-			state.rows = h / lifeCellSize
-			state.randomize()
-		}
-		return p.invertForLight(state.render())
-	})
-
-	go func() {
-		ticker := time.NewTicker(lifeFrameInterval)
-		defer ticker.Stop()
-
-		for range ticker.C {
-			if p.stopped.Load() {
-				return
-			}
-			state.step()
-			pc.ScheduleRefresh()
-		}
-	}()
-}
-
-func (s *lifeState) randomize() {
-	size := s.cols * s.rows
-	s.cells = make([]bool, size)
-	s.prev = make([]bool, size)
-	s.generation = 0
-	s.staleCount = 0
-
-	for i := range s.cells {
-		s.cells[i] = rand.Float64() < lifeInitDensity
-	}
-}
-
-func (s *lifeState) step() {
-	size := s.cols * s.rows
-	if size == 0 {
-		return
-	}
-
-	next := make([]bool, size)
-	for row := range s.rows {
-		for col := range s.cols {
-			neighbors := s.countNeighbors(row, col)
-			idx := row*s.cols + col
-			alive := s.cells[idx]
-
-			if alive {
-				next[idx] = neighbors == 2 || neighbors == 3
-			} else {
-				next[idx] = neighbors == 3
-			}
-		}
-	}
-
-	// Detect stale patterns (identical to previous generation = oscillator)
-	if s.cellsEqual(next, s.prev) || s.cellsEqual(next, s.cells) {
-		s.staleCount++
-	} else {
-		s.staleCount = 0
-	}
-
-	copy(s.prev, s.cells)
-	s.cells = next
-	s.generation++
-
-	// Re-seed if the pattern has become static or died out
-	if s.staleCount > 3 || s.countAlive() < size/20 {
-		s.randomize()
-	}
-}
-
-func (s *lifeState) countNeighbors(row, col int) int {
-	count := 0
-	for dr := -1; dr <= 1; dr++ {
-		for dc := -1; dc <= 1; dc++ {
-			if dr == 0 && dc == 0 {
-				continue
-			}
-			// Wrap around edges (toroidal grid)
-			r := (row + dr + s.rows) % s.rows
-			c := (col + dc + s.cols) % s.cols
-			if s.cells[r*s.cols+c] {
-				count++
-			}
-		}
-	}
-	return count
-}
-
-func (s *lifeState) countAlive() int {
-	count := 0
-	for _, alive := range s.cells {
-		if alive {
-			count++
-		}
-	}
-	return count
-}
-
-func (s *lifeState) cellsEqual(a, b []bool) bool {
-	if len(a) != len(b) {
-		return false
-	}
-	for i := range a {
-		if a[i] != b[i] {
-			return false
-		}
-	}
-	return true
-}
-
-func (s *lifeState) render() []uint8 {
-	w, h := s.width, s.height
-	if w == 0 || h == 0 {
-		return make([]uint8, rgbaChannels)
-	}
-
-	pixels := make([]uint8, w*h*rgbaChannels)
-
-	for row := range s.rows {
-		for col := range s.cols {
-			idx := row*s.cols + col
-			alive := s.cells[idx]
-			wasAlive := s.prev[idx]
-
-			if alive {
-				s.drawCell(pixels, col, row, lifeAlpha)
-			} else if wasAlive {
-				// Fading ghost of recently dead cells
-				s.drawCell(pixels, col, row, lifeFadeAlpha)
-			}
-		}
-	}
-
-	return pixels
-}
-
-func (s *lifeState) drawCell(pixels []uint8, col, row int, alpha uint8) {
-	startX := col * lifeCellSize
-	startY := row * lifeCellSize
-
-	// Draw cell with 1px gap for grid appearance
-	for dy := 0; dy < lifeCellSize-1; dy++ {
-		for dx := 0; dx < lifeCellSize-1; dx++ {
-			px := startX + dx
-			py := startY + dy
-			if px >= s.width || py >= s.height {
-				continue
-			}
-			offset := (py*s.width + px) * rgbaChannels
-			if alpha > pixels[offset+3] {
-				pixels[offset] = 180   // R — slight green tint
-				pixels[offset+1] = 255 // G
-				pixels[offset+2] = 180 // B
-				pixels[offset+3] = alpha
-			}
-		}
-	}
-}
-
-// --- Space Invaders Effect ---
-
-const (
-	invRows            = 4
-	invCols            = 8
-	invMoveInterval    = 12 // frames between invader movements
-	invShootChance     = 0.015
-	invBulletSpeed     = 4.0
-	invPlayerBulletSpd = -5.0
-	invBulletWidth     = 2
-	invBulletHeight    = 5
-	invAlpha           = 120
-	invBulletAlpha     = 100
-	invExplosionFrames = 8
-)
-
-// Classic 8x8 invader sprites (1 = filled pixel in the sprite grid)
-var invaderSprites = [3][8]uint8{
-	// Squid (top rows)
-	{0x18, 0x3C, 0x7E, 0xDB, 0xFF, 0x24, 0x5A, 0xA5},
-	// Crab (middle rows)
-	{0x24, 0x3C, 0x7E, 0xDB, 0xFF, 0x7E, 0x24, 0x42},
-	// Octopus (bottom rows)
-	{0x18, 0x3C, 0x7E, 0xDB, 0xFF, 0xBD, 0x24, 0x42},
-}
-
-type invaderEntity struct {
-	alive bool
-	x, y  float64
-	kind  int // 0=squid, 1=crab, 2=octopus
-}
-
-type invBullet struct {
-	x, y float64
-	vy   float64
-}
-
-type invExplosion struct {
-	x, y   float64
-	frames int
-}
-
-type invadersState struct {
-	width, height int
-
-	invaders   []invaderEntity
-	bullets    []invBullet
-	explosions []invExplosion
-
-	// Movement
-	moveDir    float64 // +1 right, -1 left
-	moveTimer  int
-	dropNext   bool
-	moveAmount float64
-
-	// Player
-	playerX     float64
-	playerPhase float64 // sine oscillation phase around invader center
-
-	// Dynamic scale (computed from window size)
-	pixelSize    int
-	spacingX     int
-	spacingY     int
-	playerWidth  int
-	playerHeight int
-
-	// Game state
-	frameCount int
-}
-
-func (p *shenanigans) startInvaders(pc linkquisition.PickerCanvas) {
-	state := &invadersState{
-		width:  pc.Width(),
-		height: pc.Height(),
-	}
-	if state.width == 0 {
-		state.width = 600
-	}
-	if state.height == 0 {
-		state.height = 400
-	}
-	state.reset()
-
-	pc.AddRasterOverlay(0.4, func(w, h int) []uint8 {
-		if w != state.width || h != state.height {
-			state.width = w
-			state.height = h
-			state.reset()
-		}
-		return p.invertForLight(state.render())
-	})
-
-	go func() {
-		ticker := time.NewTicker(frameInterval)
-		defer ticker.Stop()
-
-		for range ticker.C {
-			if p.stopped.Load() {
-				return
-			}
-			state.update()
-			pc.ScheduleRefresh()
-		}
-	}()
-}
-
-func (s *invadersState) reset() {
-	// Compute scale from window dimensions — fill ~70% of width with the formation
-	formationWidthCells := invCols * 10 // 8px sprite + 2px gap per invader in cell units
-	s.pixelSize = s.width * 7 / (formationWidthCells * 10)
-	if s.pixelSize < 2 {
-		s.pixelSize = 2
-	}
-	if s.pixelSize > 6 {
-		s.pixelSize = 6
-	}
-	s.spacingX = s.pixelSize*8 + s.pixelSize*2 // sprite width + gap
-	s.spacingY = s.pixelSize*8 + s.pixelSize*2
-	s.playerWidth = s.pixelSize * 6
-	s.playerHeight = s.pixelSize * 3
-
-	s.invaders = make([]invaderEntity, invRows*invCols)
-	startX := float64(s.width)/2 - float64(invCols*s.spacingX)/2
-	startY := float64(s.height) * 0.08
-
-	for row := range invRows {
-		for col := range invCols {
-			idx := row*invCols + col
-			s.invaders[idx] = invaderEntity{
-				alive: true,
-				x:     startX + float64(col*s.spacingX),
-				y:     startY + float64(row*s.spacingY),
-				kind:  row % 3,
-			}
-		}
-	}
-
-	s.bullets = nil
-	s.explosions = nil
-	s.moveDir = 1
-	s.moveTimer = 0
-	s.dropNext = false
-	s.moveAmount = float64(s.pixelSize) * 0.7
-	s.playerX = float64(s.width) / 2
-	s.playerPhase = 0
-	s.frameCount = 0
-}
-
-func (s *invadersState) update() {
-	s.frameCount++
-
-	// Move invaders
-	s.moveTimer++
-	if s.moveTimer >= invMoveInterval {
-		s.moveTimer = 0
-		s.moveInvaders()
-	}
-
-	// Move player (auto-AI: oscillate around the invader formation center)
-	s.playerPhase += 0.04
-	centerX := s.invadersCenterX()
-	oscillation := sinApprox(s.playerPhase) * float64(s.width) * 0.12
-	targetX := centerX + oscillation
-	diff := targetX - s.playerX
-	maxPlayerSpeed := 3.0
-	s.playerX += clampPaddleMove(diff, maxPlayerSpeed)
-
-	// Player shoots occasionally
-	if s.frameCount%20 == 0 {
-		s.bullets = append(s.bullets, invBullet{
-			x: s.playerX, y: float64(s.height - s.playerHeight*2 - 2), vy: invPlayerBulletSpd,
-		})
-	}
-
-	// Invaders shoot randomly
-	for i := range s.invaders {
-		if s.invaders[i].alive && rand.Float64() < invShootChance {
-			spriteSize := float64(s.pixelSize * 8)
-			s.bullets = append(s.bullets, invBullet{
-				x: s.invaders[i].x + spriteSize/2, y: s.invaders[i].y + spriteSize, vy: invBulletSpeed,
-			})
-		}
-	}
-
-	// Update bullets
-	s.updateBullets()
-
-	// Update explosions
-	alive := s.explosions[:0]
-	for i := range s.explosions {
-		s.explosions[i].frames--
-		if s.explosions[i].frames > 0 {
-			alive = append(alive, s.explosions[i])
-		}
-	}
-	s.explosions = alive
-
-	// Check if all invaders are dead — reset
-	allDead := true
-	for i := range s.invaders {
-		if s.invaders[i].alive {
-			allDead = false
-			break
-		}
-	}
-	if allDead {
-		s.reset()
-	}
-}
-
-func (s *invadersState) moveInvaders() {
-	dropAmount := float64(s.pixelSize * 4)
-	if s.dropNext {
-		for i := range s.invaders {
-			if s.invaders[i].alive {
-				s.invaders[i].y += dropAmount
-			}
-		}
-		s.dropNext = false
-		s.moveDir = -s.moveDir
-		return
-	}
-
-	// Check if any invader hits the edge
-	spriteSize := float64(s.pixelSize * 8)
-	margin := float64(s.pixelSize * 2)
-	hitEdge := false
-	for i := range s.invaders {
-		if !s.invaders[i].alive {
-			continue
-		}
-		newX := s.invaders[i].x + s.moveAmount*s.moveDir
-		if newX < margin || newX+spriteSize > float64(s.width)-margin {
-			hitEdge = true
-			break
-		}
-	}
-
-	if hitEdge {
-		s.dropNext = true
-	} else {
-		for i := range s.invaders {
-			if s.invaders[i].alive {
-				s.invaders[i].x += s.moveAmount * s.moveDir
-			}
-		}
-	}
-
-	// If invaders reach the bottom, reset
-	bottomLimit := float64(s.height - s.playerHeight*3)
-	for i := range s.invaders {
-		if s.invaders[i].alive && s.invaders[i].y > bottomLimit {
-			s.reset()
-			return
-		}
-	}
-}
-
-func (s *invadersState) updateBullets() {
-	remaining := s.bullets[:0]
-
-	for i := range s.bullets {
-		s.bullets[i].y += s.bullets[i].vy
-
-		// Remove off-screen bullets
-		if s.bullets[i].y < -10 || s.bullets[i].y > float64(s.height)+10 {
-			continue
-		}
-
-		// Player bullets hit invaders
-		if s.bullets[i].vy < 0 {
-			hit := false
-			for j := range s.invaders {
-				if !s.invaders[j].alive {
-					continue
-				}
-				if s.bulletHitsInvader(s.bullets[i], s.invaders[j]) {
-					s.invaders[j].alive = false
-					s.explosions = append(s.explosions, invExplosion{
-						x: s.invaders[j].x, y: s.invaders[j].y, frames: invExplosionFrames,
-					})
-					hit = true
-					break
-				}
-			}
-			if hit {
-				continue
-			}
-		}
-
-		remaining = append(remaining, s.bullets[i])
-	}
-
-	s.bullets = remaining
-}
-
-func (s *invadersState) bulletHitsInvader(b invBullet, inv invaderEntity) bool {
-	spriteSize := float64(s.pixelSize * 8)
-	return b.x >= inv.x-2 && b.x <= inv.x+spriteSize && b.y >= inv.y-2 && b.y <= inv.y+spriteSize
-}
-
-// invadersCenterX returns the average X position of all living invaders.
-func (s *invadersState) invadersCenterX() float64 {
-	sumX := 0.0
-	count := 0
-	for i := range s.invaders {
-		if s.invaders[i].alive {
-			sumX += s.invaders[i].x
-			count++
-		}
-	}
-	if count == 0 {
-		return float64(s.width) / 2
-	}
-	return sumX / float64(count)
-}
-
-func (s *invadersState) render() []uint8 {
-	w, h := s.width, s.height
-	if w == 0 || h == 0 {
-		return make([]uint8, rgbaChannels)
-	}
-
-	pixels := make([]uint8, w*h*rgbaChannels)
-
-	// Draw invaders
-	for i := range s.invaders {
-		if s.invaders[i].alive {
-			s.drawInvader(pixels, s.invaders[i])
-		}
-	}
-
-	// Draw explosions
-	for i := range s.explosions {
-		s.drawExplosion(pixels, s.explosions[i])
-	}
-
-	// Draw bullets
-	for i := range s.bullets {
-		s.drawBullet(pixels, s.bullets[i])
-	}
-
-	// Draw player ship
-	s.drawPlayer(pixels)
-
-	return pixels
-}
-
-func (s *invadersState) drawInvader(pixels []uint8, inv invaderEntity) {
-	sprite := invaderSprites[inv.kind]
-	for row := range 8 {
-		for col := range 8 {
-			if sprite[row]&(1<<(7-col)) != 0 {
-				px := int(inv.x) + col*s.pixelSize
-				py := int(inv.y) + row*s.pixelSize
-				s.drawBlock(pixels, px, py, s.pixelSize, 255, 255, 255, invAlpha)
-			}
-		}
-	}
-}
-
-func (s *invadersState) drawExplosion(pixels []uint8, exp invExplosion) {
-	// Simple expanding cross pattern
-	size := (invExplosionFrames - exp.frames) * s.pixelSize
-	spriteCenter := s.pixelSize * 4
-	cx, cy := int(exp.x)+spriteCenter, int(exp.y)+spriteCenter
-	alpha := uint8(exp.frames * 15)
-
-	for d := -size; d <= size; d++ {
-		s.setInvPixel(pixels, cx+d, cy, 255, 200, 100, alpha)
-		s.setInvPixel(pixels, cx, cy+d, 255, 200, 100, alpha)
-		s.setInvPixel(pixels, cx+d, cy+d, 255, 150, 50, alpha/2)
-		s.setInvPixel(pixels, cx+d, cy-d, 255, 150, 50, alpha/2)
-	}
-}
-
-func (s *invadersState) drawBullet(pixels []uint8, b invBullet) {
-	bx, by := int(b.x), int(b.y)
-	for dy := 0; dy < invBulletHeight; dy++ {
-		for dx := 0; dx < invBulletWidth; dx++ {
-			s.setInvPixel(pixels, bx+dx, by+dy, 255, 255, 255, invBulletAlpha)
-		}
-	}
-}
-
-func (s *invadersState) drawPlayer(pixels []uint8) {
-	px := int(s.playerX) - s.playerWidth/2
-	py := s.height - s.playerHeight*2
-
-	// Simple ship shape: flat base with a turret
-	for dx := 0; dx < s.playerWidth; dx++ {
-		for dy := 0; dy < s.playerHeight/2; dy++ {
-			s.setInvPixel(pixels, px+dx, py+s.playerHeight/2+dy, 255, 255, 255, invAlpha)
-		}
-	}
-	// Turret (narrower top)
-	turretX := px + s.playerWidth/4
-	turretW := s.playerWidth / 2
-	for dx := 0; dx < turretW; dx++ {
-		for dy := 0; dy < s.playerHeight/2; dy++ {
-			s.setInvPixel(pixels, turretX+dx, py+dy, 255, 255, 255, invAlpha)
-		}
-	}
-	// Nose
-	noseX := px + s.playerWidth/2 - 1
-	s.setInvPixel(pixels, noseX, py-1, 255, 255, 255, invAlpha)
-	s.setInvPixel(pixels, noseX+1, py-1, 255, 255, 255, invAlpha)
-}
-
-func (s *invadersState) drawBlock(pixels []uint8, x, y, size int, r, g, b, a uint8) {
-	for dy := range size {
-		for dx := range size {
-			s.setInvPixel(pixels, x+dx, y+dy, r, g, b, a)
-		}
-	}
-}
-
-func (s *invadersState) setInvPixel(pixels []uint8, x, y int, r, g, b, a uint8) {
-	if x < 0 || x >= s.width || y < 0 || y >= s.height {
-		return
-	}
-	offset := (y*s.width + x) * rgbaChannels
-	if a > pixels[offset+3] {
-		pixels[offset] = r
-		pixels[offset+1] = g
-		pixels[offset+2] = b
-		pixels[offset+3] = a
-	}
-}
-
-// --- Snake Effect ---
-
-const (
-	snakeInitLength    = 5
-	snakeAlpha         = 110
-	snakeFoodAlpha     = 130
-	snakeFrameInterval = 90 * time.Millisecond
-	snakeTargetCols    = 40
-)
-
-type snakeDir int
-
-const (
-	snakeDirUp snakeDir = iota
-	snakeDirDown
-	snakeDirRight
-	snakeDirLeft
-)
-
-type snakePoint struct {
-	x, y int
-}
-
-type snakeState struct {
-	width, height int
-	cols, rows    int
-	cellSize      int
-	body          []snakePoint
-	dir           snakeDir
-	food          snakePoint
-	frameCount    int
-}
-
-func (p *shenanigans) startSnake(pc linkquisition.PickerCanvas) {
-	state := &snakeState{
-		width:  pc.Width(),
-		height: pc.Height(),
-	}
-	if state.width == 0 {
-		state.width = 600
-	}
-	if state.height == 0 {
-		state.height = 400
-	}
-	state.computeGrid()
-	state.reset()
-
-	pc.AddRasterOverlay(0.5, func(w, h int) []uint8 {
-		if w != state.width || h != state.height {
-			state.width = w
-			state.height = h
-			state.computeGrid()
-			state.reset()
-		}
-		return p.invertForLight(state.render())
-	})
-
-	go func() {
-		ticker := time.NewTicker(snakeFrameInterval)
-		defer ticker.Stop()
-
-		for range ticker.C {
-			if p.stopped.Load() {
-				return
-			}
-			state.update()
-			pc.ScheduleRefresh()
-		}
-	}()
-}
-
-func (s *snakeState) computeGrid() {
-	s.cellSize = s.width / snakeTargetCols
-	if s.cellSize < 6 {
-		s.cellSize = 6
-	}
-	if s.cellSize > 16 {
-		s.cellSize = 16
-	}
-	s.cols = s.width / s.cellSize
-	s.rows = s.height / s.cellSize
-}
-func (s *snakeState) reset() {
-	// Start in the center going right
-	cx := s.cols / 2
-	cy := s.rows / 2
-	s.body = make([]snakePoint, snakeInitLength)
-	for i := range s.body {
-		s.body[i] = snakePoint{x: cx - i, y: cy}
-	}
-	s.dir = snakeDirRight
-	s.placeFood()
-}
-
-func (s *snakeState) placeFood() {
-	// Place food at a random empty cell
-	for range 100 {
-		p := snakePoint{x: rand.IntN(s.cols), y: rand.IntN(s.rows)}
-		if !s.isBody(p) {
-			s.food = p
-			return
-		}
-	}
-	// Fallback: just place it anywhere
-	s.food = snakePoint{x: rand.IntN(s.cols), y: rand.IntN(s.rows)}
-}
-
-func (s *snakeState) isBody(p snakePoint) bool {
-	for _, seg := range s.body {
-		if seg == p {
-			return true
-		}
-	}
-	return false
-}
-
-func (s *snakeState) update() {
-	s.frameCount++
-
-	// AI: decide direction every frame
-	s.chooseDirection()
-
-	// Move head
-	head := s.body[0]
-	switch s.dir {
-	case snakeDirUp:
-		head.y--
-	case snakeDirDown:
-		head.y++
-	case snakeDirRight:
-		head.x++
-	case snakeDirLeft:
-		head.x--
-	}
-
-	// Wrap around edges
-	head.x = (head.x + s.cols) % s.cols
-	head.y = (head.y + s.rows) % s.rows
-
-	// Check self-collision — reset if hit
-	if s.isBody(head) {
-		s.reset()
-		return
-	}
-
-	// Grow or move
-	s.body = append([]snakePoint{head}, s.body...)
-	if head == s.food {
-		s.placeFood()
-	} else {
-		s.body = s.body[:len(s.body)-1]
-	}
-}
-
-func (s *snakeState) chooseDirection() {
-	// Score each valid direction by distance to food after moving
-	type option struct {
-		dir  snakeDir
-		dist int
-	}
-
-	var options []option
-	for _, d := range []snakeDir{snakeDirUp, snakeDirDown, snakeDirLeft, snakeDirRight} {
-		if d == s.oppositeDir() {
-			continue
-		}
-		next := s.nextHead(d)
-		if s.isBody(next) {
-			continue
-		}
-		dist := s.wrapDist(next, s.food)
-		options = append(options, option{d, dist})
-	}
-
-	if len(options) == 0 {
-		return // no safe move, will hit self and reset
-	}
-
-	// Pick the direction that minimizes distance to food
-	best := options[0]
-	for _, o := range options[1:] {
-		if o.dist < best.dist {
-			best = o
-		}
-	}
-
-	s.dir = best.dir
-}
-
-// wrapDist returns the Manhattan distance accounting for toroidal wrapping.
-func (s *snakeState) wrapDist(a, b snakePoint) int {
-	dx := abs(a.x - b.x)
-	dy := abs(a.y - b.y)
-	if dx > s.cols/2 {
-		dx = s.cols - dx
-	}
-	if dy > s.rows/2 {
-		dy = s.rows - dy
-	}
-	return dx + dy
-}
-
-func (s *snakeState) nextHead(d snakeDir) snakePoint {
-	head := s.body[0]
-	switch d {
-	case snakeDirUp:
-		head.y--
-	case snakeDirDown:
-		head.y++
-	case snakeDirRight:
-		head.x++
-	case snakeDirLeft:
-		head.x--
-	}
-	head.x = (head.x + s.cols) % s.cols
-	head.y = (head.y + s.rows) % s.rows
-	return head
-}
-
-func (s *snakeState) oppositeDir() snakeDir {
-	switch s.dir {
-	case snakeDirUp:
-		return snakeDirDown
-	case snakeDirDown:
-		return snakeDirUp
-	case snakeDirLeft:
-		return snakeDirRight
-	case snakeDirRight:
-		return snakeDirLeft
-	}
-	return snakeDirLeft
-}
-
-func abs(x int) int {
-	if x < 0 {
-		return -x
-	}
-	return x
-}
-
-func (s *snakeState) render() []uint8 {
-	w, h := s.width, s.height
-	if w == 0 || h == 0 {
-		return make([]uint8, rgbaChannels)
-	}
-
-	pixels := make([]uint8, w*h*rgbaChannels)
-
-	// Draw body with gradient (head brighter, tail dimmer)
-	for i, seg := range s.body {
-		fade := float64(i) / float64(len(s.body))
-		alpha := uint8(float64(snakeAlpha) * (1.0 - fade*0.6))
-		s.drawSnakeCell(pixels, seg.x, seg.y, 150, 255, 150, alpha)
-	}
-
-	// Draw food (pulsing)
-	pulse := uint8(sinNorm(float64(s.frameCount)*0.15) * 60)
-	s.drawSnakeCell(pixels, s.food.x, s.food.y, 255, 100, 100, snakeFoodAlpha+pulse)
-
-	return pixels
-}
-
-func (s *snakeState) drawSnakeCell(pixels []uint8, cx, cy int, r, g, b, a uint8) {
-	startX := cx * s.cellSize
-	startY := cy * s.cellSize
-
-	// Draw with 1px padding for grid look
-	for dy := 1; dy < s.cellSize-1; dy++ {
-		for dx := 1; dx < s.cellSize-1; dx++ {
-			px := startX + dx
-			py := startY + dy
-			if px >= s.width || py >= s.height {
-				continue
-			}
-			offset := (py*s.width + px) * rgbaChannels
-			if a > pixels[offset+3] {
-				pixels[offset] = r
-				pixels[offset+1] = g
-				pixels[offset+2] = b
-				pixels[offset+3] = a
-			}
-		}
-	}
-}
-
-// --- Rain Effect ---
-
-const (
-	rainDropCount   = 150
-	rainAlpha       = 120
-	rainSplashAlpha = 100
-)
-
-type raindrop struct {
-	x      float64
-	y      float64
-	speed  float64
-	length float64
-	width  float64
-}
-
-type rainState struct {
-	width, height int
-	drops         []raindrop
-	initialized   bool
-}
-
-func (p *shenanigans) startRain(pc linkquisition.PickerCanvas) {
-	state := &rainState{
-		width:  pc.Width(),
-		height: pc.Height(),
-	}
-	if state.width == 0 {
-		state.width = 600
-	}
-	if state.height == 0 {
-		state.height = 400
-	}
-
-	pc.AddRasterOverlay(0.5, func(w, h int) []uint8 {
-		if !state.initialized || w != state.width || h != state.height {
-			state.width = w
-			state.height = h
-			state.init()
-			state.initialized = true
-		}
-		return p.invertForLight(state.render())
-	})
-
-	go func() {
-		ticker := time.NewTicker(frameInterval)
-		defer ticker.Stop()
-
-		for range ticker.C {
-			if p.stopped.Load() {
-				return
-			}
-			state.update()
-			pc.ScheduleRefresh()
-		}
-	}()
-}
-
-func (s *rainState) init() {
-	s.drops = make([]raindrop, rainDropCount)
-	for i := range s.drops {
-		s.drops[i] = s.newDrop(true)
-	}
-}
-
-func (s *rainState) newDrop(onScreen bool) raindrop {
-	y := -(rand.Float64() * float64(s.height))
-	if onScreen {
-		y = rand.Float64() * float64(s.height)
-	}
-
-	return raindrop{
-		x:      rand.Float64() * float64(s.width),
-		y:      y,
-		speed:  8 + rand.Float64()*12,
-		length: 15 + rand.Float64()*30,
-		width:  1 + rand.Float64()*1.5,
-	}
-}
-
-func (s *rainState) update() {
-	for i := range s.drops {
-		s.drops[i].y += s.drops[i].speed
-
-		// Slight wind drift
-		s.drops[i].x += 0.5
-
-		// Respawn above when falling off bottom
-		if s.drops[i].y > float64(s.height)+s.drops[i].length {
-			s.drops[i] = s.newDrop(false)
-		}
-	}
-}
-
-func (s *rainState) render() []uint8 {
-	w, h := s.width, s.height
-	if w == 0 || h == 0 {
-		return make([]uint8, rgbaChannels)
-	}
-
-	pixels := make([]uint8, w*h*rgbaChannels)
-
-	for i := range s.drops {
-		s.drawDrop(pixels, s.drops[i])
-	}
-
-	return pixels
-}
-
-func (s *rainState) drawDrop(pixels []uint8, d raindrop) { //nolint:gocyclo
-	// Draw a vertical streak
-	startY := int(d.y - d.length)
-	endY := int(d.y)
-	x := int(d.x)
-	dropWidth := int(d.width)
-
-	for py := startY; py <= endY; py++ {
-		if py < 0 || py >= s.height {
-			continue
-		}
-
-		// Fade: brighter at the bottom (leading edge), dimmer at top
-		progress := float64(py-startY) / d.length
-		alpha := uint8(float64(rainAlpha) * progress)
-
-		for dx := 0; dx < dropWidth; dx++ {
-			px := x + dx
-			if px < 0 || px >= s.width {
-				continue
-			}
-			offset := (py*s.width + px) * rgbaChannels
-			if alpha > pixels[offset+3] {
-				pixels[offset] = 180   // R — cool blue-white
-				pixels[offset+1] = 200 // G
-				pixels[offset+2] = 255 // B
-				pixels[offset+3] = alpha
-			}
-		}
-	}
-
-	// Splash at the bottom when the drop hits — wider and a few pixels tall
-	if endY >= s.height-3 && endY < s.height {
-		splashWidth := int(d.width) + 4
-		for dy := 0; dy < 3; dy++ {
-			py := s.height - 1 - dy
-			splAlpha := rainSplashAlpha - uint8(dy*30)
-			for dx := -splashWidth; dx <= splashWidth; dx++ {
-				px := x + dx
-				if px >= 0 && px < s.width && py >= 0 {
-					offset := (py*s.width + px) * rgbaChannels
-					if splAlpha > pixels[offset+3] {
-						pixels[offset] = 200
-						pixels[offset+1] = 220
-						pixels[offset+2] = 255
-						pixels[offset+3] = splAlpha
-					}
-				}
-			}
-		}
-	}
-}
-
-// --- Breakout Effect ---
-
-const (
-	breakoutRows        = 5
-	breakoutCols        = 10
-	breakoutBallSize    = 5
-	breakoutPaddleH     = 6
-	breakoutAlpha       = 120
-	breakoutBallAlpha   = 140
-	breakoutPaddleAlpha = 110
-)
-
-var breakoutColors = [5][3]uint8{
-	{220, 50, 50},  // red
-	{220, 150, 0},  // orange
-	{220, 220, 0},  // yellow
-	{50, 200, 50},  // green
-	{50, 120, 220}, // blue
-}
-
-type breakoutBrick struct {
-	alive bool
-	color int
-}
-
-type breakoutState struct {
-	width, height int
-
-	// Grid dimensions (computed from window)
-	brickW, brickH int
-	marginTop      int
-
-	// Bricks
-	bricks []breakoutBrick
-
-	// Ball
-	ballX, ballY   float64
-	ballVX, ballVY float64
-
-	// Paddle
-	paddleX     float64
-	paddleW     int
-	paddleY     int
-	paddlePhase float64
-}
-
-func (p *shenanigans) startBreakout(pc linkquisition.PickerCanvas) {
-	state := &breakoutState{
-		width:  pc.Width(),
-		height: pc.Height(),
-	}
-	if state.width == 0 {
-		state.width = 600
-	}
-	if state.height == 0 {
-		state.height = 400
-	}
-	state.reset()
-
-	pc.AddRasterOverlay(0.4, func(w, h int) []uint8 {
-		if w != state.width || h != state.height {
-			state.width = w
-			state.height = h
-			state.reset()
-		}
-		return p.invertForLight(state.render())
-	})
-
-	go func() {
-		ticker := time.NewTicker(frameInterval)
-		defer ticker.Stop()
-
-		for range ticker.C {
-			if p.stopped.Load() {
-				return
-			}
-			state.update()
-			pc.ScheduleRefresh()
-		}
-	}()
-}
-
-func (s *breakoutState) reset() {
-	// Compute brick sizes to fill ~80% of width
-	s.brickW = (s.width * 8) / (breakoutCols * 10)
-	if s.brickW < 10 {
-		s.brickW = 10
-	}
-	s.brickH = s.brickW / 3
-	if s.brickH < 6 {
-		s.brickH = 6
-	}
-	s.marginTop = s.height / 8
-	s.paddleW = s.brickW * 2
-	s.paddleY = s.height - s.height/8
-
-	// Create bricks
-	s.bricks = make([]breakoutBrick, breakoutRows*breakoutCols)
-	for row := range breakoutRows {
-		for col := range breakoutCols {
-			s.bricks[row*breakoutCols+col] = breakoutBrick{
-				alive: true,
-				color: row % len(breakoutColors),
-			}
-		}
-	}
-
-	// Reset ball to center, random angle
-	s.ballX = float64(s.width) / 2
-	s.ballY = float64(s.height) * 0.6
-	s.ballVX = 3.0 + rand.Float64()*2.0
-	if rand.IntN(2) == 0 {
-		s.ballVX = -s.ballVX
-	}
-	s.ballVY = -(3.0 + rand.Float64()*1.5)
-
-	// Paddle starts centered
-	s.paddleX = float64(s.width) / 2
-	s.paddlePhase = rand.Float64() * 6.28
-}
-
-func (s *breakoutState) update() {
-	// Move ball
-	s.ballX += s.ballVX
-	s.ballY += s.ballVY
-
-	// Bounce off walls
-	if s.ballX <= 0 || s.ballX >= float64(s.width)-1 {
-		s.ballVX = -s.ballVX
-		s.ballX = clampFloat(s.ballX, 1, float64(s.width)-2)
-	}
-	if s.ballY <= 0 {
-		s.ballVY = -s.ballVY
-		s.ballY = 1
-	}
-
-	// Ball falls below paddle — reset
-	if s.ballY > float64(s.height) {
-		s.reset()
-		return
-	}
-
-	// Paddle AI: follow ball with oscillation
-	s.paddlePhase += 0.03
-	oscillation := sinApprox(s.paddlePhase) * float64(s.width) * 0.05
-	target := s.ballX + oscillation
-	diff := target - s.paddleX
-	s.paddleX += clampPaddleMove(diff, 4.0)
-	s.paddleX = clampFloat(s.paddleX, float64(s.paddleW/2), float64(s.width-s.paddleW/2))
-
-	// Ball-paddle collision
-	halfPaddle := float64(s.paddleW) / 2
-	if s.ballVY > 0 && s.ballY >= float64(s.paddleY)-float64(breakoutPaddleH) &&
-		s.ballY <= float64(s.paddleY) &&
-		s.ballX >= s.paddleX-halfPaddle && s.ballX <= s.paddleX+halfPaddle {
-		s.ballVY = -s.ballVY
-		// Spin based on where it hit
-		offset := (s.ballX - s.paddleX) / halfPaddle
-		s.ballVX += offset * 2.0
-		s.ballY = float64(s.paddleY) - float64(breakoutPaddleH) - 1
-	}
-
-	// Ball-brick collision
-	s.checkBrickCollisions()
-
-	// Speed cap
-	maxSpd := 7.0
-	s.ballVX = clampFloat(s.ballVX, -maxSpd, maxSpd)
-	s.ballVY = clampFloat(s.ballVY, -maxSpd, maxSpd)
-
-	// All bricks destroyed — reset
-	allDead := true
-	for i := range s.bricks {
-		if s.bricks[i].alive {
-			allDead = false
-			break
-		}
-	}
-	if allDead {
-		s.reset()
-	}
-}
-
-func (s *breakoutState) checkBrickCollisions() {
-	startX := (s.width - breakoutCols*s.brickW) / 2
-
-	for row := range breakoutRows {
-		for col := range breakoutCols {
-			idx := row*breakoutCols + col
-			if !s.bricks[idx].alive {
-				continue
-			}
-
-			bx := float64(startX + col*s.brickW)
-			by := float64(s.marginTop + row*(s.brickH+2))
-
-			// Simple AABB check
-			if s.ballX >= bx-float64(breakoutBallSize) &&
-				s.ballX <= bx+float64(s.brickW)+float64(breakoutBallSize) &&
-				s.ballY >= by-float64(breakoutBallSize) &&
-				s.ballY <= by+float64(s.brickH)+float64(breakoutBallSize) {
-				s.bricks[idx].alive = false
-				// Determine bounce direction
-				centerX := bx + float64(s.brickW)/2
-				centerY := by + float64(s.brickH)/2
-				dx := s.ballX - centerX
-				dy := s.ballY - centerY
-				if abs(int(dx))*s.brickH > abs(int(dy))*s.brickW {
-					s.ballVX = -s.ballVX
-				} else {
-					s.ballVY = -s.ballVY
-				}
-				return // only break one brick per frame
-			}
-		}
-	}
-}
-
-func (s *breakoutState) render() []uint8 {
-	w, h := s.width, s.height
-	if w == 0 || h == 0 {
-		return make([]uint8, rgbaChannels)
-	}
-
-	pixels := make([]uint8, w*h*rgbaChannels)
-	startX := (s.width - breakoutCols*s.brickW) / 2
-
-	// Draw bricks
-	for row := range breakoutRows {
-		for col := range breakoutCols {
-			idx := row*breakoutCols + col
-			if !s.bricks[idx].alive {
-				continue
-			}
-			c := breakoutColors[s.bricks[idx].color]
-			bx := startX + col*s.brickW
-			by := s.marginTop + row*(s.brickH+2)
-			s.drawRect(pixels, bx+1, by+1, s.brickW-2, s.brickH-2, c[0], c[1], c[2], breakoutAlpha)
-		}
-	}
-
-	// Draw paddle
-	px := int(s.paddleX) - s.paddleW/2
-	s.drawRect(pixels, px, s.paddleY, s.paddleW, breakoutPaddleH, 200, 200, 200, breakoutPaddleAlpha)
-
-	// Draw ball
-	bx := int(s.ballX) - breakoutBallSize/2
-	by := int(s.ballY) - breakoutBallSize/2
-	s.drawRect(pixels, bx, by, breakoutBallSize, breakoutBallSize, 255, 255, 255, breakoutBallAlpha)
-
-	return pixels
-}
-
-func (s *breakoutState) drawRect(pixels []uint8, x, y, rw, rh int, r, g, b, a uint8) {
-	for dy := 0; dy < rh; dy++ {
-		for dx := 0; dx < rw; dx++ {
-			px := x + dx
-			py := y + dy
-			if px < 0 || px >= s.width || py < 0 || py >= s.height {
-				continue
-			}
-			offset := (py*s.width + px) * rgbaChannels
-			if a > pixels[offset+3] {
-				pixels[offset] = r
-				pixels[offset+1] = g
-				pixels[offset+2] = b
-				pixels[offset+3] = a
-			}
-		}
-	}
-}
-
-// --- Dino Runner Effect ---
-
-const (
-	dinoGroundY     = 0.82 // fraction of height for ground line
-	dinoGravity     = 0.6
-	dinoJumpForce   = -10.0
-	dinoRunSpeed    = 4.0
-	dinoAlpha       = 130
-	dinoCactusAlpha = 110
-	dinoGroundAlpha = 60
-)
-
-// Dino sprites — two leg frames for running animation (16 wide x 16 tall)
-// Designed to look like the Chrome offline T-rex
-var dinoSpriteRun1 = [16]uint16{
-	0x0000, // ................
-	0x03FC, // ......XXXXXXXX..
-	0x07FE, // .....XXXXXXXXX..
-	0x07DE, // .....XXXXX.XXXX.
-	0x07FE, // .....XXXXXXXXX..
-	0x07F0, // .....XXXXXXX....
-	0x7FE0, // .XXXXXXXXXX.....
-	0xFFF0, // XXXXXXXXXXXX....
-	0xFFF0, // XXXXXXXXXXXX....
-	0x7FE0, // .XXXXXXXXXX.....
-	0x3FC0, // ..XXXXXXXX......
-	0x1F80, // ...XXXXXX.......
-	0x0F00, // ....XXXX........
-	0x0600, // .....XX.........
-	0x0300, // ......XX........
-	0x0000, // ................
-}
-
-var dinoSpriteRun2 = [16]uint16{
-	0x0000, // ................
-	0x03FC, // ......XXXXXXXX..
-	0x07FE, // .....XXXXXXXXX..
-	0x07DE, // .....XXXXX.XXXX.
-	0x07FE, // .....XXXXXXXXX..
-	0x07F0, // .....XXXXXXX....
-	0x7FE0, // .XXXXXXXXXX.....
-	0xFFF0, // XXXXXXXXXXXX....
-	0xFFF0, // XXXXXXXXXXXX....
-	0x7FE0, // .XXXXXXXXXX.....
-	0x3FC0, // ..XXXXXXXX......
-	0x1F80, // ...XXXXXX.......
-	0x0F00, // ....XXXX........
-	0x0300, // ......XX........
-	0x0600, // .....XX.........
-	0x0000, // ................
-}
-
-// Dead dino (eyes become X)
-var dinoSpriteDead = [16]uint16{
-	0x0000, // ................
-	0x03FC, // ......XXXXXXXX..
-	0x05FA, // .....X.XXXXX.X..
-	0x07FE, // .....XXXXXXXXX..
-	0x07FE, // .....XXXXXXXXX..
-	0x07F0, // .....XXXXXXX....
-	0x7FE0, // .XXXXXXXXXX.....
-	0xFFF0, // XXXXXXXXXXXX....
-	0xFFF0, // XXXXXXXXXXXX....
-	0x7FE0, // .XXXXXXXXXX.....
-	0x3FC0, // ..XXXXXXXX......
-	0x1F80, // ...XXXXXX.......
-	0x0F00, // ....XXXX........
-	0x0900, // ....X..X........
-	0x0900, // ....X..X........
-	0x0000, // ................
-}
-
-type dinoCactus struct {
-	x      float64
-	height int // 1=small, 2=medium, 3=tall
-}
-
-type dinoState struct {
-	width, height int
-	scale         int
-
-	// Dino
-	dinoY      float64
-	dinoVY     float64
-	groundY    int
-	dead       bool
-	deathTimer int
-
-	// Obstacles
-	cacti      []dinoCactus
-	spawnTimer int
-	spawnDelay int
-
-	// Game
-	speed      float64
-	frameCount int
-}
-
-func (p *shenanigans) startDino(pc linkquisition.PickerCanvas) {
-	state := &dinoState{
-		width:  pc.Width(),
-		height: pc.Height(),
-	}
-	if state.width == 0 {
-		state.width = 600
-	}
-	if state.height == 0 {
-		state.height = 400
-	}
-	state.reset()
-
-	pc.AddRasterOverlay(0.45, func(w, h int) []uint8 {
-		if w != state.width || h != state.height {
-			state.width = w
-			state.height = h
-			state.reset()
-		}
-		return p.invertForLight(state.render())
-	})
-
-	go func() {
-		ticker := time.NewTicker(frameInterval)
-		defer ticker.Stop()
-
-		for range ticker.C {
-			if p.stopped.Load() {
-				return
-			}
-			state.update()
-			pc.ScheduleRefresh()
-		}
-	}()
-}
-
-func (s *dinoState) reset() {
-	s.scale = s.width / 200
-	if s.scale < 2 {
-		s.scale = 2
-	}
-	if s.scale > 5 {
-		s.scale = 5
-	}
-
-	s.groundY = int(float64(s.height) * dinoGroundY)
-	s.dinoY = float64(s.groundY)
-	s.dinoVY = 0
-	s.dead = false
-	s.deathTimer = 0
-	s.cacti = nil
-	s.spawnTimer = 0
-	s.spawnDelay = 40 + rand.IntN(30)
-	s.speed = dinoRunSpeed
-	s.frameCount = 0
-}
-
-func (s *dinoState) update() {
-	s.frameCount++
-
-	// If dead, show death sprite for a bit then reset
-	if s.dead {
-		s.deathTimer++
-		if s.deathTimer > 40 {
-			s.reset()
-		}
-		return
-	}
-
-	// Gradually increase speed
-	if s.frameCount%100 == 0 && s.speed < 9.0 {
-		s.speed += 0.2
-	}
-
-	// Jump AI: jump when a cactus is close
-	onGround := s.dinoY >= float64(s.groundY)
-	if onGround {
-		for _, c := range s.cacti {
-			dinoRight := float64(s.scale*8 + s.scale*16) // dino x + sprite width
-			dist := c.x - dinoRight
-			if dist > 0 && dist < s.speed*12 {
-				s.dinoVY = dinoJumpForce * float64(s.scale) / 3.0
-				break
-			}
-		}
-	}
-
-	// Apply gravity
-	s.dinoVY += dinoGravity
-	s.dinoY += s.dinoVY
-	if s.dinoY > float64(s.groundY) {
-		s.dinoY = float64(s.groundY)
-		s.dinoVY = 0
-	}
-
-	// Move cacti
-	for i := range s.cacti {
-		s.cacti[i].x -= s.speed
-	}
-
-	// Check collision
-	s.checkCollision()
-
-	// Remove off-screen cacti
-	alive := s.cacti[:0]
-	for _, c := range s.cacti {
-		if c.x > -float64(s.scale*20) {
-			alive = append(alive, c)
-		}
-	}
-	s.cacti = alive
-
-	// Spawn new cacti
-	s.spawnTimer++
-	if s.spawnTimer >= s.spawnDelay {
-		s.spawnTimer = 0
-		s.spawnDelay = 30 + rand.IntN(40)
-		s.cacti = append(s.cacti, dinoCactus{
-			x:      float64(s.width + s.scale*5),
-			height: 1 + rand.IntN(3),
-		})
-	}
-}
-
-func (s *dinoState) checkCollision() {
-	dinoX := s.scale * 8
-	dinoW := s.scale * 12 // effective body width (not full 16, trimmed)
-	dinoH := s.scale * 14
-	dinoTop := int(s.dinoY) + s.scale*2 // skip top empty row
-
-	for _, c := range s.cacti {
-		cx := int(c.x)
-		cactusW := s.scale * 3
-		cactusH := s.scale * (5 + c.height*3)
-		cactusTop := s.groundY + s.scale*16 - cactusH
-
-		// AABB overlap check
-		if dinoX+dinoW > cx && dinoX < cx+cactusW &&
-			dinoTop+dinoH > cactusTop && dinoTop < cactusTop+cactusH {
-			s.dead = true
-			s.deathTimer = 0
-			return
-		}
-	}
-}
-
-func (s *dinoState) render() []uint8 {
-	w, h := s.width, s.height
-	if w == 0 || h == 0 {
-		return make([]uint8, rgbaChannels)
-	}
-
-	pixels := make([]uint8, w*h*rgbaChannels)
-
-	// Draw ground line
-	for x := 0; x < w; x++ {
-		for dy := 0; dy < s.scale; dy++ {
-			py := s.groundY + s.scale*16 + dy
-			if py < h {
-				offset := (py*w + x) * rgbaChannels
-				pixels[offset] = 180
-				pixels[offset+1] = 180
-				pixels[offset+2] = 180
-				pixels[offset+3] = dinoGroundAlpha
-			}
-		}
-	}
-
-	// Draw dino
-	s.drawDino(pixels)
-
-	// Draw cacti
-	for _, c := range s.cacti {
-		s.drawCactus(pixels, c)
-	}
-
-	return pixels
-}
-
-func (s *dinoState) drawDino(pixels []uint8) {
-	dinoX := s.scale * 8 // fixed X position from left edge
-	baseY := int(s.dinoY)
-
-	// Choose sprite based on state
-	var sprite *[16]uint16
-	if s.dead {
-		sprite = &dinoSpriteDead
-	} else if s.dinoY < float64(s.groundY) {
-		// In the air — use run1 (legs together)
-		sprite = &dinoSpriteRun1
-	} else {
-		// On ground — alternate legs every 4 frames
-		if (s.frameCount/4)%2 == 0 {
-			sprite = &dinoSpriteRun1
-		} else {
-			sprite = &dinoSpriteRun2
-		}
-	}
-
-	for row := range 16 {
-		bits := sprite[row]
-		for col := range 16 {
-			if bits&(1<<(15-col)) != 0 {
-				px := dinoX + col*s.scale
-				py := baseY + row*s.scale
-				s.drawBlock(pixels, px, py, s.scale, 255, 255, 255, dinoAlpha)
-			}
-		}
-	}
-}
-
-func (s *dinoState) drawCactus(pixels []uint8, c dinoCactus) {
-	cx := int(c.x)
-	cactusW := s.scale * 3
-	cactusH := s.scale * (5 + c.height*3)
-	cy := s.groundY + s.scale*16 - cactusH
-
-	// Main stem
-	s.drawRect(pixels, cx, cy, cactusW, cactusH, 100, 200, 100, dinoCactusAlpha)
-
-	// Arms for taller cacti
-	if c.height >= 2 {
-		armY := cy + cactusH/3
-		// Left arm
-		s.drawRect(pixels, cx-s.scale*2, armY, s.scale*2, s.scale*2, 100, 200, 100, dinoCactusAlpha)
-		s.drawRect(pixels, cx-s.scale*2, armY-s.scale*2, s.scale, s.scale*2, 100, 200, 100, dinoCactusAlpha)
-	}
-	if c.height >= 3 {
-		armY := cy + cactusH*2/3
-		// Right arm
-		s.drawRect(pixels, cx+cactusW, armY, s.scale*2, s.scale*2, 100, 200, 100, dinoCactusAlpha)
-		s.drawRect(pixels, cx+cactusW+s.scale, armY-s.scale*2, s.scale, s.scale*2, 100, 200, 100, dinoCactusAlpha)
-	}
-}
-
-func (s *dinoState) drawRect(pixels []uint8, x, y, rw, rh int, r, g, b, a uint8) { //nolint:unparam
-	for dy := range rh {
-		for dx := range rw {
-			px := x + dx
-			py := y + dy
-			if px < 0 || px >= s.width || py < 0 || py >= s.height {
-				continue
-			}
-			offset := (py*s.width + px) * rgbaChannels
-			if a > pixels[offset+3] {
-				pixels[offset] = r
-				pixels[offset+1] = g
-				pixels[offset+2] = b
-				pixels[offset+3] = a
-			}
-		}
-	}
-}
-
-func (s *dinoState) drawBlock(pixels []uint8, x, y, size int, r, g, b, a uint8) {
-	for dy := range size {
-		for dx := range size {
-			px := x + dx
-			py := y + dy
-			if px < 0 || px >= s.width || py < 0 || py >= s.height {
-				continue
-			}
-			offset := (py*s.width + px) * rgbaChannels
-			if a > pixels[offset+3] {
-				pixels[offset] = r
-				pixels[offset+1] = g
-				pixels[offset+2] = b
-				pixels[offset+3] = a
-			}
-		}
-	}
-}
-
-// --- Asteroids Effect ---
-
-const (
-	astShipSize      = 10
-	astAlpha         = 120
-	astBulletAlpha   = 140
-	astBulletSpeed   = 6.0
-	astBulletLife    = 40
-	astMaxAsteroids  = 8
-	astShootInterval = 15
-	astTurnSpeed     = 0.05
-	astThrust        = 0.12
-	astFriction      = 0.98
-)
-
-type astVec struct{ x, y float64 }
-
-type asteroid struct {
-	pos    astVec
-	vel    astVec
-	radius float64
-	edges  int // number of vertices (6-10)
-}
-
-type astBullet struct {
-	pos  astVec
-	vel  astVec
-	life int
-}
-
-type asteroidsState struct {
-	width, height int
-	scale         float64
-
-	// Ship
-	shipPos   astVec
-	shipVel   astVec
-	shipAngle float64
-
-	// Objects
-	asteroids  []asteroid
-	bullets    []astBullet
-	shootTimer int
-	frameCount int
-}
-
-func (p *shenanigans) startAsteroids(pc linkquisition.PickerCanvas) {
-	state := &asteroidsState{
-		width:  pc.Width(),
-		height: pc.Height(),
-	}
-	if state.width == 0 {
-		state.width = 600
-	}
-	if state.height == 0 {
-		state.height = 400
-	}
-	state.reset()
-
-	pc.AddRasterOverlay(0.45, func(w, h int) []uint8 {
-		if w != state.width || h != state.height {
-			state.width = w
-			state.height = h
-			state.reset()
-		}
-		return p.invertForLight(state.render())
-	})
-
-	go func() {
-		ticker := time.NewTicker(frameInterval)
-		defer ticker.Stop()
-
-		for range ticker.C {
-			if p.stopped.Load() {
-				return
-			}
-			state.update()
-			pc.ScheduleRefresh()
-		}
-	}()
-}
-
-func (s *asteroidsState) reset() {
-	s.scale = float64(s.width) / 600.0
-	if s.scale < 0.5 {
-		s.scale = 0.5
-	}
-
-	s.shipPos = astVec{float64(s.width) / 2, float64(s.height) / 2}
-	s.shipVel = astVec{0, 0}
-	s.shipAngle = 0
-	s.bullets = nil
-	s.shootTimer = 0
-	s.frameCount = 0
-	s.spawnAsteroids()
-}
-
-func (s *asteroidsState) spawnAsteroids() {
-	s.asteroids = make([]asteroid, astMaxAsteroids)
-	for i := range s.asteroids {
-		// Spawn at edges
-		var pos astVec
-		if rand.IntN(2) == 0 {
-			pos.x = rand.Float64() * float64(s.width)
-			if rand.IntN(2) == 0 {
-				pos.y = 0
-			} else {
-				pos.y = float64(s.height)
-			}
-		} else {
-			pos.y = rand.Float64() * float64(s.height)
-			if rand.IntN(2) == 0 {
-				pos.x = 0
-			} else {
-				pos.x = float64(s.width)
-			}
-		}
-		s.asteroids[i] = asteroid{
-			pos:    pos,
-			vel:    astVec{(rand.Float64() - 0.5) * 2, (rand.Float64() - 0.5) * 2},
-			radius: (20 + rand.Float64()*15) * s.scale,
-			edges:  6 + rand.IntN(4),
-		}
-	}
-}
-
-func (s *asteroidsState) update() {
-	s.frameCount++
-
-	// Ship AI: rotate toward nearest asteroid, thrust toward it, shoot
-	nearest := s.nearestAsteroid()
-	if nearest != nil {
-		// Angle to nearest asteroid
-		dx := nearest.pos.x - s.shipPos.x
-		dy := nearest.pos.y - s.shipPos.y
-		targetAngle := atan2Approx(dy, dx)
-
-		// Rotate toward target
-		diff := targetAngle - s.shipAngle
-		for diff > 3.14159 {
-			diff -= 6.28318
-		}
-		for diff < -3.14159 {
-			diff += 6.28318
-		}
-		if diff > astTurnSpeed {
-			s.shipAngle += astTurnSpeed
-		} else if diff < -astTurnSpeed {
-			s.shipAngle -= astTurnSpeed
-		}
-
-		// Thrust
-		s.shipVel.x += cosApprox(s.shipAngle) * astThrust
-		s.shipVel.y += sinApprox(s.shipAngle) * astThrust
-	}
-
-	// Friction
-	s.shipVel.x *= astFriction
-	s.shipVel.y *= astFriction
-
-	// Move ship
-	s.shipPos.x += s.shipVel.x
-	s.shipPos.y += s.shipVel.y
-
-	// Wrap ship
-	s.shipPos.x = s.wrapX(s.shipPos.x)
-	s.shipPos.y = s.wrapY(s.shipPos.y)
-
-	// Shoot
-	s.shootTimer++
-	if s.shootTimer >= astShootInterval && nearest != nil {
-		s.shootTimer = 0
-		s.bullets = append(s.bullets, astBullet{
-			pos:  s.shipPos,
-			vel:  astVec{cosApprox(s.shipAngle) * astBulletSpeed, sinApprox(s.shipAngle) * astBulletSpeed},
-			life: astBulletLife,
-		})
-	}
-
-	// Move bullets
-	alive := s.bullets[:0]
-	for i := range s.bullets {
-		s.bullets[i].pos.x += s.bullets[i].vel.x
-		s.bullets[i].pos.y += s.bullets[i].vel.y
-		s.bullets[i].life--
-		if s.bullets[i].life > 0 {
-			alive = append(alive, s.bullets[i])
-		}
-	}
-	s.bullets = alive
-
-	// Move asteroids
-	for i := range s.asteroids {
-		s.asteroids[i].pos.x = s.wrapX(s.asteroids[i].pos.x + s.asteroids[i].vel.x)
-		s.asteroids[i].pos.y = s.wrapY(s.asteroids[i].pos.y + s.asteroids[i].vel.y)
-	}
-
-	// Bullet-asteroid collision
-	s.checkBulletHits()
-
-	// Respawn if all destroyed
-	if len(s.asteroids) == 0 {
-		s.spawnAsteroids()
-	}
-}
-
-func (s *asteroidsState) checkBulletHits() {
-	var newBullets []astBullet
-	for _, b := range s.bullets {
-		hit := false
-		for i := range s.asteroids {
-			dx := b.pos.x - s.asteroids[i].pos.x
-			dy := b.pos.y - s.asteroids[i].pos.y
-			dist := dx*dx + dy*dy
-			r := s.asteroids[i].radius
-			if dist < r*r {
-				// Split asteroid
-				if s.asteroids[i].radius > 12*s.scale {
-					newR := s.asteroids[i].radius * 0.55
-					for range 2 {
-						s.asteroids = append(s.asteroids, asteroid{
-							pos:    s.asteroids[i].pos,
-							vel:    astVec{(rand.Float64() - 0.5) * 3, (rand.Float64() - 0.5) * 3},
-							radius: newR,
-							edges:  5 + rand.IntN(4),
-						})
-					}
-				}
-				// Remove hit asteroid
-				s.asteroids = append(s.asteroids[:i], s.asteroids[i+1:]...)
-				hit = true
-				break
-			}
-		}
-		if !hit {
-			newBullets = append(newBullets, b)
-		}
-	}
-	s.bullets = newBullets
-}
-
-func (s *asteroidsState) nearestAsteroid() *asteroid {
-	if len(s.asteroids) == 0 {
-		return nil
-	}
-	best := 0
-	bestDist := 1e18
-	for i := range s.asteroids {
-		dx := s.asteroids[i].pos.x - s.shipPos.x
-		dy := s.asteroids[i].pos.y - s.shipPos.y
-		d := dx*dx + dy*dy
-		if d < bestDist {
-			bestDist = d
-			best = i
-		}
-	}
-	return &s.asteroids[best]
-}
-
-func (s *asteroidsState) wrapX(x float64) float64 {
-	w := float64(s.width)
-	for x < 0 {
-		x += w
-	}
-	for x >= w {
-		x -= w
-	}
-	return x
-}
-
-func (s *asteroidsState) wrapY(y float64) float64 {
-	h := float64(s.height)
-	for y < 0 {
-		y += h
-	}
-	for y >= h {
-		y -= h
-	}
-	return y
-}
-
-func (s *asteroidsState) render() []uint8 {
-	w, h := s.width, s.height
-	if w == 0 || h == 0 {
-		return make([]uint8, rgbaChannels)
-	}
-
-	pixels := make([]uint8, w*h*rgbaChannels)
-
-	// Draw asteroids as wireframe polygons
-	for i := range s.asteroids {
-		s.drawAsteroid(pixels, s.asteroids[i])
-	}
-
-	// Draw bullets
-	for _, b := range s.bullets {
-		bx, by := int(b.pos.x), int(b.pos.y)
-		for dy := -1; dy <= 1; dy++ {
-			for dx := -1; dx <= 1; dx++ {
-				s.setAstPixel(pixels, bx+dx, by+dy, 255, 255, 255, astBulletAlpha)
-			}
-		}
-	}
-
-	// Draw ship (triangle)
-	s.drawShip(pixels)
-
-	return pixels
-}
-
-func (s *asteroidsState) drawAsteroid(pixels []uint8, a asteroid) {
-	cx, cy := a.pos.x, a.pos.y
-	n := a.edges
-
-	for i := range n {
-		angle1 := float64(i) * 6.28318 / float64(n)
-		angle2 := float64(i+1) * 6.28318 / float64(n)
-		x1 := cx + cosApprox(angle1)*a.radius
-		y1 := cy + sinApprox(angle1)*a.radius
-		x2 := cx + cosApprox(angle2)*a.radius
-		y2 := cy + sinApprox(angle2)*a.radius
-		s.drawLine(pixels, int(x1), int(y1), int(x2), int(y2), 200, 200, 200, astAlpha)
-	}
-}
-
-func (s *asteroidsState) drawShip(pixels []uint8) {
-	// Triangle pointing in shipAngle direction
-	size := float64(astShipSize) * s.scale
-	cx, cy := s.shipPos.x, s.shipPos.y
-
-	// Nose
-	nx := cx + cosApprox(s.shipAngle)*size
-	ny := cy + sinApprox(s.shipAngle)*size
-	// Left wing
-	lx := cx + cosApprox(s.shipAngle+2.5)*size*0.6
-	ly := cy + sinApprox(s.shipAngle+2.5)*size*0.6
-	// Right wing
-	rx := cx + cosApprox(s.shipAngle-2.5)*size*0.6
-	ry := cy + sinApprox(s.shipAngle-2.5)*size*0.6
-
-	s.drawLine(pixels, int(nx), int(ny), int(lx), int(ly), 255, 255, 255, astAlpha)
-	s.drawLine(pixels, int(nx), int(ny), int(rx), int(ry), 255, 255, 255, astAlpha)
-	s.drawLine(pixels, int(lx), int(ly), int(rx), int(ry), 255, 255, 255, astAlpha)
-}
-
-// drawLine draws a line using Bresenham's algorithm.
-func (s *asteroidsState) drawLine(pixels []uint8, x0, y0, x1, y1 int, r, g, b, a uint8) { //nolint:unparam
-	dx := abs(x1 - x0)
-	dy := -abs(y1 - y0)
-	sx := 1
-	if x0 > x1 {
-		sx = -1
-	}
-	sy := 1
-	if y0 > y1 {
-		sy = -1
-	}
-	err := dx + dy
-
-	for {
-		s.setAstPixel(pixels, x0, y0, r, g, b, a)
-		if x0 == x1 && y0 == y1 {
-			break
-		}
-		e2 := 2 * err
-		if e2 >= dy {
-			err += dy
-			x0 += sx
-		}
-		if e2 <= dx {
-			err += dx
-			y0 += sy
-		}
-	}
-}
-
-func (s *asteroidsState) setAstPixel(pixels []uint8, x, y int, r, g, b, a uint8) {
-	if x < 0 || x >= s.width || y < 0 || y >= s.height {
-		return
-	}
-	offset := (y*s.width + x) * rgbaChannels
-	if a > pixels[offset+3] {
-		pixels[offset] = r
-		pixels[offset+1] = g
-		pixels[offset+2] = b
-		pixels[offset+3] = a
-	}
-}
-
-// cosApprox uses sinApprox shifted by pi/2.
-func cosApprox(x float64) float64 {
-	return sinApprox(x + 1.5707963)
-}
-
-// atan2Approx is a rough atan2 approximation.
-func atan2Approx(y, x float64) float64 {
-	// Simple approximation using the identity and sinApprox
-	const pi = 3.14159265
-	if x == 0 {
-		if y > 0 {
-			return pi / 2
-		}
-		return -pi / 2
-	}
-	a := y / x
-	// Clamp for stability
-	if a > 10 {
-		a = 10
-	} else if a < -10 {
-		a = -10
-	}
-	// Polynomial approximation of atan for small values
-	result := a / (1 + 0.28*a*a)
-	if x < 0 {
-		if y >= 0 {
-			result += pi
-		} else {
-			result -= pi
-		}
-	}
-	return result
-}
-
-// --- Pac-Man Effect ---
-
-const (
-	pacCellSize      = 10
-	pacAlpha         = 130
-	pacGhostAlpha    = 110
-	pacDotAlpha      = 80
-	pacWallAlpha     = 60
-	pacFrameInterval = 120 * time.Millisecond
-	pacMazeW         = 21
-	pacMazeH         = 11
-)
-
-// Simple maze layout: 1=wall, 0=path
-var pacMaze = [pacMazeH][pacMazeW]uint8{
-	{1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1},
-	{1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1},
-	{1, 0, 1, 1, 0, 1, 1, 1, 1, 0, 1, 0, 1, 1, 1, 1, 0, 1, 1, 0, 1},
-	{1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1},
-	{1, 0, 1, 1, 0, 1, 0, 1, 1, 1, 1, 1, 1, 1, 0, 1, 0, 1, 1, 0, 1},
-	{1, 0, 0, 0, 0, 1, 0, 0, 0, 0, 1, 0, 0, 0, 0, 1, 0, 0, 0, 0, 1},
-	{1, 0, 1, 1, 0, 1, 0, 1, 1, 1, 1, 1, 1, 1, 0, 1, 0, 1, 1, 0, 1},
-	{1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1},
-	{1, 0, 1, 1, 0, 1, 1, 1, 1, 0, 1, 0, 1, 1, 1, 1, 0, 1, 1, 0, 1},
-	{1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1},
-	{1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1},
-}
-
-type pacEntity struct {
-	x, y int
-	dir  int // 0=right, 1=down, 2=left, 3=up
-}
-
-type pacmanState struct {
-	width, height int
-	cellSize      int
-	offsetX       int
-	offsetY       int
-
-	pacman     pacEntity
-	ghosts     [4]pacEntity
-	dots       [pacMazeH][pacMazeW]bool
-	mouthOpen  bool
-	frameCount int
-}
-
-func (p *shenanigans) startPacman(pc linkquisition.PickerCanvas) {
-	state := &pacmanState{
-		width:  pc.Width(),
-		height: pc.Height(),
-	}
-	if state.width == 0 {
-		state.width = 600
-	}
-	if state.height == 0 {
-		state.height = 400
-	}
-	state.reset()
-
-	pc.AddRasterOverlay(0.45, func(w, h int) []uint8 {
-		if w != state.width || h != state.height {
-			state.width = w
-			state.height = h
-			state.reset()
-		}
-		return p.invertForLight(state.render())
-	})
-
-	go func() {
-		ticker := time.NewTicker(pacFrameInterval)
-		defer ticker.Stop()
-
-		for range ticker.C {
-			if p.stopped.Load() {
-				return
-			}
-			state.update()
-			pc.ScheduleRefresh()
-		}
-	}()
-}
-
-func (s *pacmanState) reset() {
-	// Scale maze to fit window
-	cellW := s.width / pacMazeW
-	cellH := s.height / pacMazeH
-	s.cellSize = cellW
-	if cellH < cellW {
-		s.cellSize = cellH
-	}
-	if s.cellSize < 6 {
-		s.cellSize = 6
-	}
-
-	// Center the maze
-	s.offsetX = (s.width - pacMazeW*s.cellSize) / 2
-	s.offsetY = (s.height - pacMazeH*s.cellSize) / 2
-
-	// Place pac-man
-	s.pacman = pacEntity{x: 1, y: 1, dir: 0}
-
-	// Place ghosts
-	s.ghosts = [4]pacEntity{
-		{x: 9, y: 5, dir: 0},
-		{x: 10, y: 5, dir: 2},
-		{x: 11, y: 5, dir: 0},
-		{x: 10, y: 3, dir: 1},
-	}
-
-	// Fill dots on all path cells
-	for row := range pacMazeH {
-		for col := range pacMazeW {
-			s.dots[row][col] = pacMaze[row][col] == 0
-		}
-	}
-	s.frameCount = 0
-}
-
-func (s *pacmanState) update() {
-	s.frameCount++
-	s.mouthOpen = (s.frameCount/3)%2 == 0
-
-	// Move pac-man (AI: prefer direction toward nearest dot)
-	s.movePacman()
-
-	// Eat dot
-	s.dots[s.pacman.y][s.pacman.x] = false
-
-	// Move ghosts
-	for i := range s.ghosts {
-		s.moveGhost(i)
-	}
-
-	// Check if all dots eaten — reset
-	hasDots := false
-	for row := range pacMazeH {
-		for col := range pacMazeW {
-			if s.dots[row][col] {
-				hasDots = true
-				break
-			}
-		}
-		if hasDots {
-			break
-		}
-	}
-	if !hasDots {
-		s.reset()
-	}
-}
-
-func (s *pacmanState) movePacman() {
-	// Find nearest dot
-	bestDist := 9999
-	bestDir := s.pacman.dir
-
-	for _, d := range []int{0, 1, 2, 3} {
-		nx, ny := s.nextCell(s.pacman.x, s.pacman.y, d)
-		if pacMaze[ny][nx] == 1 {
-			continue
-		}
-		// Don't reverse unless no other option
-		if d == (s.pacman.dir+2)%4 {
-			continue
-		}
-		dist := s.distToNearestDot(nx, ny)
-		if dist < bestDist {
-			bestDist = dist
-			bestDir = d
-		}
-	}
-
-	// If no forward option, allow reverse
-	nx, ny := s.nextCell(s.pacman.x, s.pacman.y, bestDir)
-	if pacMaze[ny][nx] == 1 {
-		bestDir = (s.pacman.dir + 2) % 4
-	}
-
-	s.pacman.dir = bestDir
-	nx, ny = s.nextCell(s.pacman.x, s.pacman.y, bestDir)
-	if pacMaze[ny][nx] == 0 {
-		s.pacman.x = nx
-		s.pacman.y = ny
-	}
-}
-
-func (s *pacmanState) moveGhost(idx int) {
-	g := &s.ghosts[idx]
-
-	// Simple AI: pick a random valid direction (not reverse) at intersections
-	var options []int
-	for _, d := range []int{0, 1, 2, 3} {
-		if d == (g.dir+2)%4 {
-			continue
-		}
-		nx, ny := s.nextCell(g.x, g.y, d)
-		if pacMaze[ny][nx] == 0 {
-			options = append(options, d)
-		}
-	}
-
-	if len(options) == 0 {
-		// Dead end, reverse
-		g.dir = (g.dir + 2) % 4
-	} else {
-		g.dir = options[rand.IntN(len(options))]
-	}
-
-	nx, ny := s.nextCell(g.x, g.y, g.dir)
-	if pacMaze[ny][nx] == 0 {
-		g.x = nx
-		g.y = ny
-	}
-}
-
-func (s *pacmanState) nextCell(x, y, dir int) (nx, ny int) {
-	nx, ny = x, y
-	switch dir {
-	case 0:
-		nx++
-	case 1:
-		ny++
-	case 2:
-		nx--
-	case 3:
-		ny--
-	}
-	// Clamp to maze bounds
-	if nx < 0 {
-		nx = 0
-	} else if nx >= pacMazeW {
-		nx = pacMazeW - 1
-	}
-	if ny < 0 {
-		ny = 0
-	} else if ny >= pacMazeH {
-		ny = pacMazeH - 1
-	}
-	return nx, ny
-}
-
-func (s *pacmanState) distToNearestDot(fromX, fromY int) int {
-	best := 9999
-	for row := range pacMazeH {
-		for col := range pacMazeW {
-			if s.dots[row][col] {
-				d := abs(col-fromX) + abs(row-fromY)
-				if d < best {
-					best = d
-				}
-			}
-		}
-	}
-	return best
-}
-
-func (s *pacmanState) render() []uint8 {
-	w, h := s.width, s.height
-	if w == 0 || h == 0 {
-		return make([]uint8, rgbaChannels)
-	}
-
-	pixels := make([]uint8, w*h*rgbaChannels)
-
-	// Draw maze walls
-	for row := range pacMazeH {
-		for col := range pacMazeW {
-			if pacMaze[row][col] == 1 {
-				px := s.offsetX + col*s.cellSize
-				py := s.offsetY + row*s.cellSize
-				s.drawPacRect(pixels, px+1, py+1, s.cellSize-2, s.cellSize-2, 30, 30, 180, pacWallAlpha)
-			}
-		}
-	}
-
-	// Draw dots
-	dotR := s.cellSize / 6
-	if dotR < 1 {
-		dotR = 1
-	}
-	for row := range pacMazeH {
-		for col := range pacMazeW {
-			if s.dots[row][col] {
-				cx := s.offsetX + col*s.cellSize + s.cellSize/2
-				cy := s.offsetY + row*s.cellSize + s.cellSize/2
-				s.drawCircle(pixels, cx, cy, dotR, 255, 200, 150, pacDotAlpha)
-			}
-		}
-	}
-
-	// Draw pac-man
-	s.drawPacman(pixels)
-
-	// Draw ghosts
-	ghostColors := [4][3]uint8{{255, 0, 0}, {255, 150, 200}, {0, 220, 220}, {255, 150, 0}}
-	for i, g := range s.ghosts {
-		cx := s.offsetX + g.x*s.cellSize + s.cellSize/2
-		cy := s.offsetY + g.y*s.cellSize + s.cellSize/2
-		r := s.cellSize * 2 / 5
-		c := ghostColors[i]
-		s.drawCircle(pixels, cx, cy, r, c[0], c[1], c[2], pacGhostAlpha)
-		// Flat bottom for ghost shape
-		s.drawPacRect(pixels, cx-r, cy, r*2, r, c[0], c[1], c[2], pacGhostAlpha)
-	}
-
-	return pixels
-}
-
-func (s *pacmanState) drawPacman(pixels []uint8) { //nolint:gocyclo
-	cx := s.offsetX + s.pacman.x*s.cellSize + s.cellSize/2
-	cy := s.offsetY + s.pacman.y*s.cellSize + s.cellSize/2
-	r := s.cellSize * 2 / 5
-
-	// Draw as circle with a mouth gap
-	for dy := -r; dy <= r; dy++ {
-		for dx := -r; dx <= r; dx++ {
-			if dx*dx+dy*dy > r*r {
-				continue
-			}
-			// Mouth: skip pixels in a wedge in the facing direction
-			if s.mouthOpen {
-				skip := false
-				switch s.pacman.dir {
-				case 0: // right
-					skip = dx > 0 && abs(dy) < dx
-				case 2: // left
-					skip = dx < 0 && abs(dy) < -dx
-				case 1: // down
-					skip = dy > 0 && abs(dx) < dy
-				case 3: // up
-					skip = dy < 0 && abs(dx) < -dy
-				}
-				if skip {
-					continue
-				}
-			}
-			px := cx + dx
-			py := cy + dy
-			if px >= 0 && px < s.width && py >= 0 && py < s.height {
-				offset := (py*s.width + px) * rgbaChannels
-				if pacAlpha > pixels[offset+3] {
-					pixels[offset] = 255
-					pixels[offset+1] = 220
-					pixels[offset+2] = 0
-					pixels[offset+3] = pacAlpha
-				}
-			}
-		}
-	}
-}
-
-func (s *pacmanState) drawCircle(pixels []uint8, cx, cy, r int, red, g, b, a uint8) {
-	for dy := -r; dy <= r; dy++ {
-		for dx := -r; dx <= r; dx++ {
-			if dx*dx+dy*dy > r*r {
-				continue
-			}
-			px := cx + dx
-			py := cy + dy
-			if px >= 0 && px < s.width && py >= 0 && py < s.height {
-				offset := (py*s.width + px) * rgbaChannels
-				if a > pixels[offset+3] {
-					pixels[offset] = red
-					pixels[offset+1] = g
-					pixels[offset+2] = b
-					pixels[offset+3] = a
-				}
-			}
-		}
-	}
-}
-
-func (s *pacmanState) drawPacRect(pixels []uint8, x, y, rw, rh int, r, g, b, a uint8) {
-	for dy := range rh {
-		for dx := range rw {
-			px := x + dx
-			py := y + dy
-			if px >= 0 && px < s.width && py >= 0 && py < s.height {
-				offset := (py*s.width + px) * rgbaChannels
-				if a > pixels[offset+3] {
-					pixels[offset] = r
-					pixels[offset+1] = g
-					pixels[offset+2] = b
-					pixels[offset+3] = a
-				}
-			}
-		}
-	}
-}
-
-// --- Tetris Effect ---
-
-const (
-	tetrisRows          = 20
-	tetrisCols          = 10
-	tetrisAlpha         = 70
-	tetrisGhostAlpha    = 20
-	tetrisGridAlpha     = 20
-	tetrisFrameInterval = 400 * time.Millisecond
-)
-
-// Standard Tetris piece shapes as 4x4 bitmaps (row-major).
-// Each piece is defined in its spawn orientation.
-var tetrisPieces = [7][4][4]bool{
-	// I-piece
-	{
-		{false, false, false, false},
-		{true, true, true, true},
-		{false, false, false, false},
-		{false, false, false, false},
-	},
-	// O-piece
-	{
-		{false, true, true, false},
-		{false, true, true, false},
-		{false, false, false, false},
-		{false, false, false, false},
-	},
-	// T-piece
-	{
-		{false, true, false, false},
-		{true, true, true, false},
-		{false, false, false, false},
-		{false, false, false, false},
-	},
-	// S-piece
-	{
-		{false, true, true, false},
-		{true, true, false, false},
-		{false, false, false, false},
-		{false, false, false, false},
-	},
-	// Z-piece
-	{
-		{true, true, false, false},
-		{false, true, true, false},
-		{false, false, false, false},
-		{false, false, false, false},
-	},
-	// J-piece
-	{
-		{true, false, false, false},
-		{true, true, true, false},
-		{false, false, false, false},
-		{false, false, false, false},
-	},
-	// L-piece
-	{
-		{false, false, true, false},
-		{true, true, true, false},
-		{false, false, false, false},
-		{false, false, false, false},
-	},
-}
-
-// NES-style Tetris colors for each piece type.
-var tetrisColors = [7][3]uint8{
-	{0, 240, 240}, // I - cyan
-	{240, 240, 0}, // O - yellow
-	{160, 0, 240}, // T - purple
-	{0, 240, 0},   // S - green
-	{240, 0, 0},   // Z - red
-	{0, 0, 240},   // J - blue
-	{240, 160, 0}, // L - orange
-}
-
-type tetrisCell struct {
-	filled bool
-	color  int // piece type index for color lookup
-}
-
-type tetrisPiece struct {
-	kind   int
-	shape  [4][4]bool
-	x, y   int
-	ghostY int
-}
-
-type tetrisState struct {
-	width, height int
-	cellSize      int
-	offsetX       int // pixel offset to center the grid
-	offsetY       int
-	grid          [tetrisRows][tetrisCols]tetrisCell
-	current       tetrisPiece
-	linesCleared  int
-	flashRows     []int
-	flashTimer    int
-}
-
-func (p *shenanigans) startTetris(pc linkquisition.PickerCanvas) {
-	state := &tetrisState{
-		width:  pc.Width(),
-		height: pc.Height(),
-	}
-	if state.width == 0 {
-		state.width = 600
-	}
-	if state.height == 0 {
-		state.height = 400
-	}
-	state.computeLayout()
-	state.spawnPiece()
-
-	pc.AddRasterOverlay(0.6, func(w, h int) []uint8 {
-		if w != state.width || h != state.height {
-			state.width = w
-			state.height = h
-			state.computeLayout()
-		}
-		return p.invertForLight(state.render())
-	})
-
-	go func() {
-		ticker := time.NewTicker(tetrisFrameInterval)
-		defer ticker.Stop()
-
-		for range ticker.C {
-			if p.stopped.Load() {
-				return
-			}
-			state.update()
-			pc.ScheduleRefresh()
-		}
-	}()
-}
-
-func (s *tetrisState) computeLayout() {
-	// Size cells to fill the window height, using most of the vertical space
-	s.cellSize = (s.height - 4) / tetrisRows
-	if s.cellSize < 4 {
-		s.cellSize = 4
-	}
-	// Also cap by width so the grid doesn't overflow horizontally
-	maxByWidth := (s.width * 3 / 4) / tetrisCols
-	if s.cellSize > maxByWidth {
-		s.cellSize = maxByWidth
-	}
-	// Center the grid horizontally
-	gridW := tetrisCols * s.cellSize
-	s.offsetX = (s.width - gridW) / 2
-	// Center vertically with a small top margin
-	gridH := tetrisRows * s.cellSize
-	s.offsetY = (s.height - gridH) / 2
-}
-
-func (s *tetrisState) spawnPiece() {
-	kind := rand.IntN(len(tetrisPieces))
-	s.current = tetrisPiece{
-		kind:  kind,
-		shape: tetrisPieces[kind],
-		x:     tetrisCols/2 - 2,
-		y:     0,
-	}
-	s.computeGhost()
-}
-
-func (s *tetrisState) computeGhost() {
-	// Drop the piece as far as it can go to find the ghost position
-	ghost := s.current
-	for s.validPosition(ghost.shape, ghost.x, ghost.y+1) {
-		ghost.y++
-	}
-	s.current.ghostY = ghost.y
-}
-
-func (s *tetrisState) update() {
-	// Handle line clear flash
-	if len(s.flashRows) > 0 {
-		s.flashTimer--
-		if s.flashTimer <= 0 {
-			s.clearFlashRows()
-			s.flashRows = nil
-		}
-		return
-	}
-
-	// Try to move the piece down
-	if s.validPosition(s.current.shape, s.current.x, s.current.y+1) {
-		s.current.y++
-	} else {
-		// Lock the piece
-		s.lockPiece()
-		// Check for completed lines
-		s.checkLines()
-		// Spawn a new piece
-		s.spawnPiece()
-		// If the new piece immediately collides, game over — reset
-		if !s.validPosition(s.current.shape, s.current.x, s.current.y) {
-			s.reset()
-		}
-	}
-
-	// AI: try to make smart moves
-	s.aiMove()
-}
-
-func (s *tetrisState) aiMove() {
-	// Simple AI: evaluate all positions and rotations, pick the best one.
-	// We do one horizontal move or rotation per tick to make it look natural.
-	bestX, bestRot := s.findBestPlacement()
-
-	// Rotate toward the best rotation
-	currentRot := 0
-	testShape := tetrisPieces[s.current.kind]
-	for r := range 4 {
-		if testShape == s.current.shape {
-			currentRot = r
-			break
-		}
-		testShape = tetrisRotate(testShape)
-	}
-
-	if currentRot != bestRot {
-		// Rotate once
-		rotated := tetrisRotate(s.current.shape)
-		if s.validPosition(rotated, s.current.x, s.current.y) {
-			s.current.shape = rotated
-			s.computeGhost()
-		}
-		return
-	}
-
-	// Move horizontally toward best position
-	if s.current.x < bestX {
-		if s.validPosition(s.current.shape, s.current.x+1, s.current.y) {
-			s.current.x++
-			s.computeGhost()
-		}
-	} else if s.current.x > bestX {
-		if s.validPosition(s.current.shape, s.current.x-1, s.current.y) {
-			s.current.x--
-			s.computeGhost()
-		}
-	}
-}
-
-func (s *tetrisState) findBestPlacement() (bestX, bestRot int) {
-	bestScore := -1000000
-	bestX = s.current.x
-	bestRot = 0
-
-	shape := tetrisPieces[s.current.kind]
-	for rot := range 4 {
-		for x := -2; x < tetrisCols; x++ {
-			// Drop piece to bottom
-			y := s.current.y
-			for s.validPositionShape(shape, x, y+1) {
-				y++
-			}
-			if !s.validPositionShape(shape, x, y) {
-				continue
-			}
-
-			// Score the placement
-			score := s.scorePlacement(shape, x, y)
-			if score > bestScore {
-				bestScore = score
-				bestX = x
-				bestRot = rot
-			}
-		}
-		shape = tetrisRotate(shape)
-	}
-	return bestX, bestRot
-}
-
-func (s *tetrisState) scorePlacement(shape [4][4]bool, px, py int) int {
-	score := 0
-
-	// Reward lower placement (more toward the bottom)
-	score += py * 4
-
-	// Simulate placing the piece and check for completed lines
-	var tempGrid [tetrisRows][tetrisCols]tetrisCell
-	tempGrid = s.grid
-	for row := range 4 {
-		for col := range 4 {
-			if shape[row][col] {
-				gy := py + row
-				gx := px + col
-				if gy >= 0 && gy < tetrisRows && gx >= 0 && gx < tetrisCols {
-					tempGrid[gy][gx] = tetrisCell{filled: true, color: 0}
-				}
-			}
-		}
-	}
-
-	// Count complete lines
-	lines := 0
-	for row := range tetrisRows {
-		full := true
-		for col := range tetrisCols {
-			if !tempGrid[row][col].filled {
-				full = false
-				break
-			}
-		}
-		if full {
-			lines++
-		}
-	}
-	score += lines * 100
-
-	// Penalize holes (empty cells with filled cells above them)
-	for col := range tetrisCols {
-		foundFilled := false
-		for row := range tetrisRows {
-			if tempGrid[row][col].filled {
-				foundFilled = true
-			} else if foundFilled {
-				score -= 30
-			}
-		}
-	}
-
-	// Penalize height differences between adjacent columns
-	for col := 0; col < tetrisCols-1; col++ {
-		h1 := tetrisColumnHeight(&tempGrid, col)
-		h2 := tetrisColumnHeight(&tempGrid, col+1)
-		diff := h1 - h2
-		if diff < 0 {
-			diff = -diff
-		}
-		score -= diff * 3
-	}
-
-	return score
-}
-
-func tetrisColumnHeight(grid *[tetrisRows][tetrisCols]tetrisCell, col int) int {
-	for row := range tetrisRows {
-		if grid[row][col].filled {
-			return tetrisRows - row
-		}
-	}
-	return 0
-}
-
-func (s *tetrisState) validPosition(shape [4][4]bool, px, py int) bool {
-	return s.validPositionShape(shape, px, py)
-}
-
-func (s *tetrisState) validPositionShape(shape [4][4]bool, px, py int) bool {
-	for row := range 4 {
-		for col := range 4 {
-			if !shape[row][col] {
-				continue
-			}
-			gx := px + col
-			gy := py + row
-			if gx < 0 || gx >= tetrisCols || gy >= tetrisRows {
-				return false
-			}
-			if gy >= 0 && s.grid[gy][gx].filled {
-				return false
-			}
-		}
-	}
-	return true
-}
-
-func (s *tetrisState) lockPiece() {
-	for row := range 4 {
-		for col := range 4 {
-			if s.current.shape[row][col] {
-				gy := s.current.y + row
-				gx := s.current.x + col
-				if gy >= 0 && gy < tetrisRows && gx >= 0 && gx < tetrisCols {
-					s.grid[gy][gx] = tetrisCell{filled: true, color: s.current.kind}
-				}
-			}
-		}
-	}
-}
-
-func (s *tetrisState) checkLines() {
-	var completed []int
-	for row := range tetrisRows {
-		full := true
-		for col := range tetrisCols {
-			if !s.grid[row][col].filled {
-				full = false
-				break
-			}
-		}
-		if full {
-			completed = append(completed, row)
-		}
-	}
-	if len(completed) > 0 {
-		s.flashRows = completed
-		s.flashTimer = 3 // flash for 3 ticks
-		s.linesCleared += len(completed)
-	}
-}
-
-func (s *tetrisState) clearFlashRows() {
-	for _, row := range s.flashRows {
-		// Move everything above down by one
-		for r := row; r > 0; r-- {
-			s.grid[r] = s.grid[r-1]
-		}
-		// Clear top row
-		for col := range tetrisCols {
-			s.grid[0][col] = tetrisCell{}
-		}
-	}
-}
-
-func (s *tetrisState) reset() {
-	s.grid = [tetrisRows][tetrisCols]tetrisCell{}
-	s.linesCleared = 0
-	s.flashRows = nil
-	s.spawnPiece()
-}
-
-func tetrisRotate(shape [4][4]bool) [4][4]bool {
-	var rotated [4][4]bool
-	for row := range 4 {
-		for col := range 4 {
-			rotated[col][3-row] = shape[row][col]
-		}
-	}
-	return rotated
-}
-
-func (s *tetrisState) render() []uint8 {
-	w, h := s.width, s.height
-	if w == 0 || h == 0 {
-		return make([]uint8, rgbaChannels)
-	}
-
-	pixels := make([]uint8, w*h*rgbaChannels)
-
-	// Draw grid border
-	s.drawGridBorder(pixels)
-
-	// Draw locked cells
-	for row := range tetrisRows {
-		for col := range tetrisCols {
-			cell := s.grid[row][col]
-			if cell.filled {
-				// Flash effect on completed rows
-				if s.isFlashRow(row) && s.flashTimer%2 == 0 {
-					s.drawTetrisCell(pixels, col, row, 255, 255, 255, tetrisAlpha+20)
-				} else {
-					c := tetrisColors[cell.color]
-					s.drawTetrisCell(pixels, col, row, c[0], c[1], c[2], tetrisAlpha)
-				}
-			}
-		}
-	}
-
-	// Draw ghost piece
-	if len(s.flashRows) == 0 {
-		c := tetrisColors[s.current.kind]
-		for row := range 4 {
-			for col := range 4 {
-				if s.current.shape[row][col] {
-					gy := s.current.ghostY + row
-					gx := s.current.x + col
-					if gy >= 0 && gy < tetrisRows && gx >= 0 && gx < tetrisCols {
-						s.drawTetrisCell(pixels, gx, gy, c[0], c[1], c[2], tetrisGhostAlpha)
-					}
-				}
-			}
-		}
-
-		// Draw current piece
-		for row := range 4 {
-			for col := range 4 {
-				if s.current.shape[row][col] {
-					gy := s.current.y + row
-					gx := s.current.x + col
-					if gy >= 0 && gy < tetrisRows && gx >= 0 && gx < tetrisCols {
-						s.drawTetrisCell(pixels, gx, gy, c[0], c[1], c[2], tetrisAlpha)
-					}
-				}
-			}
-		}
-	}
-
-	return pixels
-}
-
-func (s *tetrisState) isFlashRow(row int) bool {
-	for _, r := range s.flashRows {
-		if r == row {
-			return true
-		}
-	}
-	return false
-}
-
-func (s *tetrisState) drawTetrisCell(pixels []uint8, col, row int, r, g, b, a uint8) {
-	px := s.offsetX + col*s.cellSize
-	py := s.offsetY + row*s.cellSize
-
-	// Draw filled cell with 1px border for grid look
-	for dy := 1; dy < s.cellSize-1; dy++ {
-		for dx := 1; dx < s.cellSize-1; dx++ {
-			x := px + dx
-			y := py + dy
-			if x >= 0 && x < s.width && y >= 0 && y < s.height {
-				offset := (y*s.width + x) * rgbaChannels
-				if a > pixels[offset+3] {
-					pixels[offset] = r
-					pixels[offset+1] = g
-					pixels[offset+2] = b
-					pixels[offset+3] = a
-				}
-			}
-		}
-	}
-
-	// Draw highlight on top edge for subtle 3D bevel effect
-	highlight := uint8(min(int(a)+20, 255))
-	for dx := 1; dx < s.cellSize-1; dx++ {
-		x := px + dx
-		y := py + 1
-		if x >= 0 && x < s.width && y >= 0 && y < s.height {
-			offset := (y*s.width + x) * rgbaChannels
-			if highlight > pixels[offset+3] {
-				pixels[offset] = min(r+30, 255)
-				pixels[offset+1] = min(g+30, 255)
-				pixels[offset+2] = min(b+30, 255)
-				pixels[offset+3] = highlight
-			}
-		}
-	}
-}
-
-func (s *tetrisState) drawGridBorder(pixels []uint8) {
-	// Draw a subtle border around the playing field
-	left := s.offsetX - 1
-	right := s.offsetX + tetrisCols*s.cellSize
-	top := s.offsetY - 1
-	bottom := s.offsetY + tetrisRows*s.cellSize
-
-	for y := top; y <= bottom; y++ {
-		if y < 0 || y >= s.height {
-			continue
-		}
-		// Left border
-		if left >= 0 && left < s.width {
-			offset := (y*s.width + left) * rgbaChannels
-			pixels[offset] = 100
-			pixels[offset+1] = 100
-			pixels[offset+2] = 100
-			pixels[offset+3] = tetrisGridAlpha + 20
-		}
-		// Right border
-		if right >= 0 && right < s.width {
-			offset := (y*s.width + right) * rgbaChannels
-			pixels[offset] = 100
-			pixels[offset+1] = 100
-			pixels[offset+2] = 100
-			pixels[offset+3] = tetrisGridAlpha + 20
-		}
-	}
-	for x := left; x <= right; x++ {
-		if x < 0 || x >= s.width {
-			continue
-		}
-		// Bottom border
-		if bottom >= 0 && bottom < s.height {
-			offset := (bottom*s.width + x) * rgbaChannels
-			pixels[offset] = 100
-			pixels[offset+1] = 100
-			pixels[offset+2] = 100
-			pixels[offset+3] = tetrisGridAlpha + 20
-		}
-	}
-}
-
-// --- Frogger Effect ---
-
-const (
-	froggerLanes         = 10
-	froggerAlpha         = 60
-	froggerFrogAlpha     = 80
-	froggerFrameInterval = 50 * time.Millisecond
-	froggerHopInterval   = 12 // frames between frog hops
-)
-
-// Lane types for the frogger grid.
-const (
-	froggerLaneRoad  = 0
-	froggerLaneRiver = 1
-	froggerLaneSafe  = 2
-)
-
-type froggerVehicle struct {
-	x     float64
-	speed float64
-	width int
-	color [3]uint8
-}
-
-type froggerLog struct {
-	x     float64
-	speed float64
-	width int
-}
-
-type froggerLane struct {
-	kind     int
-	vehicles []froggerVehicle
-	logs     []froggerLog
-}
-
-type froggerState struct {
-	width, height int
-	laneHeight    int
-	offsetY       int
-	lanes         []froggerLane
-	frogX, frogY  int // grid positions (lane index + column)
-	frogPixelX    float64
-	hopTimer      int
-	frameCount    int
-	frogSize      int
-	colWidth      int
-	gridCols      int
-}
-
-func (p *shenanigans) startFrogger(pc linkquisition.PickerCanvas) {
-	state := &froggerState{
-		width:  pc.Width(),
-		height: pc.Height(),
-	}
-	if state.width == 0 {
-		state.width = 600
-	}
-	if state.height == 0 {
-		state.height = 400
-	}
-	state.init()
-
-	pc.AddRasterOverlay(0.6, func(w, h int) []uint8 {
-		if w != state.width || h != state.height {
-			state.width = w
-			state.height = h
-			state.init()
-		}
-		return p.invertForLight(state.render())
-	})
-
-	go func() {
-		ticker := time.NewTicker(froggerFrameInterval)
-		defer ticker.Stop()
-
-		for range ticker.C {
-			if p.stopped.Load() {
-				return
-			}
-			state.update()
-			pc.ScheduleRefresh()
-		}
-	}()
-}
-
-func (s *froggerState) init() {
-	// Layout: lanes fill the window vertically
-	totalLanes := froggerLanes + 2 // +2 for start and goal safe zones
-	s.laneHeight = s.height / totalLanes
-	if s.laneHeight < 10 {
-		s.laneHeight = 10
-	}
-	s.offsetY = (s.height - totalLanes*s.laneHeight) / 2
-	s.frogSize = s.laneHeight * 2 / 3
-	s.gridCols = 13
-	s.colWidth = s.width / s.gridCols
-
-	// Build lanes: safe, 4 road, safe (median), 4 river, safe (goal)
-	s.lanes = make([]froggerLane, totalLanes)
-	s.lanes[0] = froggerLane{kind: froggerLaneSafe}                // start
-	s.lanes[totalLanes-1] = froggerLane{kind: froggerLaneSafe}     // goal
-	s.lanes[froggerLanes/2+1] = froggerLane{kind: froggerLaneSafe} // median
-
-	// Road lanes (bottom half)
-	roadColors := [][3]uint8{
-		{200, 60, 60},  // red car
-		{60, 60, 200},  // blue truck
-		{200, 200, 60}, // yellow car
-		{200, 100, 60}, // orange van
-	}
-	for i := 1; i <= froggerLanes/2; i++ {
-		dir := 1.0
-		if i%2 == 0 {
-			dir = -1.0
-		}
-		speed := (0.8 + rand.Float64()*1.2) * dir
-		count := 2 + rand.IntN(2)
-		vehicles := make([]froggerVehicle, count)
-		spacing := float64(s.width) / float64(count)
-		vWidth := s.colWidth + rand.IntN(s.colWidth)
-		for j := range vehicles {
-			vehicles[j] = froggerVehicle{
-				x:     float64(j) * spacing,
-				speed: speed,
-				width: vWidth,
-				color: roadColors[(i+j)%len(roadColors)],
-			}
-		}
-		s.lanes[i] = froggerLane{kind: froggerLaneRoad, vehicles: vehicles}
-	}
-
-	// River lanes (top half)
-	for i := froggerLanes/2 + 2; i < froggerLanes+1; i++ {
-		dir := 1.0
-		if i%2 == 0 {
-			dir = -1.0
-		}
-		speed := (0.5 + rand.Float64()*0.8) * dir
-		count := 2 + rand.IntN(2)
-		logs := make([]froggerLog, count)
-		spacing := float64(s.width) / float64(count)
-		logWidth := s.colWidth*2 + rand.IntN(s.colWidth)
-		for j := range logs {
-			logs[j] = froggerLog{
-				x:     float64(j) * spacing,
-				speed: speed,
-				width: logWidth,
-			}
-		}
-		s.lanes[i] = froggerLane{kind: froggerLaneRiver, logs: logs}
-	}
-
-	s.resetFrog()
-}
-
-func (s *froggerState) resetFrog() {
-	s.frogX = s.gridCols / 2
-	s.frogY = 0
-	s.frogPixelX = float64(s.frogX * s.colWidth)
-	s.hopTimer = 0
-}
-
-func (s *froggerState) update() {
-	s.frameCount++
-
-	// Move vehicles and logs
-	for i := range s.lanes {
-		lane := &s.lanes[i]
-		for j := range lane.vehicles {
-			lane.vehicles[j].x += lane.vehicles[j].speed
-			// Wrap around
-			if lane.vehicles[j].x > float64(s.width) {
-				lane.vehicles[j].x = -float64(lane.vehicles[j].width)
-			} else if lane.vehicles[j].x < -float64(lane.vehicles[j].width) {
-				lane.vehicles[j].x = float64(s.width)
-			}
-		}
-		for j := range lane.logs {
-			lane.logs[j].x += lane.logs[j].speed
-			if lane.logs[j].x > float64(s.width) {
-				lane.logs[j].x = -float64(lane.logs[j].width)
-			} else if lane.logs[j].x < -float64(lane.logs[j].width) {
-				lane.logs[j].x = float64(s.width)
-			}
-		}
-	}
-
-	// If frog is on a log, move with it
-	if s.frogY > 0 && s.frogY < len(s.lanes) {
-		lane := &s.lanes[s.frogY]
-		if lane.kind == froggerLaneRiver {
-			for _, log := range lane.logs {
-				if s.frogOnLog(log) {
-					s.frogPixelX += log.speed
-					break
-				}
-			}
-		}
-	}
-
-	// Frog AI: hop forward periodically
-	s.hopTimer++
-	if s.hopTimer >= froggerHopInterval {
-		s.hopTimer = 0
-		s.frogAI()
-	}
-
-	// Check if frog reached the goal
-	if s.frogY >= len(s.lanes)-1 {
-		s.resetFrog()
-	}
-
-	// Check if frog is off-screen or hit
-	if s.frogPixelX < -float64(s.frogSize) || s.frogPixelX > float64(s.width) {
-		s.resetFrog()
-		return
-	}
-
-	// Check collision with vehicles
-	if s.frogY > 0 && s.frogY < len(s.lanes) {
-		lane := &s.lanes[s.frogY]
-		if lane.kind == froggerLaneRoad {
-			for _, v := range lane.vehicles {
-				if s.frogHitsVehicle(v) {
-					s.resetFrog()
-					return
-				}
-			}
-		}
-		// Check if frog is in river but not on a log
-		if lane.kind == froggerLaneRiver {
-			onLog := false
-			for _, log := range lane.logs {
-				if s.frogOnLog(log) {
-					onLog = true
-					break
-				}
-			}
-			if !onLog {
-				s.resetFrog()
-				return
-			}
-		}
-	}
-}
-
-func (s *froggerState) frogAI() {
-	// Try to move forward (up), with some lateral dodging
-	nextY := s.frogY + 1
-	if nextY >= len(s.lanes) {
-		s.frogY = nextY
-		return
-	}
-
-	// Check if moving straight forward is safe
-	if s.isSafePosition(s.frogPixelX, nextY) {
-		s.frogY = nextY
-		return
-	}
-
-	// Try left or right
-	leftX := s.frogPixelX - float64(s.colWidth)
-	rightX := s.frogPixelX + float64(s.colWidth)
-
-	if s.isSafePosition(leftX, nextY) {
-		s.frogPixelX = leftX
-		s.frogY = nextY
-	} else if s.isSafePosition(rightX, nextY) {
-		s.frogPixelX = rightX
-		s.frogY = nextY
-	} else if s.isSafePosition(leftX, s.frogY) {
-		// Dodge sideways in current lane
-		s.frogPixelX = leftX
-	} else if s.isSafePosition(rightX, s.frogY) {
-		s.frogPixelX = rightX
-	}
-	// Otherwise stay put and wait
-}
-
-func (s *froggerState) isSafePosition(px float64, laneIdx int) bool {
-	if laneIdx < 0 || laneIdx >= len(s.lanes) {
-		return false
-	}
-	if px < 0 || px > float64(s.width-s.frogSize) {
-		return false
-	}
-	lane := &s.lanes[laneIdx]
-	frogCenter := px + float64(s.frogSize)/2
-
-	if lane.kind == froggerLaneRoad {
-		for _, v := range lane.vehicles {
-			vLeft := v.x
-			vRight := v.x + float64(v.width)
-			if frogCenter > vLeft && frogCenter < vRight {
-				return false
-			}
-		}
-		return true
-	}
-	if lane.kind == froggerLaneRiver {
-		for _, log := range lane.logs {
-			logLeft := log.x
-			logRight := log.x + float64(log.width)
-			if frogCenter > logLeft && frogCenter < logRight {
-				return true
-			}
-		}
-		return false // not on any log
-	}
-	return true // safe lane
-}
-
-func (s *froggerState) frogOnLog(log froggerLog) bool {
-	frogCenter := s.frogPixelX + float64(s.frogSize)/2
-	return frogCenter > log.x && frogCenter < log.x+float64(log.width)
-}
-
-func (s *froggerState) frogHitsVehicle(v froggerVehicle) bool {
-	frogCenter := s.frogPixelX + float64(s.frogSize)/2
-	return frogCenter > v.x && frogCenter < v.x+float64(v.width)
-}
-
-func (s *froggerState) render() []uint8 {
-	w, h := s.width, s.height
-	if w == 0 || h == 0 {
-		return make([]uint8, rgbaChannels)
-	}
-
-	pixels := make([]uint8, w*h*rgbaChannels)
-
-	// Draw lanes
-	for i, lane := range s.lanes {
-		laneTop := s.offsetY + (len(s.lanes)-1-i)*s.laneHeight
-
-		switch lane.kind {
-		case froggerLaneSafe:
-			s.drawLaneRect(pixels, 0, laneTop, w, s.laneHeight, 60, 140, 60, froggerAlpha/2)
-		case froggerLaneRoad:
-			s.drawLaneRect(pixels, 0, laneTop, w, s.laneHeight, 50, 50, 50, froggerAlpha/2)
-			// Lane markings
-			for mx := 0; mx < w; mx += s.colWidth {
-				markY := laneTop + s.laneHeight/2
-				s.drawLaneRect(pixels, mx+s.colWidth/4, markY, s.colWidth/2, 2, 200, 200, 100, froggerAlpha/3)
-			}
-			// Draw vehicles
-			for _, v := range lane.vehicles {
-				vy := laneTop + (s.laneHeight-s.frogSize)/2
-				s.drawLaneRect(pixels, int(v.x), vy, v.width, s.frogSize, v.color[0], v.color[1], v.color[2], froggerAlpha)
-			}
-		case froggerLaneRiver:
-			s.drawLaneRect(pixels, 0, laneTop, w, s.laneHeight, 30, 60, 150, froggerAlpha/2)
-			// Draw logs
-			for _, log := range lane.logs {
-				ly := laneTop + (s.laneHeight-s.frogSize*2/3)/2
-				s.drawLaneRect(pixels, int(log.x), ly, log.width, s.frogSize*2/3, 140, 90, 40, froggerAlpha)
-			}
-		}
-	}
-
-	// Draw frog
-	frogScreenY := s.offsetY + (len(s.lanes)-1-s.frogY)*s.laneHeight + (s.laneHeight-s.frogSize)/2
-	s.drawFrog(pixels, int(s.frogPixelX), frogScreenY)
-
-	return pixels
-}
-
-func (s *froggerState) drawFrog(pixels []uint8, fx, fy int) {
-	// Simple frog shape: a green square with eyes
-	size := s.frogSize
-	// Body
-	for dy := 2; dy < size-2; dy++ {
-		for dx := 2; dx < size-2; dx++ {
-			px := fx + dx
-			py := fy + dy
-			if px >= 0 && px < s.width && py >= 0 && py < s.height {
-				offset := (py*s.width + px) * rgbaChannels
-				if froggerFrogAlpha > pixels[offset+3] {
-					pixels[offset] = 50
-					pixels[offset+1] = 200
-					pixels[offset+2] = 50
-					pixels[offset+3] = froggerFrogAlpha
-				}
-			}
-		}
-	}
-	// Eyes (two small white dots near the top)
-	eyeY := fy + size/4
-	eyeSize := max(size/6, 2)
-	s.drawFrogEye(pixels, fx+size/4, eyeY, eyeSize)
-	s.drawFrogEye(pixels, fx+size*3/4-eyeSize, eyeY, eyeSize)
-}
-
-func (s *froggerState) drawFrogEye(pixels []uint8, ex, ey, size int) {
-	for dy := range size {
-		for dx := range size {
-			px := ex + dx
-			py := ey + dy
-			if px >= 0 && px < s.width && py >= 0 && py < s.height {
-				offset := (py*s.width + px) * rgbaChannels
-				a := uint8(froggerFrogAlpha + 20)
-				if a > pixels[offset+3] {
-					pixels[offset] = 255
-					pixels[offset+1] = 255
-					pixels[offset+2] = 255
-					pixels[offset+3] = a
-				}
-			}
-		}
-	}
-}
-
-func (s *froggerState) drawLaneRect(pixels []uint8, x, y, rw, rh int, r, g, b, a uint8) {
-	for dy := range rh {
-		for dx := range rw {
-			px := x + dx
-			py := y + dy
-			if px >= 0 && px < s.width && py >= 0 && py < s.height {
-				offset := (py*s.width + px) * rgbaChannels
-				if a > pixels[offset+3] {
-					pixels[offset] = r
-					pixels[offset+1] = g
-					pixels[offset+2] = b
-					pixels[offset+3] = a
-				}
-			}
-		}
-	}
-}
-
-// --- Minesweeper Effect ---
-
-const (
-	minesweeperMineRatio     = 0.15
-	minesweeperAlpha         = 60
-	minesweeperRevealedAlpha = 50
-	minesweeperNumberAlpha   = 70
-	minesweeperFrameInterval = 120 * time.Millisecond
-	minesweeperTargetCols    = 20
-)
-
-// Cell states for minesweeper.
-const (
-	msCellHidden   = 0
-	msCellRevealed = 1
-	msCellExploded = 3
-)
-
-type minesweeperCell struct {
-	mine     bool
-	state    int
-	adjacent int
-}
-
-type minesweeperState struct {
-	width, height int
-	cols, rows    int
-	cellSize      int
-	offsetX       int
-	offsetY       int
-	grid          [][]minesweeperCell
-	revealQueue   []msPoint
-	gameOver      bool
-	resetTimer    int
-	frameCount    int
-}
-
-type msPoint struct {
-	x, y int
-}
-
-func (p *shenanigans) startMinesweeper(pc linkquisition.PickerCanvas) {
-	state := &minesweeperState{
-		width:  pc.Width(),
-		height: pc.Height(),
-	}
-	if state.width == 0 {
-		state.width = 600
-	}
-	if state.height == 0 {
-		state.height = 400
-	}
-	state.init()
-
-	pc.AddRasterOverlay(0.6, func(w, h int) []uint8 {
-		if w != state.width || h != state.height {
-			state.width = w
-			state.height = h
-			state.init()
-		}
-		return p.invertForLight(state.render())
-	})
-
-	go func() {
-		ticker := time.NewTicker(minesweeperFrameInterval)
-		defer ticker.Stop()
-
-		for range ticker.C {
-			if p.stopped.Load() {
-				return
-			}
-			state.update()
-			pc.ScheduleRefresh()
-		}
-	}()
-}
-
-func (s *minesweeperState) init() {
-	s.cellSize = s.width / minesweeperTargetCols
-	if s.cellSize < 8 {
-		s.cellSize = 8
-	}
-	if s.cellSize > 20 {
-		s.cellSize = 20
-	}
-	s.cols = s.width / s.cellSize
-	s.rows = s.height / s.cellSize
-	s.offsetX = (s.width - s.cols*s.cellSize) / 2
-	s.offsetY = (s.height - s.rows*s.cellSize) / 2
-	s.reset()
-}
-
-func (s *minesweeperState) reset() {
-	s.grid = make([][]minesweeperCell, s.rows)
-	for r := range s.grid {
-		s.grid[r] = make([]minesweeperCell, s.cols)
-	}
-
-	// Place mines
-	totalCells := s.rows * s.cols
-	mineCount := int(float64(totalCells) * minesweeperMineRatio)
-	for placed := 0; placed < mineCount; {
-		x := rand.IntN(s.cols)
-		y := rand.IntN(s.rows)
-		if !s.grid[y][x].mine {
-			s.grid[y][x].mine = true
-			placed++
-		}
-	}
-
-	// Calculate adjacency numbers
-	for r := range s.rows {
-		for c := range s.cols {
-			if s.grid[r][c].mine {
-				continue
-			}
-			count := 0
-			for dr := -1; dr <= 1; dr++ {
-				for dc := -1; dc <= 1; dc++ {
-					nr, nc := r+dr, c+dc
-					if nr >= 0 && nr < s.rows && nc >= 0 && nc < s.cols && s.grid[nr][nc].mine {
-						count++
-					}
-				}
-			}
-			s.grid[r][c].adjacent = count
-		}
-	}
-
-	// Start by revealing a random safe cell (no adjacent mines) to seed the flood
-	s.revealQueue = nil
-	s.gameOver = false
-	s.resetTimer = 0
-	s.seedReveal()
-}
-
-func (s *minesweeperState) seedReveal() {
-	// Find a cell with 0 adjacent mines to start a nice flood fill
-	candidates := make([]msPoint, 0, 32)
-	for r := range s.rows {
-		for c := range s.cols {
-			if !s.grid[r][c].mine && s.grid[r][c].adjacent == 0 {
-				candidates = append(candidates, msPoint{c, r})
-			}
-		}
-	}
-	if len(candidates) > 0 {
-		start := candidates[rand.IntN(len(candidates))]
-		s.revealQueue = append(s.revealQueue, start)
-	} else {
-		// Fallback: pick any non-mine cell
-		for r := range s.rows {
-			for c := range s.cols {
-				if !s.grid[r][c].mine {
-					s.revealQueue = append(s.revealQueue, msPoint{c, r})
-					return
-				}
-			}
-		}
-	}
-}
-
-func (s *minesweeperState) update() {
-	s.frameCount++
-
-	if s.gameOver {
-		s.resetTimer--
-		if s.resetTimer <= 0 {
-			s.reset()
-		}
-		return
-	}
-
-	// Reveal cells from the queue (a few per frame for wave effect)
-	revealPerFrame := max(1, (s.cols*s.rows)/80)
-	for range revealPerFrame {
-		if len(s.revealQueue) == 0 {
-			break
-		}
-		// Pop from queue
-		pt := s.revealQueue[0]
-		s.revealQueue = s.revealQueue[1:]
-
-		cell := &s.grid[pt.y][pt.x]
-		if cell.state != msCellHidden {
-			continue
-		}
-
-		if cell.mine {
-			// Hit a mine — game over!
-			cell.state = msCellExploded
-			s.gameOver = true
-			s.resetTimer = 15
-			return
-		}
-
-		cell.state = msCellRevealed
-
-		// If no adjacent mines, flood fill neighbors
-		if cell.adjacent == 0 {
-			for dr := -1; dr <= 1; dr++ {
-				for dc := -1; dc <= 1; dc++ {
-					nr, nc := pt.y+dr, pt.x+dc
-					if nr >= 0 && nr < s.rows && nc >= 0 && nc < s.cols {
-						if s.grid[nr][nc].state == msCellHidden {
-							s.revealQueue = append(s.revealQueue, msPoint{nc, nr})
-						}
-					}
-				}
-			}
-		}
-	}
-
-	// If queue is empty and game not over, pick a new unrevealed safe cell
-	if len(s.revealQueue) == 0 && !s.gameOver {
-		s.pickNextReveal()
-	}
-}
-
-func (s *minesweeperState) pickNextReveal() {
-	// Prefer cells adjacent to already-revealed cells
-	for r := range s.rows {
-		for c := range s.cols {
-			if s.grid[r][c].state == msCellHidden && !s.grid[r][c].mine {
-				if s.hasRevealedNeighbor(r, c) {
-					s.revealQueue = append(s.revealQueue, msPoint{c, r})
-					return
-				}
-			}
-		}
-	}
-	// Fallback: pick any hidden non-mine cell
-	for r := range s.rows {
-		for c := range s.cols {
-			if s.grid[r][c].state == msCellHidden && !s.grid[r][c].mine {
-				s.revealQueue = append(s.revealQueue, msPoint{c, r})
-				return
-			}
-		}
-	}
-	// All safe cells revealed — occasionally step on a mine for fun, then reset
-	for r := range s.rows {
-		for c := range s.cols {
-			if s.grid[r][c].state == msCellHidden && s.grid[r][c].mine {
-				s.revealQueue = append(s.revealQueue, msPoint{c, r})
-				return
-			}
-		}
-	}
-}
-
-func (s *minesweeperState) hasRevealedNeighbor(r, c int) bool {
-	for dr := -1; dr <= 1; dr++ {
-		for dc := -1; dc <= 1; dc++ {
-			nr, nc := r+dr, c+dc
-			if nr >= 0 && nr < s.rows && nc >= 0 && nc < s.cols {
-				if s.grid[nr][nc].state == msCellRevealed {
-					return true
-				}
-			}
-		}
-	}
-	return false
-}
-
-func (s *minesweeperState) render() []uint8 {
-	w, h := s.width, s.height
-	if w == 0 || h == 0 {
-		return make([]uint8, rgbaChannels)
-	}
-
-	pixels := make([]uint8, w*h*rgbaChannels)
-
-	for r := range s.rows {
-		for c := range s.cols {
-			cell := s.grid[r][c]
-			px := s.offsetX + c*s.cellSize
-			py := s.offsetY + r*s.cellSize
-
-			switch cell.state {
-			case msCellHidden:
-				// Raised button look
-				s.drawMSCell(pixels, px, py, 130, 130, 130, minesweeperAlpha, true)
-			case msCellRevealed:
-				// Flat revealed cell
-				s.drawMSCell(pixels, px, py, 180, 180, 180, minesweeperRevealedAlpha, false)
-				// Draw number if > 0
-				if cell.adjacent > 0 {
-					s.drawMSNumber(pixels, px, py, cell.adjacent)
-				}
-			case msCellExploded:
-				// Red explosion
-				s.drawMSCell(pixels, px, py, 220, 50, 50, minesweeperAlpha+20, false)
-				s.drawMSMine(pixels, px, py)
-			}
-		}
-	}
-
-	// If game over, reveal all mines
-	if s.gameOver {
-		for r := range s.rows {
-			for c := range s.cols {
-				if s.grid[r][c].mine && s.grid[r][c].state != msCellExploded {
-					px := s.offsetX + c*s.cellSize
-					py := s.offsetY + r*s.cellSize
-					s.drawMSCell(pixels, px, py, 180, 180, 180, minesweeperRevealedAlpha, false)
-					s.drawMSMine(pixels, px, py)
-				}
-			}
-		}
-	}
-
-	return pixels
-}
-
-func (s *minesweeperState) drawMSCell(pixels []uint8, px, py int, r, g, b, a uint8, raised bool) {
-	cs := s.cellSize
-	for dy := 0; dy < cs; dy++ {
-		for dx := 0; dx < cs; dx++ {
-			x := px + dx
-			y := py + dy
-			if x < 0 || x >= s.width || y < 0 || y >= s.height {
-				continue
-			}
-			cr, cg, cb := r, g, b
-			// Border effect for raised cells
-			if raised {
-				if dx == 0 || dy == 0 {
-					cr = min(r+40, 255)
-					cg = min(g+40, 255)
-					cb = min(b+40, 255)
-				} else if dx == cs-1 || dy == cs-1 {
-					cr = r - min(r, 40)
-					cg = g - min(g, 40)
-					cb = b - min(b, 40)
-				}
-			} else if dx == 0 || dy == 0 || dx == cs-1 || dy == cs-1 {
-				// Flat border
-				cr = r - min(r, 20)
-				cg = g - min(g, 20)
-				cb = b - min(b, 20)
-			}
-			offset := (y*s.width + x) * rgbaChannels
-			if a > pixels[offset+3] {
-				pixels[offset] = cr
-				pixels[offset+1] = cg
-				pixels[offset+2] = cb
-				pixels[offset+3] = a
-			}
-		}
-	}
-}
-
-// Minesweeper number colors (classic Windows style).
-var msNumberColors = [8][3]uint8{
-	{0, 0, 200},     // 1 - blue
-	{0, 130, 0},     // 2 - green
-	{200, 0, 0},     // 3 - red
-	{0, 0, 130},     // 4 - dark blue
-	{130, 0, 0},     // 5 - dark red
-	{0, 130, 130},   // 6 - teal
-	{100, 100, 100}, // 7 - gray
-	{60, 60, 60},    // 8 - dark gray
-}
-
-func (s *minesweeperState) drawMSNumber(pixels []uint8, px, py, num int) {
-	if num < 1 || num > 8 {
-		return
-	}
-	color := msNumberColors[num-1]
-	cs := s.cellSize
-
-	// Draw a simple representation of the number using a small centered block pattern
-	// Each digit is a 3x5 bitmap
-	digitBitmaps := [9][5]uint8{
-		{},                                // 0 (unused)
-		{0x4, 0x4, 0x4, 0x4, 0x4},        // 1: center column
-		{0x7, 0x1, 0x7, 0x4, 0x7},        // 2
-		{0x7, 0x1, 0x7, 0x1, 0x7},        // 3
-		{0x5, 0x5, 0x7, 0x1, 0x1},        // 4
-		{0x7, 0x4, 0x7, 0x1, 0x7},        // 5
-		{0x7, 0x4, 0x7, 0x5, 0x7},        // 6
-		{0x7, 0x1, 0x1, 0x1, 0x1},        // 7
-		{0x7, 0x5, 0x7, 0x5, 0x7},        // 8
-	}
-
-	bitmap := digitBitmaps[num]
-	// Scale the 3x5 bitmap to fit within the cell
-	pixSize := max(cs/6, 1)
-	startX := px + (cs-3*pixSize)/2
-	startY := py + (cs-5*pixSize)/2
-
-	for row := range 5 {
-		for col := range 3 {
-			if bitmap[row]&(1<<(2-col)) != 0 {
-				for dy := range pixSize {
-					for dx := range pixSize {
-						x := startX + col*pixSize + dx
-						y := startY + row*pixSize + dy
-						if x >= 0 && x < s.width && y >= 0 && y < s.height {
-							offset := (y*s.width + x) * rgbaChannels
-							if minesweeperNumberAlpha > pixels[offset+3] {
-								pixels[offset] = color[0]
-								pixels[offset+1] = color[1]
-								pixels[offset+2] = color[2]
-								pixels[offset+3] = minesweeperNumberAlpha
-							}
-						}
-					}
-				}
-			}
-		}
-	}
-}
-
-func (s *minesweeperState) drawMSMine(pixels []uint8, px, py int) {
-	cs := s.cellSize
-	cx := px + cs/2
-	cy := py + cs/2
-	radius := cs / 4
-
-	// Draw a small circle for the mine
-	for dy := -radius; dy <= radius; dy++ {
-		for dx := -radius; dx <= radius; dx++ {
-			if dx*dx+dy*dy > radius*radius {
-				continue
-			}
-			x := cx + dx
-			y := cy + dy
-			if x >= 0 && x < s.width && y >= 0 && y < s.height {
-				offset := (y*s.width + x) * rgbaChannels
-				a := uint8(minesweeperAlpha + 20)
-				if a > pixels[offset+3] {
-					pixels[offset] = 30
-					pixels[offset+1] = 30
-					pixels[offset+2] = 30
-					pixels[offset+3] = a
-				}
-			}
-		}
-	}
-
-	// Draw spikes (4 lines)
-	for d := -radius - 1; d <= radius+1; d++ {
-		points := [][2]int{
-			{cx + d, cy}, {cx, cy + d},
-			{cx + d*7/10, cy + d*7/10}, {cx + d*7/10, cy - d*7/10},
-		}
-		for _, pt := range points {
-			x, y := pt[0], pt[1]
-			if x >= 0 && x < s.width && y >= 0 && y < s.height {
-				offset := (y*s.width + x) * rgbaChannels
-				a := uint8(minesweeperAlpha + 10)
-				if a > pixels[offset+3] {
-					pixels[offset] = 30
-					pixels[offset+1] = 30
-					pixels[offset+2] = 30
-					pixels[offset+3] = a
-				}
-			}
-		}
-	}
-}
-
-// --- Flappy Bird Effect ---
-
-const (
-	flappyGravity       = 0.35
-	flappyJumpVelocity  = -5.0
-	flappyPipeSpeed     = 2.0
-	flappyPipeGap       = 0.28 // fraction of window height
-	flappyPipeWidth     = 40
-	flappyBirdSize      = 14
-	flappyAlpha         = 65
-	flappyBirdAlpha     = 80
-	flappyFrameInterval = 25 * time.Millisecond
-	flappyPipeSpacing   = 180 // pixels between pipe centers
-)
-
-type flappyPipe struct {
-	x      float64
-	gapY   float64 // center of the gap (0-1 normalized)
-	scored bool
-}
-
-type flappyState struct {
-	width, height int
-
-	// Bird
-	birdY  float64
-	birdVY float64
-	birdX  float64
-
-	// Pipes
-	pipes []flappyPipe
-
-	// Game state
-	score      int
-	frameCount int
-	dead       bool
-	deadTimer  int
-}
-
-func (p *shenanigans) startFlappy(pc linkquisition.PickerCanvas) {
-	state := &flappyState{
-		width:  pc.Width(),
-		height: pc.Height(),
-	}
-	if state.width == 0 {
-		state.width = 600
-	}
-	if state.height == 0 {
-		state.height = 400
-	}
-	state.reset()
-
-	pc.AddRasterOverlay(0.6, func(w, h int) []uint8 {
-		if w != state.width || h != state.height {
-			state.width = w
-			state.height = h
-			state.reset()
-		}
-		return p.invertForLight(state.render())
-	})
-
-	go func() {
-		ticker := time.NewTicker(flappyFrameInterval)
-		defer ticker.Stop()
-
-		for range ticker.C {
-			if p.stopped.Load() {
-				return
-			}
-			state.update()
-			pc.ScheduleRefresh()
-		}
-	}()
-}
-
-func (s *flappyState) reset() {
-	s.birdX = float64(s.width) * 0.25
-	s.birdY = float64(s.height) * 0.5
-	s.birdVY = 0
-	s.score = 0
-	s.dead = false
-	s.deadTimer = 0
-	s.frameCount = 0
-
-	// Initialize pipes — enough to fill the screen plus buffer off the right edge
-	s.pipes = nil
-	pipeCount := s.width/flappyPipeSpacing + 3
-	startX := float64(s.width) + float64(flappyPipeSpacing)/2
-	for i := range pipeCount {
-		s.pipes = append(s.pipes, flappyPipe{
-			x:    startX + float64(i*flappyPipeSpacing),
-			gapY: 0.3 + rand.Float64()*0.4,
-		})
-	}
-}
-
-func (s *flappyState) update() {
-	s.frameCount++
-
-	if s.dead {
-		s.deadTimer--
-		if s.deadTimer <= 0 {
-			s.reset()
-		}
-		return
-	}
-
-	// Apply gravity
-	s.birdVY += flappyGravity
-	s.birdY += s.birdVY
-
-	// AI: flap when needed
-	s.flappyAI()
-
-	// Move pipes
-	for i := range s.pipes {
-		s.pipes[i].x -= flappyPipeSpeed
-	}
-
-	// Recycle pipes that have scrolled off-screen
-	if len(s.pipes) > 0 && s.pipes[0].x < -float64(flappyPipeWidth) {
-		s.pipes = s.pipes[1:]
-		// Add a new pipe — ensure it spawns at or beyond the right edge
-		newX := s.pipes[len(s.pipes)-1].x + float64(flappyPipeSpacing)
-		rightEdge := float64(s.width)
-		if newX < rightEdge {
-			newX = rightEdge
-		}
-		s.pipes = append(s.pipes, flappyPipe{
-			x:    newX,
-			gapY: 0.25 + rand.Float64()*0.5,
-		})
-	}
-
-	// Check collisions
-	if s.checkCollision() {
-		s.dead = true
-		s.deadTimer = 30
-		return
-	}
-
-	// Score
-	for i := range s.pipes {
-		if !s.pipes[i].scored && s.pipes[i].x+float64(flappyPipeWidth) < s.birdX {
-			s.pipes[i].scored = true
-			s.score++
-		}
-	}
-}
-
-func (s *flappyState) flappyAI() {
-	// Look at the next pipe ahead of the bird
-	var nextPipe *flappyPipe
-	for i := range s.pipes {
-		if s.pipes[i].x+float64(flappyPipeWidth) > s.birdX {
-			nextPipe = &s.pipes[i]
-			break
-		}
-	}
-	if nextPipe == nil {
-		return
-	}
-
-	// Target the center of the gap
-	gapCenterY := nextPipe.gapY * float64(s.height)
-	gapHalf := flappyPipeGap * float64(s.height) / 2
-
-	// Flap if bird is below the target zone or falling too fast toward it
-	targetY := gapCenterY - gapHalf*0.2 // aim slightly above center
-	if s.birdY > targetY && s.birdVY > -2.0 {
-		s.birdVY = flappyJumpVelocity
-	}
-
-	// Don't let bird go above screen
-	if s.birdY < float64(flappyBirdSize) {
-		s.birdVY = 1.0
-	}
-}
-
-func (s *flappyState) checkCollision() bool {
-	h := float64(s.height)
-
-	// Floor/ceiling
-	if s.birdY > h-float64(flappyBirdSize) || s.birdY < 0 {
-		return true
-	}
-
-	// Pipe collision
-	birdLeft := s.birdX - float64(flappyBirdSize)/2
-	birdRight := s.birdX + float64(flappyBirdSize)/2
-	birdTop := s.birdY - float64(flappyBirdSize)/2
-	birdBottom := s.birdY + float64(flappyBirdSize)/2
-
-	for _, pipe := range s.pipes {
-		pipeLeft := pipe.x
-		pipeRight := pipe.x + float64(flappyPipeWidth)
-
-		// Check horizontal overlap
-		if birdRight > pipeLeft && birdLeft < pipeRight {
-			gapTop := pipe.gapY*h - flappyPipeGap*h/2
-			gapBottom := pipe.gapY*h + flappyPipeGap*h/2
-
-			// Check if bird is outside the gap
-			if birdTop < gapTop || birdBottom > gapBottom {
-				return true
-			}
-		}
-	}
-
-	return false
-}
-
-func (s *flappyState) render() []uint8 {
-	w, h := s.width, s.height
-	if w == 0 || h == 0 {
-		return make([]uint8, rgbaChannels)
-	}
-
-	pixels := make([]uint8, w*h*rgbaChannels)
-
-	// Draw pipes
-	for _, pipe := range s.pipes {
-		s.drawPipe(pixels, pipe)
-	}
-
-	// Draw bird
-	s.drawBird(pixels)
-
-	return pixels
-}
-
-func (s *flappyState) drawPipe(pixels []uint8, pipe flappyPipe) {
-	w, h := s.width, s.height
-	pipeLeft := int(pipe.x)
-	pipeRight := pipeLeft + flappyPipeWidth
-
-	gapTop := int(pipe.gapY*float64(h) - flappyPipeGap*float64(h)/2)
-	gapBottom := int(pipe.gapY*float64(h) + flappyPipeGap*float64(h)/2)
-
-	// Top pipe
-	for py := 0; py < gapTop; py++ {
-		for px := pipeLeft; px < pipeRight; px++ {
-			if px < 0 || px >= w || py < 0 || py >= h {
-				continue
-			}
-			offset := (py*w + px) * rgbaChannels
-			// Green pipe with darker edges
-			r, g, b := uint8(60), uint8(180), uint8(60)
-			if px == pipeLeft || px == pipeRight-1 {
-				r, g, b = 40, 120, 40
-			}
-			if py >= gapTop-4 && py < gapTop {
-				// Lip at the bottom of top pipe
-				r, g, b = 50, 150, 50
-			}
-			if flappyAlpha > pixels[offset+3] {
-				pixels[offset] = r
-				pixels[offset+1] = g
-				pixels[offset+2] = b
-				pixels[offset+3] = flappyAlpha
-			}
-		}
-	}
-
-	// Bottom pipe
-	for py := gapBottom; py < h; py++ {
-		for px := pipeLeft; px < pipeRight; px++ {
-			if px < 0 || px >= w || py < 0 || py >= h {
-				continue
-			}
-			offset := (py*w + px) * rgbaChannels
-			r, g, b := uint8(60), uint8(180), uint8(60)
-			if px == pipeLeft || px == pipeRight-1 {
-				r, g, b = 40, 120, 40
-			}
-			if py >= gapBottom && py < gapBottom+4 {
-				// Lip at the top of bottom pipe
-				r, g, b = 50, 150, 50
-			}
-			if flappyAlpha > pixels[offset+3] {
-				pixels[offset] = r
-				pixels[offset+1] = g
-				pixels[offset+2] = b
-				pixels[offset+3] = flappyAlpha
-			}
-		}
-	}
-}
-
-func (s *flappyState) drawBird(pixels []uint8) {
-	w, h := s.width, s.height
-	cx := int(s.birdX)
-	cy := int(s.birdY)
-	size := flappyBirdSize
-
-	// Bird body (yellow circle-ish)
-	for dy := -size / 2; dy <= size/2; dy++ {
-		for dx := -size / 2; dx <= size/2; dx++ {
-			if dx*dx+dy*dy > (size/2)*(size/2) {
-				continue
-			}
-			px := cx + dx
-			py := cy + dy
-			if px < 0 || px >= w || py < 0 || py >= h {
-				continue
-			}
-			offset := (py*w + px) * rgbaChannels
-			if flappyBirdAlpha > pixels[offset+3] {
-				pixels[offset] = 240
-				pixels[offset+1] = 200
-				pixels[offset+2] = 50
-				pixels[offset+3] = flappyBirdAlpha
-			}
-		}
-	}
-
-	// Eye (white dot with black pupil)
-	eyeX := cx + size/4
-	eyeY := cy - size/6
-	for dy := -2; dy <= 2; dy++ {
-		for dx := -2; dx <= 2; dx++ {
-			px := eyeX + dx
-			py := eyeY + dy
-			if px >= 0 && px < w && py >= 0 && py < h {
-				offset := (py*w + px) * rgbaChannels
-				pixels[offset] = 255
-				pixels[offset+1] = 255
-				pixels[offset+2] = 255
-				pixels[offset+3] = flappyBirdAlpha
-			}
-		}
-	}
-	// Pupil
-	if eyeX+1 >= 0 && eyeX+1 < w && eyeY >= 0 && eyeY < h {
-		offset := (eyeY*w + eyeX + 1) * rgbaChannels
-		pixels[offset] = 0
-		pixels[offset+1] = 0
-		pixels[offset+2] = 0
-		pixels[offset+3] = flappyBirdAlpha
-	}
-
-	// Beak (small orange triangle to the right)
-	for dy := -1; dy <= 1; dy++ {
-		for dx := 0; dx < 4-abs(dy); dx++ {
-			px := cx + size/2 + dx
-			py := cy + dy + 1
-			if px >= 0 && px < w && py >= 0 && py < h {
-				offset := (py*w + px) * rgbaChannels
-				if flappyBirdAlpha > pixels[offset+3] {
-					pixels[offset] = 240
-					pixels[offset+1] = 120
-					pixels[offset+2] = 30
-					pixels[offset+3] = flappyBirdAlpha
-				}
-			}
-		}
-	}
-}
-
-// --- Lava Lamp Effect ---
-
-const (
-	lavaBlobCount     = 6
-	lavaAlpha         = 60
-	lavaFrameInterval = 40 * time.Millisecond
-	lavaThreshold     = 1.0
-	lavaResDiv        = 4 // render at 1/4 resolution for performance
-)
-
-type lavaBlob struct {
-	x, y       float64
-	radius     float64
-	vy         float64
-	phase      float64
-	phaseSpeed float64
-}
-
-type lavaState struct {
-	width, height int
-	blobs         []lavaBlob
-	time          float64
-	hueShift      float64
-}
-
-func (p *shenanigans) startLava(pc linkquisition.PickerCanvas) {
-	state := &lavaState{
-		width:  pc.Width(),
-		height: pc.Height(),
-	}
-	if state.width == 0 {
-		state.width = 600
-	}
-	if state.height == 0 {
-		state.height = 400
-	}
-	state.init()
-
-	pc.AddRasterOverlay(0.6, func(w, h int) []uint8 {
-		if w != state.width || h != state.height {
-			state.width = w
-			state.height = h
-		}
-		return state.render()
-	})
-
-	go func() {
-		ticker := time.NewTicker(lavaFrameInterval)
-		defer ticker.Stop()
-
-		for range ticker.C {
-			if p.stopped.Load() {
-				return
-			}
-			state.update()
-			pc.ScheduleRefresh()
-		}
-	}()
-}
-
-func (s *lavaState) init() {
-	s.blobs = make([]lavaBlob, lavaBlobCount)
-	for i := range s.blobs {
-		s.blobs[i] = lavaBlob{
-			x:          0.2 + rand.Float64()*0.6,
-			y:          rand.Float64(),
-			radius:     0.06 + rand.Float64()*0.08,
-			vy:         0.002 + rand.Float64()*0.003,
-			phase:      rand.Float64() * 6.28,
-			phaseSpeed: 0.01 + rand.Float64()*0.02,
-		}
-		if i%2 == 0 {
-			s.blobs[i].vy = -s.blobs[i].vy
-		}
-	}
-}
-
-func (s *lavaState) update() {
-	s.time += 0.02
-	s.hueShift += 0.003
-
-	for i := range s.blobs {
-		b := &s.blobs[i]
-		b.y += b.vy
-		b.phase += b.phaseSpeed
-		b.x += sinApprox(b.phase) * 0.002
-
-		if b.y < -0.1 {
-			b.vy = absF(b.vy)
-		} else if b.y > 1.1 {
-			b.vy = -absF(b.vy)
-		}
-		if b.x < 0.1 {
-			b.x = 0.1
-		} else if b.x > 0.9 {
-			b.x = 0.9
-		}
-	}
-}
-
-func (s *lavaState) render() []uint8 {
-	w, h := s.width, s.height
-	if w == 0 || h == 0 {
-		return make([]uint8, rgbaChannels)
-	}
-
-	pixels := make([]uint8, w*h*rgbaChannels)
-	step := lavaResDiv
-
-	for py := 0; py < h; py += step {
-		fy := float64(py) / float64(h)
-		for px := 0; px < w; px += step {
-			fx := float64(px) / float64(w)
-
-			field := 0.0
-			for _, b := range s.blobs {
-				dx := fx - b.x
-				dy := fy - b.y
-				distSq := dx*dx + dy*dy
-				if distSq < 0.0001 {
-					distSq = 0.0001
-				}
-				field += (b.radius * b.radius) / distSq
-			}
-
-			if field < lavaThreshold {
-				continue
-			}
-
-			intensity := (field - lavaThreshold) / lavaThreshold
-			if intensity > 1.0 {
-				intensity = 1.0
-			}
-
-			r, g, b := s.lavaColor(intensity, fy)
-			a := uint8(float64(lavaAlpha) * (0.5 + intensity*0.5))
-
-			for dy := 0; dy < step && py+dy < h; dy++ {
-				for dx := 0; dx < step && px+dx < w; dx++ {
-					offset := ((py+dy)*w + px + dx) * rgbaChannels
-					if a > pixels[offset+3] {
-						pixels[offset] = r
-						pixels[offset+1] = g
-						pixels[offset+2] = b
-						pixels[offset+3] = a
-					}
-				}
-			}
-		}
-	}
-
-	return pixels
-}
-
-func (s *lavaState) lavaColor(intensity, fy float64) (uint8, uint8, uint8) {
-	hue := s.hueShift + fy*0.3
-	h := hue - float64(int(hue))
-	var r, g, b float64
-
-	switch {
-	case h < 0.25:
-		t := h / 0.25
-		r, g, b = 1.0, 0.2+t*0.4, 0.0
-	case h < 0.5:
-		t := (h - 0.25) / 0.25
-		r, g, b = 1.0, 0.6-t*0.5, t*0.6
-	case h < 0.75:
-		t := (h - 0.5) / 0.25
-		r, g, b = 1.0-t*0.3, 0.1, 0.6+t*0.3
-	default:
-		t := (h - 0.75) / 0.25
-		r, g, b = 0.7+t*0.3, 0.1+t*0.1, 0.9-t*0.9
-	}
-
-	r = r*0.6 + r*0.4*intensity
-	g = g*0.4 + g*0.6*intensity
-	b = b*0.5 + b*0.5*intensity
-
-	return uint8(r * 255), uint8(g * 255), uint8(b * 255)
-}
-
-// --- Sine Scroller Effect ---
-
-const (
-	sineScrollAlpha         = 55
-	sineScrollBarAlpha      = 35
-	sineScrollFrameInterval = 25 * time.Millisecond
-	sineScrollMessage       = "  NOBODY EXPECTS THE LINKQUISITION!   ...GREETINGS TO ALL DEMOSCENE ENTHUSIASTS...   "
-)
-
-type sineScrollState struct {
-	width, height int
-	scrollX       float64
-	time          float64
-}
-
-func (p *shenanigans) startSineScroll(pc linkquisition.PickerCanvas) {
-	state := &sineScrollState{
-		width:  pc.Width(),
-		height: pc.Height(),
-	}
-	if state.width == 0 {
-		state.width = 600
-	}
-	if state.height == 0 {
-		state.height = 400
-	}
-
-	pc.AddRasterOverlay(0.6, func(w, h int) []uint8 {
-		if w != state.width || h != state.height {
-			state.width = w
-			state.height = h
-		}
-		return p.invertForLight(state.render())
-	})
-
-	go func() {
-		ticker := time.NewTicker(sineScrollFrameInterval)
-		defer ticker.Stop()
-
-		for range ticker.C {
-			if p.stopped.Load() {
-				return
-			}
-			state.update()
-			pc.ScheduleRefresh()
-		}
-	}()
-}
-
-func (s *sineScrollState) update() {
-	s.scrollX += 2.5
-	s.time += 0.04
-}
-
-func (s *sineScrollState) render() []uint8 {
-	w, h := s.width, s.height
-	if w == 0 || h == 0 {
-		return make([]uint8, rgbaChannels)
-	}
-
-	pixels := make([]uint8, w*h*rgbaChannels)
-
-	// Draw copper bars (horizontal color gradient bands that move)
-	s.drawCopperBars(pixels)
-
-	// Draw the sine-scrolling text
-	s.drawSineText(pixels)
-
-	return pixels
-}
-
-func (s *sineScrollState) drawCopperBars(pixels []uint8) {
-	w, h := s.width, s.height
-	barCount := 5
-	barHeight := h / 8
-
-	for i := range barCount {
-		// Each bar oscillates vertically at different speeds
-		centerY := float64(h)/2 + sinApprox(s.time*0.8+float64(i)*1.3)*float64(h)*0.3
-		barTop := int(centerY) - barHeight/2
-
-		for py := barTop; py < barTop+barHeight; py++ {
-			if py < 0 || py >= h {
-				continue
-			}
-			// Distance from bar center for smooth edges
-			dist := absF(float64(py)-centerY) / float64(barHeight/2)
-			if dist > 1.0 {
-				continue
-			}
-			fade := (1.0 - dist) * (1.0 - dist)
-
-			// Color based on bar index and time
-			phase := s.time*0.5 + float64(i)*0.8
-			r := uint8(sinNorm(phase) * 200 * fade)
-			g := uint8(sinNorm(phase+2.09) * 200 * fade)
-			b := uint8(sinNorm(phase+4.19) * 200 * fade)
-			a := uint8(float64(sineScrollBarAlpha) * fade)
-
-			for px := 0; px < w; px++ {
-				offset := (py*w + px) * rgbaChannels
-				if a > pixels[offset+3] {
-					pixels[offset] = r
-					pixels[offset+1] = g
-					pixels[offset+2] = b
-					pixels[offset+3] = a
-				}
-			}
-		}
-	}
-}
-
-func (s *sineScrollState) drawSineText(pixels []uint8) {
-	w, h := s.width, s.height
-	msg := sineScrollMessage
-	charHeight := 14
-
-	// Pixel scale for the font
-	pixScale := max(h/120, 1)
-	scaledCharW := 10 * pixScale
-	scaledCharH := charHeight * pixScale
-	totalScrollW := len(msg) * scaledCharW
-
-	// Current scroll offset within the repeating message
-	scrollOffset := int(s.scrollX) % totalScrollW
-
-	for i, ch := range msg {
-		// Character x position (scrolling)
-		cx := i*scaledCharW - scrollOffset
-
-		// Wrap: if character scrolled off left, wrap to right
-		if cx < -scaledCharW {
-			cx += totalScrollW
-		}
-		if cx > w {
-			continue
-		}
-
-		// Sine wave Y offset per character
-		wave := sinApprox((float64(cx)/float64(w)*4.0+s.time)*3.14159) * float64(h) * 0.15
-		cy := h/2 - scaledCharH/2 + int(wave)
-
-		// Color: rainbow based on position
-		hue := float64(i)/float64(len(msg)) + s.time*0.1
-		r, g, b := sineScrollColor(hue)
-
-		// Draw the character glyph
-		s.drawChar(pixels, ch, cx, cy, pixScale, r, g, b)
-	}
-}
-
-func (s *sineScrollState) drawChar(pixels []uint8, ch rune, cx, cy, scale int, r, g, b uint8) {
-	w, h := s.width, s.height
-	glyph := sineScrollGlyph(ch)
-
-	for row := range 7 {
-		for col := range 5 {
-			if glyph[row]&(1<<(4-col)) == 0 {
-				continue
-			}
-			// Draw scaled pixel
-			for sy := range scale {
-				for sx := range scale {
-					px := cx + col*scale + sx + scale
-					py := cy + row*scale*2 + sy
-					if px >= 0 && px < w && py >= 0 && py < h {
-						offset := (py*w + px) * rgbaChannels
-						if sineScrollAlpha > pixels[offset+3] {
-							pixels[offset] = r
-							pixels[offset+1] = g
-							pixels[offset+2] = b
-							pixels[offset+3] = sineScrollAlpha
-						}
-					}
-				}
-			}
-		}
-	}
-}
-
-// sineScrollColor returns a rainbow color for a given hue [0,1).
-func sineScrollColor(hue float64) (uint8, uint8, uint8) {
-	h := hue - float64(int(hue))
-	r := sinNorm(h*6.28) * 255
-	g := sinNorm(h*6.28+2.09) * 255
-	b := sinNorm(h*6.28+4.19) * 255
-	return uint8(r), uint8(g), uint8(b)
-}
-
-// Simple 5x7 font glyphs for uppercase + space/punctuation.
-func sineScrollGlyph(ch rune) [7]uint8 {
-	if ch >= 'a' && ch <= 'z' {
-		ch = ch - 'a' + 'A'
-	}
-	if g, ok := sineScrollFont[ch]; ok {
-		return g
-	}
-	return [7]uint8{}
-}
-
-var sineScrollFont = map[rune][7]uint8{
-	'A': {0x0E, 0x11, 0x11, 0x1F, 0x11, 0x11, 0x11},
-	'B': {0x1E, 0x11, 0x11, 0x1E, 0x11, 0x11, 0x1E},
-	'C': {0x0E, 0x11, 0x10, 0x10, 0x10, 0x11, 0x0E},
-	'D': {0x1E, 0x11, 0x11, 0x11, 0x11, 0x11, 0x1E},
-	'E': {0x1F, 0x10, 0x10, 0x1E, 0x10, 0x10, 0x1F},
-	'F': {0x1F, 0x10, 0x10, 0x1E, 0x10, 0x10, 0x10},
-	'G': {0x0E, 0x11, 0x10, 0x17, 0x11, 0x11, 0x0E},
-	'H': {0x11, 0x11, 0x11, 0x1F, 0x11, 0x11, 0x11},
-	'I': {0x0E, 0x04, 0x04, 0x04, 0x04, 0x04, 0x0E},
-	'J': {0x07, 0x02, 0x02, 0x02, 0x02, 0x12, 0x0C},
-	'K': {0x11, 0x12, 0x14, 0x18, 0x14, 0x12, 0x11},
-	'L': {0x10, 0x10, 0x10, 0x10, 0x10, 0x10, 0x1F},
-	'M': {0x11, 0x1B, 0x15, 0x15, 0x11, 0x11, 0x11},
-	'N': {0x11, 0x19, 0x15, 0x13, 0x11, 0x11, 0x11},
-	'O': {0x0E, 0x11, 0x11, 0x11, 0x11, 0x11, 0x0E},
-	'P': {0x1E, 0x11, 0x11, 0x1E, 0x10, 0x10, 0x10},
-	'Q': {0x0E, 0x11, 0x11, 0x11, 0x15, 0x12, 0x0D},
-	'R': {0x1E, 0x11, 0x11, 0x1E, 0x14, 0x12, 0x11},
-	'S': {0x0E, 0x11, 0x10, 0x0E, 0x01, 0x11, 0x0E},
-	'T': {0x1F, 0x04, 0x04, 0x04, 0x04, 0x04, 0x04},
-	'U': {0x11, 0x11, 0x11, 0x11, 0x11, 0x11, 0x0E},
-	'V': {0x11, 0x11, 0x11, 0x11, 0x0A, 0x0A, 0x04},
-	'W': {0x11, 0x11, 0x11, 0x15, 0x15, 0x1B, 0x11},
-	'X': {0x11, 0x11, 0x0A, 0x04, 0x0A, 0x11, 0x11},
-	'Y': {0x11, 0x11, 0x0A, 0x04, 0x04, 0x04, 0x04},
-	'Z': {0x1F, 0x01, 0x02, 0x04, 0x08, 0x10, 0x1F},
-	'!': {0x04, 0x04, 0x04, 0x04, 0x04, 0x00, 0x04},
-	'.': {0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x04},
-	' ': {0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00},
-}
-
-// --- Fireflies Effect ---
-
-const (
-	fireflyCount         = 40
-	fireflyAlpha         = 70
-	fireflyFrameInterval = 30 * time.Millisecond
-)
-
-type firefly struct {
-	x, y       float64
-	vx, vy     float64
-	phase      float64
-	phaseSpeed float64
-	glowSize   float64
-	hue        float64
-}
-
-type firefliesState struct {
-	width, height int
-	flies         []firefly
-}
-
-func (p *shenanigans) startFireflies(pc linkquisition.PickerCanvas) {
-	state := &firefliesState{
-		width:  pc.Width(),
-		height: pc.Height(),
-	}
-	if state.width == 0 {
-		state.width = 600
-	}
-	if state.height == 0 {
-		state.height = 400
-	}
-	state.init()
-
-	pc.AddRasterOverlay(0.6, func(w, h int) []uint8 {
-		if w != state.width || h != state.height {
-			state.width = w
-			state.height = h
-			state.init()
-		}
-		return p.invertForLight(state.render())
-	})
-
-	go func() {
-		ticker := time.NewTicker(fireflyFrameInterval)
-		defer ticker.Stop()
-
-		for range ticker.C {
-			if p.stopped.Load() {
-				return
-			}
-			state.update()
-			pc.ScheduleRefresh()
-		}
-	}()
-}
-
-func (s *firefliesState) init() {
-	s.flies = make([]firefly, fireflyCount)
-	for i := range s.flies {
-		s.flies[i] = firefly{
-			x:          rand.Float64() * float64(s.width),
-			y:          rand.Float64() * float64(s.height),
-			vx:         (rand.Float64() - 0.5) * 0.8,
-			vy:         (rand.Float64() - 0.5) * 0.8,
-			phase:      rand.Float64() * 6.28,
-			phaseSpeed: 0.03 + rand.Float64()*0.05,
-			glowSize:   6 + rand.Float64()*8,
-			hue:        0.1 + rand.Float64()*0.15, // warm yellow-green range
-		}
-	}
-}
-
-func (s *firefliesState) update() {
-	w := float64(s.width)
-	h := float64(s.height)
-
-	for i := range s.flies {
-		f := &s.flies[i]
-
-		// Advance blink phase
-		f.phase += f.phaseSpeed
-
-		// Gentle random drift
-		f.vx += (rand.Float64() - 0.5) * 0.1
-		f.vy += (rand.Float64() - 0.5) * 0.1
-
-		// Dampen velocity
-		f.vx *= 0.98
-		f.vy *= 0.98
-
-		// Clamp speed
-		maxSpeed := 1.5
-		if f.vx > maxSpeed {
-			f.vx = maxSpeed
-		} else if f.vx < -maxSpeed {
-			f.vx = -maxSpeed
-		}
-		if f.vy > maxSpeed {
-			f.vy = maxSpeed
-		} else if f.vy < -maxSpeed {
-			f.vy = -maxSpeed
-		}
-
-		f.x += f.vx
-		f.y += f.vy
-
-		// Soft bounce off edges
-		margin := 20.0
-		if f.x < margin {
-			f.vx += 0.1
-		} else if f.x > w-margin {
-			f.vx -= 0.1
-		}
-		if f.y < margin {
-			f.vy += 0.1
-		} else if f.y > h-margin {
-			f.vy -= 0.1
-		}
-	}
-}
-
-func (s *firefliesState) render() []uint8 {
-	w, h := s.width, s.height
-	if w == 0 || h == 0 {
-		return make([]uint8, rgbaChannels)
-	}
-
-	pixels := make([]uint8, w*h*rgbaChannels)
-
-	for _, f := range s.flies {
-		// Brightness oscillates with phase — soft blink
-		brightness := sinApprox(f.phase)
-		if brightness < 0 {
-			brightness = 0 // off half the time
-		}
-		brightness = brightness * brightness // sharper on/off
-
-		if brightness < 0.05 {
-			continue
-		}
-
-		// Glow color (warm yellow-green)
-		r, g, b := fireflyColor(f.hue, brightness)
-		alpha := uint8(float64(fireflyAlpha) * brightness)
-
-		// Draw soft glow
-		cx, cy := int(f.x), int(f.y)
-		radius := int(f.glowSize * (0.5 + brightness*0.5))
-
-		for dy := -radius; dy <= radius; dy++ {
-			for dx := -radius; dx <= radius; dx++ {
-				dist := float64(dx*dx+dy*dy) / float64(radius*radius+1)
-				if dist > 1.0 {
-					continue
-				}
-				// Soft falloff
-				falloff := (1.0 - dist) * (1.0 - dist)
-				px := cx + dx
-				py := cy + dy
-				if px >= 0 && px < w && py >= 0 && py < h {
-					a := uint8(float64(alpha) * falloff)
-					offset := (py*w + px) * rgbaChannels
-					if a > pixels[offset+3] {
-						pixels[offset] = r
-						pixels[offset+1] = g
-						pixels[offset+2] = b
-						pixels[offset+3] = a
-					}
-				}
-			}
-		}
-	}
-
-	return pixels
-}
-
-func fireflyColor(hue, brightness float64) (uint8, uint8, uint8) {
-	// Warm glow: yellow core fading to green at edges
-	r := uint8((0.9 + hue*0.5) * brightness * 255)
-	g := uint8((0.8 + hue) * brightness * 255)
-	b := uint8(0.1 * brightness * 255)
-	return r, g, b
-}
-
-// --- Boids Effect ---
-
-const (
-	boidCount         = 60
-	boidAlpha         = 60
-	boidFrameInterval = 25 * time.Millisecond
-	boidMaxSpeed      = 3.5
-	boidVisualRange   = 60.0
-	boidSeparationDist = 20.0
-)
-
-type boid struct {
-	x, y   float64
-	vx, vy float64
-	hue    float64
-}
-
-type boidsState struct {
-	width, height int
-	flock         []boid
-}
-
-func (p *shenanigans) startBoids(pc linkquisition.PickerCanvas) {
-	state := &boidsState{
-		width:  pc.Width(),
-		height: pc.Height(),
-	}
-	if state.width == 0 {
-		state.width = 600
-	}
-	if state.height == 0 {
-		state.height = 400
-	}
-	state.init()
-
-	pc.AddRasterOverlay(0.6, func(w, h int) []uint8 {
-		if w != state.width || h != state.height {
-			state.width = w
-			state.height = h
-			state.init()
-		}
-		return p.invertForLight(state.render())
-	})
-
-	go func() {
-		ticker := time.NewTicker(boidFrameInterval)
-		defer ticker.Stop()
-
-		for range ticker.C {
-			if p.stopped.Load() {
-				return
-			}
-			state.update()
-			pc.ScheduleRefresh()
-		}
-	}()
-}
-
-func (s *boidsState) init() {
-	s.flock = make([]boid, boidCount)
-	for i := range s.flock {
-		s.flock[i] = boid{
-			x:   rand.Float64() * float64(s.width),
-			y:   rand.Float64() * float64(s.height),
-			vx:  (rand.Float64() - 0.5) * boidMaxSpeed * 2,
-			vy:  (rand.Float64() - 0.5) * boidMaxSpeed * 2,
-			hue: 0.55 + rand.Float64()*0.15, // blue-teal range
-		}
-	}
-}
-
-func (s *boidsState) update() {
-	w := float64(s.width)
-	h := float64(s.height)
-
-	for i := range s.flock {
-		b := &s.flock[i]
-
-		// Flocking forces
-		var sepX, sepY float64 // separation
-		var alignVX, alignVY float64 // alignment
-		var cohX, cohY float64 // cohesion
-		neighbors := 0
-
-		for j := range s.flock {
-			if i == j {
-				continue
-			}
-			other := &s.flock[j]
-			dx := other.x - b.x
-			dy := other.y - b.y
-			dist := dx*dx + dy*dy
-
-			if dist < boidVisualRange*boidVisualRange {
-				neighbors++
-				alignVX += other.vx
-				alignVY += other.vy
-				cohX += other.x
-				cohY += other.y
-
-				// Separation: push away from very close boids
-				if dist < boidSeparationDist*boidSeparationDist && dist > 0.1 {
-					sepX -= dx
-					sepY -= dy
-				}
-			}
-		}
-
-		if neighbors > 0 {
-			nf := float64(neighbors)
-
-			// Alignment: steer toward average velocity
-			alignVX /= nf
-			alignVY /= nf
-			b.vx += (alignVX - b.vx) * 0.05
-
-			b.vy += (alignVY - b.vy) * 0.05
-
-			// Cohesion: steer toward average position
-			cohX /= nf
-			cohY /= nf
-			b.vx += (cohX - b.x) * 0.003
-			b.vy += (cohY - b.y) * 0.003
-		}
-
-		// Separation
-		b.vx += sepX * 0.05
-		b.vy += sepY * 0.05
-
-		// Edge avoidance (soft turn)
-		margin := 50.0
-		turnForce := 0.3
-		if b.x < margin {
-			b.vx += turnForce
-		} else if b.x > w-margin {
-			b.vx -= turnForce
-		}
-		if b.y < margin {
-			b.vy += turnForce
-		} else if b.y > h-margin {
-			b.vy -= turnForce
-		}
-
-		// Speed limit
-		speed := b.vx*b.vx + b.vy*b.vy
-		if speed > boidMaxSpeed*boidMaxSpeed {
-			scale := boidMaxSpeed / (absF(b.vx) + absF(b.vy))
-			b.vx *= scale * 1.4
-			b.vy *= scale * 1.4
-		}
-
-		// Move
-		b.x += b.vx
-		b.y += b.vy
-	}
-}
-
-func (s *boidsState) render() []uint8 {
-	w, h := s.width, s.height
-	if w == 0 || h == 0 {
-		return make([]uint8, rgbaChannels)
-	}
-
-	pixels := make([]uint8, w*h*rgbaChannels)
-
-	for _, b := range s.flock {
-		cx, cy := int(b.x), int(b.y)
-		if cx < 0 || cx >= w || cy < 0 || cy >= h {
-			continue
-		}
-
-		// Draw as a small triangle pointing in movement direction
-		s.drawBoid(pixels, b)
-	}
-
-	return pixels
-}
-
-func (s *boidsState) drawBoid(pixels []uint8, b boid) {
-	w, h := s.width, s.height
-
-	// Normalize velocity for direction
-	speed := absF(b.vx) + absF(b.vy)
-	if speed < 0.1 {
-		speed = 0.1
-	}
-	dirX := b.vx / speed
-	dirY := b.vy / speed
-
-	// Triangle points: nose in direction of travel, tail behind
-	size := 5.0
-	noseX := b.x + dirX*size
-	noseY := b.y + dirY*size
-	// Perpendicular for tail width
-	perpX := -dirY * size * 0.5
-	perpY := dirX * size * 0.5
-	tailX1 := b.x - dirX*size*0.3 + perpX
-	tailY1 := b.y - dirY*size*0.3 + perpY
-	tailX2 := b.x - dirX*size*0.3 - perpX
-	tailY2 := b.y - dirY*size*0.3 - perpY
-
-	// Color from hue
-	r, g, bl := boidColor(b.hue)
-
-	// Draw filled triangle using scanline
-	s.drawBoidTriangle(pixels, w, h,
-		int(noseX), int(noseY),
-		int(tailX1), int(tailY1),
-		int(tailX2), int(tailY2),
-		r, g, bl)
-}
-
-func (s *boidsState) drawBoidTriangle(pixels []uint8, w, h, x0, y0, x1, y1, x2, y2 int, r, g, b uint8) {
-	// Simple: draw lines between vertices
-	s.drawBoidLine(pixels, w, h, x0, y0, x1, y1, r, g, b)
-	s.drawBoidLine(pixels, w, h, x1, y1, x2, y2, r, g, b)
-	s.drawBoidLine(pixels, w, h, x2, y2, x0, y0, r, g, b)
-	// Fill center
-	cx := (x0 + x1 + x2) / 3
-	cy := (y0 + y1 + y2) / 3
-	if cx >= 0 && cx < w && cy >= 0 && cy < h {
-		offset := (cy*w + cx) * rgbaChannels
-		if boidAlpha > pixels[offset+3] {
-			pixels[offset] = r
-			pixels[offset+1] = g
-			pixels[offset+2] = b
-			pixels[offset+3] = boidAlpha
-		}
-	}
-}
-
-func (s *boidsState) drawBoidLine(pixels []uint8, w, h, x0, y0, x1, y1 int, r, g, b uint8) {
-	steps := abs(x1-x0) + abs(y1-y0)
-	if steps == 0 {
-		return
-	}
-	if steps > 30 {
-		steps = 30
-	}
-	for i := 0; i <= steps; i++ {
-		t := float64(i) / float64(steps)
-		px := x0 + int(float64(x1-x0)*t)
-		py := y0 + int(float64(y1-y0)*t)
-		if px >= 0 && px < w && py >= 0 && py < h {
-			offset := (py*w + px) * rgbaChannels
-			if boidAlpha > pixels[offset+3] {
-				pixels[offset] = r
-				pixels[offset+1] = g
-				pixels[offset+2] = b
-				pixels[offset+3] = boidAlpha
-			}
-		}
-	}
-}
-
-func boidColor(hue float64) (uint8, uint8, uint8) {
-	// Map hue to a nice color (blue-teal-cyan range)
-	h := hue - float64(int(hue))
-	r := uint8(sinNorm(h*6.28+4.19) * 200)
-	g := uint8(sinNorm(h*6.28+2.09) * 230)
-	b := uint8(sinNorm(h*6.28) * 255)
-	return r, g, b
-}
-
 // NewForTesting creates a fresh shenanigans instance for use in tests,
 // avoiding copying the package-level Plugin variable (which may contain a mutex).
 func NewForTesting() *shenanigans {
